% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analogizer.R
\name{aggregateByCluster}
\alias{aggregateByCluster}
\title{Aggregate gene-level measurements across cells within clusters to allow correlation across datasets}
\usage{
aggregateByCluster(object)
}
\arguments{
\item{object}{analogizer object. Should run quantile_norm and possibly clusterLouvainJaccard before calling.}
}
\value{
analogizer object with agg.data
}
\description{
Aggregate gene-level measurements across cells within clusters to allow correlation across datasets
}
\examples{
\dontrun{
Y = matrix(c(1,2,3,4,5,6,7,8,9,10,11,12),nrow=4,byrow=T)
Z = matrix(c(1,2,3,4,5,6,7,6,5,4,3,2),nrow=4,byrow=T)
analogy = Analogizer(Y,Z)
analogy@var.genes = c(1,2,3,4)
analogy = scaleNotCenter(analogy)
analogy = optimize_als(analogy,k=2,nrep=1)
analogy = quantile_norm(analogy)
analogy = clusterLouvainJaccard(object)
}
}
