% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analogizer.R
\name{alignment_metric}
\alias{alignment_metric}
\title{Calculate alignment statistic to quantify how well-aligned two or more datasets are.}
\usage{
alignment_metric(object, k = NULL, rand.seed = 1, cells.use = NULL,
  clusters.use = NULL, by_dataset = F)
}
\arguments{
\item{object}{analogizer object. Should call quantile_norm before calling.}

\item{k}{Number of nearest neighbors to use in calculating alignment.}

\item{rand.seed}{Random seed for reproducibility}

\item{cells.use}{Vector of cells across all datasets to use in calculating score}

\item{clusters.use}{Names of clusters to use in calculating score}
}
\value{
alignment statistic
}
\description{
Calculate alignment statistic to quantify how well-aligned two or more datasets are.
}
\examples{
\dontrun{
Y = matrix(c(1,2,3,4,5,6,7,8,9,10,11,12),nrow=4,byrow=T)
Z = matrix(c(1,2,3,4,5,6,7,6,5,4,3,2),nrow=4,byrow=T)
analogy = Analogizer(Y,Z)
analogy@var.genes = c(1,2,3,4)
analogy = scaleNotCenter(analogy)
analogy = optimize_als(analogy,k=2,nrep=1)
}
}
