% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analogizer.R
\name{alignment_metric_per_cluster}
\alias{alignment_metric_per_cluster}
\title{Calculate alignment statistic for each cluster in aligned datasets.}
\usage{
alignment_metric_per_cluster(object, rand.seed = 1, k = NULL,
  by_dataset = F)
}
\arguments{
\item{object}{analogizer object. Should call quantile_norm before calling.}

\item{rand.seed}{Random seed for reproducibility}

\item{k}{Number of nearest neighbors in calculating alignment
Can pass in single value or vector with same length as clusters}
}
\value{
Vector of alignment statistics (with names of clusters)
}
\description{
Calculate alignment statistic for each cluster in aligned datasets.
}
\examples{
\dontrun{
Y = matrix(c(1,2,3,4,5,6,7,8,9,10,11,12),nrow=4,byrow=T)
Z = matrix(c(1,2,3,4,5,6,7,6,5,4,3,2),nrow=4,byrow=T)
analogy = Analogizer(Y,Z)
analogy@var.genes = c(1,2,3,4)
analogy = scaleNotCenter(analogy)
analogy = optimize_als(analogy,k=2,nrep=1)
}
}
