% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{alignment_metric_per_factor}
\alias{alignment_metric_per_factor}
\title{Calculate alignment metric per factor.}
\usage{
alignment_metric_per_factor(object, k = NULL)
}
\arguments{
\item{object}{Analogizer object. Should run quantile_align_SNF before calling.}
}
\description{
Calculate alignment metric per factor.
}
