\name{Analogizer-package}
\alias{Analogizer-package}
\alias{Analogizer}
\docType{package}
\title{
\packageTitle{Analogizer}
}
\description{
\packageDescription{Analogizer}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{Analogizer}
\packageIndices{Analogizer}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{Analogizer}

Maintainer: \packageMaintainer{Analogizer}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
