% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analogizer.R
\docType{class}
\name{analogizer}
\alias{analogizer}
\alias{analogizer-class}
\title{The Analogizer Class}
\description{
The analogizer object is created from two or more single cell datasets.
The class provides functions for data preprocessing, integrative analysis, and visualization.
}
\section{Slots}{

\describe{
\item{\code{raw.data}}{List of raw data matrices, one per experiment (genes by cells)}

\item{\code{norm.data}}{List of normalized matrices}

\item{\code{scale.data}}{List of scaled matrices}

\item{\code{var.genes}}{Subset of informative genes shared across datasets to be used in matrix factorization}

\item{\code{H}}{Cell loading factors (one matrix per dataset)}

\item{\code{H.norm}}{Normalized cell loading factors (one matrix per dataset)}

\item{\code{W}}{Shared gene loading factors}

\item{\code{V}}{Dataset-specific gene loading factors}

\item{\code{tsne.coords}}{Matrix of 2D coordinates obtained from running t-SNE on H.norm}

\item{\code{clusters}}{Joint cluster assignments for cells}

\item{\code{snf}}{Shared nearest factor matrix for use in clustering and alignment}

\item{\code{agg.data}}{Data aggregated within clusters}

\item{\code{parameters}}{List of parameters used in analysis}
}}

