% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analogizer.R
\name{AnalogizerToSeurat}
\alias{AnalogizerToSeurat}
\title{Create a Seurat object containing the data from an Analogizer object}
\usage{
AnalogizerToSeurat(object, need.sparse = T)
}
\arguments{
\item{object}{Analogizer object.}

\item{need.sparse}{Whether data needs to be converted to sparse format first; most relevant for 
older Analogizer objects (default TRUE).}
}
\value{
Seurat object with raw.data, scale.data, dr$tsne, dr$inmf, and ident slots set.
}
\description{
Carries over raw.data, scale.data, tsne.coords and the iNMF factorization, plus cluster 
assignment.s
}
\examples{
\dontrun{
# Analogizer object
analogy
s.object <- AnalogizerToSeurat(analogy)
}
}
