% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bmmc}
\alias{bmmc}
\title{liger object of bone marrow subsample data with RNA and ATAC modality}
\format{
\linkS4class{liger} object with two dataset named by "rna" and "atac"
}
\source{
https://www.nature.com/articles/s41587-019-0332-7
}
\usage{
bmmc
}
\description{
liger object of bone marrow subsample data with RNA and ATAC modality
}
\references{
Jeffrey M. Granja and et. al., Nature Biotechnology, 2019
}
\keyword{datasets}
