% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liger-class.R
\name{c.liger}
\alias{c.liger}
\title{Combine multiple liger object in to one}
\usage{
\method{c}{liger}(...)
}
\arguments{
\item{...}{\linkS4class{liger} objects}
}
\value{
A new \linkS4class{liger} object containing all datasets in the order
of specification.
}
\description{
The list of \code{datasets} slot, the rows of
\code{cellMeta} slot and the list of \code{commands} slot will be simply
concatenated. Variable features in \code{varFeatures} slot will be taken a
union. The \eqn{W} and \eqn{H.norm} matrices are not taken into account for
now.
}
