% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analogizer.R
\name{calc_dataset_specificity}
\alias{calc_dataset_specificity}
\title{Uses the factorization information to calculate a dataset-specificity score for each factor}
\usage{
calc_dataset_specificity(object)
}
\arguments{
\item{object}{analogizer object. Should run optimizeALS before calling.}
}
\value{
List containing three elements. First two elements are the norm of each metagene factor for each dataset. Last element is the vector of dataset specificity scores.
}
\description{
Uses the factorization information to calculate a dataset-specificity score for each factor
}
\examples{
\dontrun{
Y = matrix(c(1,2,3,4,5,6,7,8,9,10,11,12),nrow=4,byrow=T)
Z = matrix(c(1,2,3,4,5,6,7,6,5,4,3,2),nrow=4,byrow=T)
analogy = Analogizer(Y,Z)
analogy@var.genes = c(1,2,3,4)
analogy = scaleNotCenter(analogy)
analogy = optimize_als(analogy,k=2,nrep=1)
analogy = quantile_norm(analogy)
analogy = clusterLouvainJaccard(object)
}
}
