% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analogizer.R
\name{calcARI}
\alias{calcARI}
\title{Calculate adjusted Rand index}
\usage{
calcARI(object, clusters.compare)
}
\arguments{
\item{object}{Analogizer object. Should run quantileAlignSNF before calling.}

\item{clusters.compare}{Clustering with which to compare (named vector).}
}
\value{
Value of ARI

Adjusted Rand index value.
}
\description{
Computes adjusted Rand index for Analogizer clustering and external clustering.
The Rand index ranges from 0 to 1, with 0 indicating no agreement between clusterings and 1 
indicating perfect agreement.
}
\examples{
\dontrun{
# analogizer object, factorization done
analogy
analogy <- quantileAlignSNF(analogy)
# toy clusters 
cluster1 <- sample(c('type1', 'type2', 'type3'), ncol(analogy@raw.data[[1]]), replace = T)
names(cluster1) <- colnames(analogy@raw.data[[1]])
cluster2 <- sample(c('type4', 'type5', 'type6'), ncol(analogy@raw.data[[2]]), replace = T)
names(cluster2) <- colnames(analogy@raw.data[[2]])
# get ARI for first clustering
ari1 <- calcARI(analogy, cluster1)
# get ARI for second clustering
ari2 <- calcARI(analogy, cluster2)
}
}
