% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectGenes.R
\name{calcGeneVars.H5}
\alias{calcGeneVars.H5}
\title{Calculate Gene Variance for ligerDataset object}
\usage{
calcGeneVars.H5(object, chunkSize = 1000, verbose = TRUE)
}
\arguments{
\item{object}{ligerDataset object}

\item{chunkSize}{Integer for the maximum number of cells in each chunk.
Default \code{1000}.}

\item{verbose}{Logical. Whether to show a progress bar. Default \code{TRUE}.}
}
\value{
The input \code{object} with calculated var updated in the H5 file.
}
\description{
Calculate Gene Variance for ligerDataset object
}
