% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{calcGeneVars.old}
\alias{calcGeneVars.old}
\title{Calculate variance of gene expression across cells in an online fashion}
\usage{
calcGeneVars.old(object, chunk = 1000, verbose = TRUE)
}
\arguments{
\item{object}{\code{liger} object. The input raw.data should be a list of
hdf5 files. Should call normalize and selectGenes before calling.}

\item{chunk}{size of chunks in hdf5 file. (default 1000)}

\item{verbose}{Print progress bar/messages (TRUE by default)}
}
\value{
\code{liger} object with scale.data slot set.
}
\description{
This function calculates the variance of gene expression values across cells
for hdf5 files.
}
