% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analogizer.R
\name{calcPurity}
\alias{calcPurity}
\title{Calculate purity}
\usage{
calcPurity(object, classes.compare)
}
\arguments{
\item{object}{Analogizer object. Should run quantileAlignSNF before calling.}

\item{classes.compare}{Clustering with which to compare (named vector).}
}
\value{
Purity value.
}
\description{
Calculates purity for Analogizer clustering and external clustering (true clusters/classes). 
Purity can sometimes be a more useful metric when the clustering to be tested contains more 
subgroups or clusters than the true clusters (or classes). Purity also ranges from 0 to 1,
with a score of 1 representing a pure, or accurate, clustering.
}
\examples{
\dontrun{
# analogizer object, factorization done
analogy
analogy <- quantileAlignSNF(analogy)
# toy clusters 
cluster1 <- sample(c('type1', 'type2', 'type3'), ncol(analogy@raw.data[[1]]), replace = T)
names(cluster1) <- colnames(analogy@raw.data[[1]])
cluster2 <- sample(c('type4', 'type5', 'type6'), ncol(analogy@raw.data[[2]]), replace = T)
names(cluster2) <- colnames(analogy@raw.data[[2]])
# get ARI for first clustering
ari1 <- calcARI(analogy, cluster1)
# get ARI for second clustering
ari2 <- calcARI(analogy, cluster2)
}
}
