% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liger.R
\name{calculateWilcoxon}
\alias{calculateWilcoxon}
\title{Analyze differential gene expression using presto's implementation of the Wilcoxon rank sum test.}
\usage{
calculateWilcoxon(object, compare.datasets = TRUE,
  compare.clusters = TRUE)
}
\arguments{
\item{object}{\code{liger} object. Should run quantileAlignSNF first}

\item{compare.datasets}{differential expression analysis for different datasets within in each cluster
(default TRUE)}

\item{compare.clusters}{differential expression analysis for different clusters (default TRUE).}
}
\value{
Updated \code{liger} object.
}
\description{
Analyze differential gene expression using presto's implementation of the Wilcoxon rank sum test.
}
\examples{
\dontrun{
# liger object, clustering complete 
ligerex
# fill wilcoxon slot with dataframe of information from presto's Wilcoxon rank sum test
ligerex <- calculateWilcoxon(ligerex)
}

}
