% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ligerDataset-class.R
\name{cbind.ligerDataset}
\alias{cbind.ligerDataset}
\title{cbind method for ligerDataset objects}
\usage{
\method{cbind}{ligerDataset}(x, ..., deparse.level = 1)
}
\arguments{
\item{x}{\linkS4class{ligerDataset} object}

\item{...}{\linkS4class{ligerDataset} objects}

\item{deparse.level}{Not used here.}
}
\value{
A new \linkS4class{liger} object containing all datasets in the order
of specification.
}
\description{
The list of \code{datasets} slot, the rows of
\code{cell.meta} slot and the list of \code{commands} slot will be simply
concatenated. Variable features in \code{var.features} slot will be taken a
union. The \eqn{W} and \eqn{H.norm} matrices are not taken into account for
now.
}
