% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ccGeneHuman}
\alias{ccGeneHuman}
\title{Cell cycle gene set for human}
\format{
A list of two character vectors:
\describe{
\item{s.genes}{Genes associated with S-phase}
\item{g2m.genes}{Genes associated with G2M-phase}
}
}
\source{
https://www.science.org/doi/abs/10.1126/science.aad0501
}
\usage{
ccGeneHuman
}
\description{
Copied from Seurat::cc.genes
}
\keyword{datasets}
