% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liger-class.R
\name{cell.meta<-,liger-method}
\alias{cell.meta<-,liger-method}
\title{Set cell.meta to liger object}
\usage{
\S4method{cell.meta}{liger}(x) <- value
}
\arguments{
\item{x}{liger object}

\item{value}{ligerDataset object}
}
\description{
Set cell.meta to liger object
}
