% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liger-class.R
\name{cellMeta,liger,NULL-method}
\alias{cellMeta,liger,NULL-method}
\alias{cellMeta,liger,character-method}
\alias{cellMeta,liger,missing-method}
\alias{cellMeta<-,liger,missing-method}
\alias{cellMeta<-,liger,character-method}
\title{Access cellMeta of liger object}
\usage{
\S4method{cellMeta}{liger,`NULL`}(x, columns = NULL, cellIdx = NULL, as.data.frame = FALSE, ...)

\S4method{cellMeta}{liger,character}(x, columns = NULL, cellIdx = NULL, as.data.frame = FALSE, ...)

\S4method{cellMeta}{liger,missing}(x, columns = NULL, cellIdx = NULL, as.data.frame = FALSE, ...)

\S4method{cellMeta}{liger,missing}(x, columns = NULL, check = FALSE) <- value

\S4method{cellMeta}{liger,character}(x, columns = NULL, check = FALSE) <- value
}
\arguments{
\item{x}{\linkS4class{liger} object}
}
\description{
Access cellMeta of liger object

Access cellMeta of liger object

Access cellMeta of liger object
}
