% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ligerCommand-class.R
\name{commands}
\alias{commands}
\alias{commands,liger-method}
\title{Fetch command record log}
\usage{
commands(x, funcName = NULL, arg = NULL)

\S4method{commands}{liger}(x, funcName = NULL, arg = NULL)
}
\arguments{
\item{x}{\linkS4class{liger} object}

\item{funcName}{Function name of the command of interests, or an exact
command label for the call.}

\item{arg}{Argument names}
}
\description{
Fetch command record log
}
