% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zb-generics.R, R/zc-liger-methods.R,
%   R/zc-ligerDataset_subclass-methods.R
\name{coordinate}
\alias{coordinate}
\alias{coordinate<-}
\alias{coordinate,liger,character-method}
\alias{coordinate<-,liger,character-method}
\alias{coordinate,ligerSpatialDataset,missing-method}
\alias{coordinate<-,ligerSpatialDataset,missing-method}
\title{Access ligerSpatialDataset coordinate data}
\usage{
coordinate(x, dataset)

coordinate(x, dataset, check = TRUE) <- value

\S4method{coordinate}{liger,character}(x, dataset)

\S4method{coordinate}{liger,character}(x, dataset, check = TRUE) <- value

\S4method{coordinate}{ligerSpatialDataset,missing}(x, dataset = NULL)

\S4method{coordinate}{ligerSpatialDataset,missing}(x, dataset = NULL, check = TRUE) <- value
}
\arguments{
\item{x}{\linkS4class{ligerSpatialDataset} object or a \linkS4class{liger}
object.}

\item{dataset}{Name or numeric index of an spatial dataset.}

\item{check}{Logical, whether to perform object validity check on setting new
value.}

\item{value}{\code{\link{matrix}}.}
}
\value{
The retrieved coordinate matrix or the updated \code{x} object.
}
\description{
Similar as how default \linkS4class{ligerDataset} data is
accessed.
}
