% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{createH5LigerDataset}
\alias{createH5LigerDataset}
\title{Create on-disk ligerDataset Object}
\usage{
createH5LigerDataset(
  h5file,
  formatType = NULL,
  rawData = NULL,
  normData = NULL,
  scaleData = NULL,
  barcodesName = NULL,
  genesName = NULL,
  indicesName = NULL,
  indptrName = NULL,
  modal = c("default", "rna", "atac"),
  featureMeta = NULL,
  ...
)
}
\arguments{
\item{h5file}{Filename of a H5 file}

\item{formatType}{Select preset of H5 file structure. Current available
options are \code{"10X"} and \code{"AnnData"}.}

\item{rawData, indicesName, indptrName}{The path in a H5 file for the raw
sparse matrix data. These three types of data stands for the \code{x},
\code{i}, and \code{p} slots of a \code{\link[Matrix]{dgCMatrix-class}}
object. Default \code{NULL} uses \code{formatType} preset.}

\item{normData}{The path in a H5 file for the "x" vector of the normalized
sparse matrix. Default \code{NULL}.}

\item{scaleData}{The path in a H5 file for the dense 2D scaled matrix.
Default \code{NULL}.}

\item{genesName, barcodesName}{The path in a H5 file for the gene names and
cell barcodes. Default \code{NULL} uses \code{formatType} preset.}

\item{modal}{Name of modality for this dataset. Currently options of
\code{"default"}, \code{"rna"}, and \code{"atac"} are supported. Default
\code{"default"}.}

\item{featureMeta}{Data frame for feature metadata. Default \code{NULL}.}

\item{...}{Additional slot data. See \linkS4class{ligerDataset} for detail.
Given values will be directly placed at corresponding slots.}
}
\description{
Create on-disk ligerDataset Object
}
