% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ligerDataset-class.R
\name{createLigerDataset}
\alias{createLigerDataset}
\title{Create ligerDataset object}
\usage{
createLigerDataset(
  rawData = NULL,
  modal = c("default", "rna", "atac"),
  normData = NULL,
  scaleData = NULL,
  scaleUnsharedData = NULL,
  varUnsharedFeatures = NULL,
  H = NULL,
  V = NULL,
  A = NULL,
  B = NULL,
  U = NULL,
  h5fileInfo = NULL,
  featureMeta = NULL,
  ...
)
}
\arguments{
\item{rawData}{matrix}

\item{normData}{matrix}

\item{V}{matrix}

\item{A}{matrix}

\item{B}{matrix}

\item{U}{matrix}

\item{annotation}{data.frame}
}
\description{
Create ligerDataset object
}
