% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ligerDataset-class.R
\name{raw.data,ligerDataset-method}
\alias{raw.data,ligerDataset-method}
\alias{raw.data<-,ligerDataset,ANY,matrixLike_OR_NULL-method}
\alias{raw.data<-,ligerDataset,ANY,H5D-method}
\alias{norm.data,ligerDataset-method}
\alias{norm.data<-,ligerDataset,ANY,matrixLike_OR_NULL-method}
\alias{norm.data<-,ligerDataset,ANY,H5D-method}
\alias{scale.data,ligerDataset,ANY-method}
\alias{scale.data,liger,character-method}
\alias{scale.data,liger,numeric-method}
\alias{scale.data<-,ligerDataset,ANY,matrixLike_OR_NULL-method}
\alias{scale.data<-,ligerDataset,ANY,H5D-method}
\title{Access data in a ligerDataset object}
\usage{
\S4method{raw.data}{ligerDataset}(x)

\S4method{raw.data}{ligerDataset,ANY,matrixLike_OR_NULL}(x, check = TRUE) <- value

\S4method{raw.data}{ligerDataset,ANY,H5D}(x, check = TRUE) <- value

\S4method{norm.data}{ligerDataset}(x)

\S4method{norm.data}{ligerDataset,ANY,matrixLike_OR_NULL}(x, check = TRUE) <- value

\S4method{norm.data}{ligerDataset,ANY,H5D}(x, check = TRUE) <- value

\S4method{scale.data}{ligerDataset,ANY}(x, name = NULL)

\S4method{scale.data}{liger,character}(x, name = NULL)

\S4method{scale.data}{liger,numeric}(x, name = NULL)

\S4method{scale.data}{ligerDataset,ANY,matrixLike_OR_NULL}(x, check = TRUE) <- value

\S4method{scale.data}{ligerDataset,ANY,H5D}(x, check = TRUE) <- value
}
\arguments{
\item{x}{\linkS4class{ligerDataset} object}

\item{check}{Whether to perform object validity check on setting new value.}

\item{value}{matrix like or \code{NULL}}
}
\description{
Access data in a ligerDataset object
}
