% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liger-class.R
\name{dataset,liger,character_OR_NULL-method}
\alias{dataset,liger,character_OR_NULL-method}
\alias{dataset,liger,missing-method}
\alias{dataset<-,liger,character,missing,ligerDataset-method}
\alias{dataset<-,liger,character,ANY,matrixLike-method}
\alias{dataset<-,liger,character,missing,NULL-method}
\title{Access single dataset of liger object}
\usage{
\S4method{dataset}{liger,character_OR_NULL}(x, name = NULL)

\S4method{dataset}{liger,missing}(x, name = NULL)

\S4method{dataset}{liger,character,missing,ligerDataset}(x, name, type = NULL) <- value

\S4method{dataset}{liger,character,ANY,matrixLike}(x, name, type = c("raw.data", "norm.data", "scale.data")) <- value

\S4method{dataset}{liger,character,missing,`NULL`}(x, name, type = NULL) <- value
}
\arguments{
\item{x}{\linkS4class{liger} object}

\item{name}{name of dataset}

\item{type}{Type of specified matrix. Only works when \code{value} is a
matrix like object. choose from \code{"raw.data"}, \code{"norm.data"} or
\code{"scale.data"}. Default \code{"raw.data"}.}

\item{value}{A matrix of a dataset to be added, or a constructed
\linkS4class{ligerDataset} object. \code{NULL} to remove a dataset by
\code{name}.}
}
\value{
Getter methods returns a \linkS4class{ligerDataset} object, while
setter methods update the input \code{x} (the \linkS4class{liger} object).
}
\description{
TODO: Debating on whether to have a separate (differently named)
method to directly access matrices in embedded ligerDataset objects, or to
allow accessing matrices with \code{type}.
}
