% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liger-class.R
\name{datasets,liger-method}
\alias{datasets,liger-method}
\title{Access datasets of liger object}
\usage{
\S4method{datasets}{liger}(x)
}
\arguments{
\item{x}{liger object}
}
\description{
Access datasets of liger object
}
