% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liger-class.R
\name{datasets<-,liger-method}
\alias{datasets<-,liger-method}
\title{Set datasets to liger object}
\usage{
\S4method{datasets}{liger}(x) <- value
}
\arguments{
\item{x}{liger object}

\item{value}{ligerDataset object}
}
\description{
Set datasets to liger object
}
