% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liger-class.R
\name{datasets,liger-method}
\alias{datasets,liger-method}
\alias{datasets<-,liger,logical-method}
\alias{datasets<-,liger,missing-method}
\title{Access datasets of liger object}
\usage{
\S4method{datasets}{liger}(x, check = NULL)

\S4method{datasets}{liger,logical}(x, check = TRUE) <- value

\S4method{datasets}{liger,missing}(x, check = TRUE) <- value
}
\arguments{
\item{x}{liger object}

\item{check}{Whether to perform object validity check when using setter
method. Default \code{TRUE}.}

\item{value}{list of \linkS4class{ligerDataset} object.}
}
\description{
This method access the whole slot \code{datasets}, a list, of
the \linkS4class{liger} object. So that operations like
\code{datasets(x)[[1]] <- y} modifies the inner \code{ligerDataset} object
without triggering cell metadata update, which will be done by
\code{dataset(x) <- y}.
}
\seealso{
\code{\link{dataset}}
}
