% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liger-class.R, R/ligerDataset-class.R
\name{dim,liger-method}
\alias{dim,liger-method}
\alias{dimnames,liger-method}
\alias{dimnames<-,liger,list-method}
\alias{dim,ligerDataset-method}
\alias{dimnames,ligerDataset-method}
\alias{dimnames<-,ligerDataset,list-method}
\title{Access dimensionality of liger object}
\usage{
\S4method{dim}{liger}(x)

\S4method{dimnames}{liger}(x)

\S4method{dimnames}{liger,list}(x) <- value

\S4method{dim}{ligerDataset}(x)

\S4method{dimnames}{ligerDataset}(x)

\S4method{dimnames}{ligerDataset,list}(x) <- value
}
\arguments{
\item{x}{\linkS4class{liger} object}

\item{value}{For direct call of \code{dimnames<-} method, should be a list
with \code{NULL} as the first element and valid cell identifiers as the
second element. For \code{colnames<-} method, the character vector of cell
identifiers. \code{rownames<-} method is not applicable.}
}
\value{
See Description
}
\description{
For a \linkS4class{liger} object or a
\linkS4class{ligerDataset} object, the column orientation is
assigned for cells while rows should be for features/genes. However, due to
the data structure, it is hard to define a row index for the outer container
-- \code{liger} object, which might contain datasets that vary in genes.

Therefore, for \code{liger} objects, \code{dim} and \code{dimnames} returns
\code{NA}/\code{NULL} for rows and total cell counts/barcodes for the
columns. For \code{ligerDataset} objects, gene index will be available for
rows.
}
