% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analogizer.R
\name{distortion_metric}
\alias{distortion_metric}
\title{Calculate distortion statistic to quantify how much alignment distorts
the geometry of the original datasets.}
\usage{
distortion_metric(object, dr_method = "PCA", ndims = 40, k = 10,
  use_aligned = TRUE, rand.seed = 42, return_breakdown = FALSE)
}
\arguments{
\item{object}{analogizer object. Should call quantile_norm before calling.}

\item{dr_method}{Dimensionality reduction method to use for assessing pre-alignment geometry (either "PCA", "NMF", or "ICA"). Should call quantile_norm before calling.}
}
\value{
alignment statistic
}
\description{
Calculate distortion statistic to quantify how much alignment distorts
the geometry of the original datasets.
}
\examples{
\dontrun{
Y = matrix(c(1,2,3,4,5,6,7,8,9,10,11,12),nrow=4,byrow=T)
Z = matrix(c(1,2,3,4,5,6,7,6,5,4,3,2),nrow=4,byrow=T)
analogy = Analogizer(Y,Z)
analogy@var.genes = c(1,2,3,4)
analogy = scaleNotCenter(analogy)
analogy = optimize_als(analogy,k=2,nrep=1)
}
}
