% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downsample.R
\name{downsample}
\alias{downsample}
\title{Downsample datasets}
\usage{
downsample(
  object,
  balance = c("all", "cluster", "dataset"),
  maxCells = 1000,
  useDatasets = NULL,
  seed = 1,
  ...
)
}
\arguments{
\item{object}{\linkS4class{liger} object}

\item{balance}{\code{"all"} for sampling \code{maxCells} cells from all
datasets specified by \code{useDatasets}. \code{"cluster"} for sampling
\code{maxCells} cells per cluster per dataset. \code{"dataset"} for
\code{maxCells} cells per dataset.}

\item{maxCells}{Max number of cells to sample from the grouping based on
\code{balance}.}

\item{useDatasets}{Index selection of datasets to consider. Default
\code{NULL} for using all datasets.}

\item{seed}{Random seed for reproducibility. Default \code{1}.}

\item{...}{Arguments passed to \code{\link{subsetLiger}}, where
\code{cellIdx} is occupied by internal implementation.}
}
\value{
Subset of \linkS4class{liger} \code{object}.
}
\description{
This function mainly aims at downsampling datasets to a size
suitable for plotting.
}
