% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analogizer.R
\name{extractSubset}
\alias{extractSubset}
\title{Construct an Analogizer object with a specified subset of cells or clusters.}
\usage{
extractSubset(object, cell.subset = NULL, cluster.subset = NULL)
}
\arguments{
\item{object}{analogizer object.}

\item{cell.subset}{List of names of cells to extract. Names of list elements should match names of datasets.}

\item{cluster.subset}{Clusters to extract}
}
\value{
analogizer object
}
\description{
Construct an Analogizer object with a specified subset of cells or clusters.
}
\examples{
\dontrun{
Y = matrix(c(1,2,3,4,5,6,7,8,9,10,11,12),nrow=4,byrow=T)
Z = matrix(c(1,2,3,4,5,6,7,6,5,4,3,2),nrow=4,byrow=T)
analogy = Analogizer(list(Y,Z))
analogy =
}
}
