% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ligerDataset-class.R
\name{feature.meta,ligerDataset-method}
\alias{feature.meta,ligerDataset-method}
\title{Access the feature metadata in a ligerDataset object}
\usage{
\S4method{feature.meta}{ligerDataset}(x)
}
\arguments{
\item{x}{ligerDataset object}
}
\description{
Access the feature metadata in a ligerDataset object
}
