% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liger-class.R
\name{fortify.liger}
\alias{fortify.liger}
\title{Fortify method for classes from the liger package}
\usage{
\method{fortify}{liger}(x)
}
\arguments{
\item{x}{\code{liger} object to convert into a dataframe.}

\item{...}{not used by this method}
}
\value{
A \code{data.frame} combining \code{cell.meta(x)} with
\code{H} and \code{H.norm} matrices, if any of the latters exist.
}
\description{
Fortify method for classes from the liger package
}
\examples{
\dontrun{
library(ggplot2)
# Create liger object, ligerex
ggplot(ligerex,  aes(x = dataset, y = nUMI)) + geom_violin()
}
}
