% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analogizer.R
\name{get_factor_markers}
\alias{get_factor_markers}
\title{Find shared and dataset-specific markers using the factorization.}
\usage{
get_factor_markers(object, dataset1 = NULL, dataset2 = NULL,
  factor.share.thresh = 10, dataset_specificity = NULL, log_fc_thresh = 1,
  umi_thresh = 30, frac_thresh = 0, pval_thresh = 0.05, num_genes = 30,
  print.genes = F)
}
\arguments{
\item{object}{analogizer object. Should call run_tSNE before calling.}

\item{dataset1}{Name of first dataset}

\item{dataset2}{Name of second dataset}

\item{factor.share.thresh}{Use only factors with a dataset specificity less than threshold}

\item{dataset_specificity}{Pre-calculated dataset specificity if available}

\item{log_fc_thresh}{Lower log-fold change threshold for differential expression in markers}

\item{umi_thresh}{Lower UMI threshold for markers}

\item{frac_thresh}{Lower threshold for fraction of cells expressing marker}

\item{pval_thresh}{Upper p-value threshold for Wilcoxon rank test for gene expression}

\item{num_genes}{Max number of genes to report for each dataset}

\item{print.genes}{Print ordered markers passing logfc, umi and frac thresholds}
}
\description{
Find shared and dataset-specific markers using the factorization.
}
\examples{
\dontrun{
Y = matrix(c(1,2,3,4,5,6,7,8,9,10,11,12),nrow=4,byrow=T)
Z = matrix(c(1,2,3,4,5,6,7,6,5,4,3,2),nrow=4,byrow=T)
analogy = Analogizer(Y,Z)
analogy@var.genes = c(1,2,3,4)
analogy = scaleNotCenter(analogy)
analogy = optimize_als(analogy,k=2,nrep=1)
}
}
