% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ligerDataset-class.R
\name{getH5File,liger,character-method}
\alias{getH5File,liger,character-method}
\title{Access the H5File object in a specific dataset or all datasets of a liger
object}
\usage{
\S4method{getH5File}{liger,character}(x, dataset = NULL)
}
\arguments{
\item{x}{liger object}

\item{dataset}{One or more names of datasets that can be found in \code{x}}
}
\description{
Access the H5File object in a specific dataset or all datasets of a liger
object
}
