% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ligerDataset-class.R
\name{getH5File,ligerDataset,missing-method}
\alias{getH5File,ligerDataset,missing-method}
\title{Access the H5File object in a ligerDataset object}
\usage{
\S4method{getH5File}{ligerDataset,missing}(x, dataset = NULL)
}
\arguments{
\item{x}{ligerDataset object}

\item{dataset}{missing}
}
\description{
Access the H5File object in a ligerDataset object
}
