% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ligerDataset-class.R
\name{getH5File,ligerDataset,missing-method}
\alias{getH5File,ligerDataset,missing-method}
\alias{getH5File,liger,character-method}
\title{Access the H5File object}
\usage{
\S4method{getH5File}{ligerDataset,missing}(x, dataset = NULL)

\S4method{getH5File}{liger,character}(x, dataset = NULL)
}
\arguments{
\item{x}{\linkS4class{ligerDataset} object or \linkS4class{liger} object.}

\item{dataset}{Get H5File from specific dataset(s) if using a
\linkS4class{liger} object.}
}
\description{
Access the H5File object
}
