% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ligerDataset-class.R
\name{getMatrix,ligerDataset,ANY,missing,missing-method}
\alias{getMatrix,ligerDataset,ANY,missing,missing-method}
\alias{getMatrix,liger,ANY,ANY,ANY-method}
\title{Get matrix from liger/ligerDataset object}
\usage{
\S4method{getMatrix}{ligerDataset,ANY,missing,missing}(
  x,
  slot = c("raw.data", "norm.data", "scale.data", "H", "V"),
  dataset = NULL
)

\S4method{getMatrix}{liger,ANY,ANY,ANY}(
  x,
  slot = c("raw.data", "norm.data", "scale.data", "H", "V", "W", "H.norm"),
  dataset = NULL,
  returnList = FALSE
)
}
\arguments{
\item{x}{\linkS4class{liger} or \linkS4class{ligerDataset} object}

\item{slot}{The type of matrix to retrieve. Choose from \code{"raw.data"},
\code{"norm.data"}, \code{"scale.data"}, \code{"H"}, \code{"V"} for both
classes. For \linkS4class{liger} object, \code{"W"} can also be an option,
ignoring \code{dataset}.}

\item{dataset}{When using a \linkS4class{liger} object, from which dataset(s)
to retrieve the data. Single dataset returns the matrix, multiple datasets
return a named list of the type of matrices, default \code{NULL} return a
named list of the type of matrix from all datasets.}
}
\value{
A matrix or a list, depending on \code{dataset}
}
\description{
Get matrix from liger/ligerDataset object
}
