% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h5Utility.R
\name{H5Apply}
\alias{H5Apply}
\title{Apply function to chunks of H5 data in ligerDataset object}
\usage{
H5Apply(
  object,
  FUN,
  init = NULL,
  useData = c("raw.data", "norm.data", "scale.data"),
  chunkSize = 1000,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{FUN}{A function where the first argument must be the chunk of data, the
return value must be a vector.}

\item{h5file}{H5File object}
}
\description{
h5 calculation wrapper, that run user specified calculation with
on-disk matrix in chunks
}
