% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ligerDataset-class.R
\name{h5file.info,ligerDataset,ANY,missing-method}
\alias{h5file.info,ligerDataset,ANY,missing-method}
\title{Access the H5File information in a ligerDataset object}
\usage{
\S4method{h5file.info}{ligerDataset,ANY,missing}(x, info = NULL, dataset = NULL)
}
\arguments{
\item{x}{ligerDataset object}

\item{info}{Character vector of queried information}

\item{dataset}{missing}
}
\description{
Access the H5File information in a ligerDataset object
}
