% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ligerDataset-class.R
\name{h5file.info,ligerDataset-method}
\alias{h5file.info,ligerDataset-method}
\alias{h5file.info<-,ligerDataset-method}
\title{Access the H5File information}
\usage{
\S4method{h5file.info}{ligerDataset}(x, info = NULL)

\S4method{h5file.info}{ligerDataset}(x, info = NULL, check = TRUE) <- value
}
\arguments{
\item{x}{\linkS4class{ligerDataset}}

\item{info}{Character vector of queried information. Default \code{NULL}
returns a full list of information.}
}
\description{
H5File information generally stores the "paths" to all types of
data in the H5 file. Setter method changes not only information stored in
\code{h5file.info} slot, but also data slots (i.e. \code{raw.data},
\code{norm.data} or \code{scale.data}) that a related with the corresponding
information.
}
