% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{importVignetteData}
\alias{importVignetteData}
\title{Import prepared dataset publically available}
\usage{
importVignetteData(
  dataset,
  overwrite = FALSE,
  dir = getwd(),
  method = "libcurl",
  verbose = getOption("ligerVerbose"),
  ...
)
}
\arguments{
\item{dataset}{Name of dataset, see available options with
\code{names(.manifest)}.}

\item{overwrite}{Logical, if a file exists at corresponding download
location, whether to re-download or directly use this file. Default
\code{FALSE}.}

\item{dir}{Path to download datasets. Default \code{getwd()}.}

\item{method}{\code{method} argument directly passed to
\code{\link[utils]{download.file}}. Using \code{"libcurl"} while other
options might not work depending on platform.}

\item{verbose}{Logical. Whether to show information of the progress. Default
\code{getOption("ligerVerbose")} which is \code{TRUE} if users have not set.}

\item{...}{Additional arguments passed to \code{\link{download.file}}}
}
\value{
\code{\linkS4class{liger}} object of specified dataset.
}
\description{
Import prepared dataset publically available
}
\examples{
if (FALSE) {
    pbmc <- importVignetteData("pbmc")
}
}
