% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liger.R
\name{imputateKNN}
\alias{imputateKNN}
\title{Impute the query cell expression matrix}
\usage{
imputateKNN(object, knn_k = 50, reference, query, weight = FALSE)
}
\arguments{
\item{object}{\code{liger} object.}

\item{knn_k}{The maximum number of nearest neighbors to search. The default value is set to 50.}

\item{reference}{Name of the reference data}

\item{query}{Name of the query data}

\item{weight}{Use KNN distances as weight matrix (default FALSE)}
}
\value{
\code{liger} object with impute.data slot set.
}
\description{
Impute query features from a reference dataset using KNN.
}
\examples{
\dontrun{
Y <- matrix(c(1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12), nrow = 4, byrow = T)
Z <- matrix(c(1, 2, 3, 4, 5, 6, 7, 6, 5, 4, 3, 2), nrow = 4, byrow = T)
ligerex <- createLiger(list(y_set = Y, z_set = Z))
ligerex <- normalize(ligerex)
# select genes
ligerex <- selectGenes(ligerex)
ligerex <- scaleNotCenter(ligerex)
ligerex <- imputateKNN(y_set, z_set, weight = TRUE)
}
}
