% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rliger.R
\name{imputeKNN}
\alias{imputeKNN}
\title{Impute the query cell expression matrix}
\usage{
imputeKNN(
  object,
  reference,
  queries,
  knn_k = 20,
  weight = TRUE,
  norm = TRUE,
  scale = FALSE
)
}
\arguments{
\item{object}{\code{liger} object.}

\item{reference}{Dataset containing values to impute into query dataset(s).}

\item{queries}{Dataset to be augmented by imputation. If not specified, will pass in all datasets.}

\item{knn_k}{The maximum number of nearest neighbors to search. (default 20)}

\item{weight}{Whether to use KNN distances as weight matrix (default FALSE).}

\item{norm}{Whether normalize the imputed data with default parameters (default TRUE).}

\item{scale}{Whether scale but not center the imputed data with default parameters (default TRUE).}
}
\value{
\code{liger} object with raw data in raw.data slot replaced by imputed data (genes by cells)
}
\description{
Impute query features from a reference dataset using KNN.
}
\examples{
\dontrun{
Y <- matrix(c(1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12), nrow = 4, byrow = T)
Z <- matrix(c(1, 2, 3, 4, 5, 6, 7, 6, 5, 4, 3, 2), nrow = 4, byrow = T)
X <- matrix(c(1, 2, 3, 4, 5, 6, 7, 2, 3, 4, 5, 6), nrow = 4, byrow = T)
ligerex <- createLiger(list(y_set = Y, z_set = Z, x_set = X))
ligerex <- normalize(ligerex)
# select genes
ligerex <- selectGenes(ligerex)
ligerex <- scaleNotCenter(ligerex)
ligerex <- optimizeALS(ligerex, k = 20)
ligerex <- quantileAlignSNF(ligerex)
# impute every dataset other than the reference dataset
ligerex <- imputeKNN(ligerex, reference = "y_set", weight = FALSE)
# impute only z_set dataset
ligerex <- imputeKNN(ligerex, reference = "y_set", queries = list("z_set"), knn_k = 50)
}
}
