% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zc-liger-methods.R
\name{is.newLiger}
\alias{is.newLiger}
\title{Check if given liger object if under new implementation}
\usage{
is.newLiger(object)
}
\arguments{
\item{object}{A liger object}
}
\value{
\code{TRUE} if the version of \code{object} is later than or equal to
1.99.0. Otherwise \code{FALSE}. It raises an error if input object is not of
\linkS4class{liger} class.
}
\description{
Check if given liger object if under new implementation
}
\examples{
is.newLiger(pbmc) # TRUE
}
