% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analogizer.R
\name{kSuggestion}
\alias{kSuggestion}
\title{Plot alignment and agreement for various test values of lambda. Can be used to select
appropriate value of lambda for factorization of particular dataset.}
\usage{
kSuggestion(object, k_test = seq(5, 50, 5), lambda = 5, thresh = 1e-04,
  max_iters = 100, num.cores = 1, rand.seed = 1, plot.metric = "median",
  gen.new = F, return_results = F)
}
\arguments{
\item{object}{analogizer object. Should normalize, select genes, and scale before calling.}

\item{k_test}{Set of factor numbers to test (default seq(5, 50, 5))}

\item{lambda}{Lambda for optimization}

\item{thresh}{Convergence threshold. Convergence occurs when |obj0-obj|/(mean(obj0,obj)) < thresh}

\item{max_iters}{Maximum number of block coordinate descent iterations to perform}

\item{num.cores}{Number of cores to use for optimizing factorizations in parallel}

\item{rand.seed}{Random seed for reproducibility}

\item{gen.new}{Do not use optimizeNewK in factorizations.}

\item{return_results}{Return matrix of alignment and agreement values}
}
\value{
Matrix of results or plot
}
\description{
Plot alignment and agreement for various test values of lambda. Can be used to select
appropriate value of lambda for factorization of particular dataset.
}
\examples{
\dontrun{
Y = matrix(c(1,2,3,4,5,6,7,8,9,10,11,12),nrow=4,byrow=T)
Z = matrix(c(1,2,3,4,5,6,7,6,5,4,3,2),nrow=4,byrow=T)
analogy = Analogizer(Y,Z)
analogy@var.genes = c(1,2,3,4)
analogy = scaleNotCenter(analogy)
analogy = optimize_als(analogy,k=2,nrep=1)
}
}
