% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liger-class.R
\name{datasets}
\alias{datasets}
\alias{datasets<-}
\alias{dataset}
\alias{dataset<-}
\alias{cellMeta}
\alias{cellMeta<-}
\alias{varFeatures}
\alias{varFeatures<-}
\alias{[[,liger,ANY,missing-method}
\alias{[[<-,liger,ANY,missing-method}
\alias{$,liger-method}
\alias{$<-,liger-method}
\alias{datasets,liger-method}
\alias{datasets<-,liger,logical-method}
\alias{datasets<-,liger,missing-method}
\alias{dataset,liger,character_OR_NULL-method}
\alias{dataset,liger,missing-method}
\alias{dataset,liger,numeric-method}
\alias{dataset<-,liger,character,missing,ANY,ligerDataset-method}
\alias{dataset<-,liger,character,ANY,ANY,matrixLike-method}
\alias{dataset<-,liger,character,missing,ANY,NULL-method}
\alias{cellMeta,liger,NULL-method}
\alias{cellMeta,liger,character-method}
\alias{cellMeta,liger,missing-method}
\alias{cellMeta<-,liger,missing-method}
\alias{cellMeta<-,liger,character-method}
\alias{varFeatures,liger-method}
\alias{varFeatures<-,liger,ANY,character-method}
\title{Access data in a liger object}
\usage{
datasets(x, check = NULL)

datasets(x, check = TRUE) <- value

dataset(x, name = NULL)

dataset(x, name, type = NULL, qc = TRUE) <- value

cellMeta(x, columns = NULL, cellIdx = NULL, as.data.frame = FALSE, ...)

cellMeta(x, columns = NULL, check = FALSE) <- value

varFeatures(x)

varFeatures(x, check = TRUE) <- value

\S4method{[[}{liger,ANY,missing}(x, i, j, ...)

\S4method{[[}{liger,ANY,missing}(x, i, j, ...) <- value

\S4method{$}{liger}(x, name)

\S4method{$}{liger}(x, name) <- value

\S4method{datasets}{liger}(x, check = NULL)

\S4method{datasets}{liger,logical}(x, check = TRUE) <- value

\S4method{datasets}{liger,missing}(x, check = TRUE) <- value

\S4method{dataset}{liger,character_OR_NULL}(x, name = NULL)

\S4method{dataset}{liger,missing}(x, name = NULL)

\S4method{dataset}{liger,numeric}(x, name = NULL)

\S4method{dataset}{liger,character,missing,ANY,ligerDataset}(x, name, type = NULL, qc = TRUE) <- value

\S4method{dataset}{liger,character,ANY,ANY,matrixLike}(x, name, type = c("rawData", "normData", "scaleData"), qc = TRUE) <- value

\S4method{dataset}{liger,character,missing,ANY,`NULL`}(x, name, type = NULL, qc = TRUE) <- value

\S4method{cellMeta}{liger,`NULL`}(x, columns = NULL, cellIdx = NULL, as.data.frame = FALSE, ...)

\S4method{cellMeta}{liger,character}(x, columns = NULL, cellIdx = NULL, as.data.frame = FALSE, ...)

\S4method{cellMeta}{liger,missing}(x, columns = NULL, cellIdx = NULL, as.data.frame = FALSE, ...)

\S4method{cellMeta}{liger,missing}(x, columns = NULL, check = FALSE) <- value

\S4method{cellMeta}{liger,character}(x, columns = NULL, check = FALSE) <- value

\S4method{varFeatures}{liger}(x)

\S4method{varFeatures}{liger,ANY,character}(x, check = TRUE) <- value
}
\arguments{
\item{x}{\linkS4class{liger} object}

\item{check}{Logical, whether to perform object validity check on setting new
value.}

\item{value}{Check \linkS4class{liger} class specification for required value
type.}

\item{name}{Name or numeric index of a dataset}

\item{type}{When using \code{dataset<-} with a matrix like \code{value},
specify what type the matrix is. Choose from \code{"rawData"},
\code{"normData"} or \code{"scaleData"}.}

\item{qc}{Logical, whether to perform general qc on added new dataset.}

\item{columns}{The names of available variables in \code{cellMeta} slot. When
\code{as.data.frame = TRUE}, please use variable names after coercion.}

\item{cellIdx}{Valid cell subscription to subset retrieved variables. Default
\code{NULL} uses all cells.}

\item{as.data.frame}{Logical, whether to apply
\code{\link[base]{as.data.frame}} on the subscription. Default \code{FALSE}.}

\item{...}{For \code{cellMeta}, arguments passed to \code{[]} subscriptor,
e.g. \code{drop = FALSE}.}

\item{i, j}{Feature or cell index for \code{`[`} method. For \code{`[[`}
method, use a single variable name with \code{i} and \code{j} is not
applicable.}
}
\description{
The methods listed in this manual only apply to
\linkS4class{liger} objects, while there are other generics that have methods
for both \code{liger} and \linkS4class{ligerDataset} objects. Please see
\code{\link{rawData}} page.
}
\section{Whether to use datasets() or dataset()?}{

\code{datasets()} method only accesses the \code{datasets} slot, the list of
\linkS4class{ligerDataset} objects. \code{dataset()} method accesses a single
dataset, with subsequent cell metadata updates and checks bonded when adding
or modifying a dataset. Therefore, when users want to modify something inside
a \code{ligerDataset} while no cell metadata change should happen, it is
recommended to use: \code{datasets(x)[[name]] <- ligerD} for efficiency,
though the result would be the same as \code{dataset(x, name) <- ligerD}.
}

\section{Cell metadata access}{

Three approaches are provided for access of cell metadata. A generic function
\code{cellMeta} is implemented with plenty of options and multi-variable
accessibility. Besides, users can use double-bracket (e.g.
\code{ligerObj[[Var]]}) or dollor-sign (e.g. \code{ligerObj$nUMI}) to access
or modify single variables.
}

