% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ctrl}
\alias{ctrl}
\alias{stim}
\title{dgCMatrix object of PBMC subsample data with Control and Stimulated datasets}
\format{
\code{dgCMatrix} object of gene expression matrix from PBMC study,
 named by "ctrl" and "stim".

An object of class \code{dgCMatrix} with 262 rows and 300 columns.
}
\source{
https://www.nature.com/articles/nbt.4042
}
\usage{
ctrl

stim
}
\description{
dgCMatrix object of PBMC subsample data with Control and Stimulated datasets
}
\references{
Hyun Min Kang and et. al., Nature Biotechnology, 2018
}
\keyword{datasets}
