% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{as.cellMeta}
\alias{as.cellMeta}
\alias{as.featureMeta}
\alias{dimnames.cellMeta}
\alias{row.names<-.cellMeta}
\alias{row.names<-.featureMeta}
\alias{dimnames.featureMeta}
\alias{[.cellMeta}
\alias{[.featureMeta}
\alias{[<-.cellMeta}
\alias{[<-.featureMeta}
\title{Liger cell/feature metadata}
\usage{
as.cellMeta(x, row.names = NULL)

as.featureMeta(x, row.names = NULL)

\method{dimnames}{cellMeta}(x)

\method{row.names}{cellMeta}(x) <- value

\method{row.names}{featureMeta}(x) <- value

\method{dimnames}{featureMeta}(x)

\method{[}{cellMeta}(x, i, j, drop = FALSE)

\method{[}{featureMeta}(x, i, j, drop = FALSE)

\method{[}{cellMeta}(x, i, j, ...) <- value

\method{[}{featureMeta}(x, i, j, ...) <- value
}
\arguments{
\item{x}{A data frame or tibble containing cell or feature metadata.}

\item{row.names}{If \code{x} is a tibble without the preserved unique ID
column (i.e., \code{.cellID} for cell metadata or \code{.featureID} for
feature metadata), which column should be used as the unique ID for cells or
features. This argument is ignored if \code{x} already has the preserved
unique ID column.}

\item{value}{Replacement value when using the setter method.}

\item{i, j}{Row and column subscription.}

\item{drop}{If \code{TRUE} the result is coerced to the lowest possible
dimension (see the examples). This only works for extracting elements, not
for the replacement.}

\item{...}{Additional arguments passed to \verb{[.data.frame} method.}
}
\value{
For \code{as.*} functions, a tibble with class \code{cellMeta} or
\code{featureMeta}. The first column is the preserved unique ID column (i.e.,
\code{.cellID} or \code{.featureID}).

For \code{dimnames} method, a list of two character vectors: the first vector
is the unique IDs, and the second is the column names.

For \code{[]} method, a subset of the original metadata tibble.
}
\description{
The rationale for defining these classes is to have a data frame extension
class holding the data, so that printed view looks tidy. \code{tibble} is
selected as the base class for metadata. However, \code{tibble} does not
allow \code{rownames}. Therefore, we lightly extend the class to preserve
a column for unique IDs of cells or features, and bring \code{rownames}
and character subscription of rows (\verb{[]}) back to it.
}
\examples{
as.cellMeta(pbmc@cellMeta)
}
