\name{liger-package}
\alias{liger-package}
\docType{package}
\title{
\packageTitle{liger}
}
\description{
\packageDescription{liger}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{liger}
\packageIndices{liger}
 An overview of how to use the package, including the most important functions 
}
\author{
\packageAuthor{liger}

Maintainer: \packageMaintainer{liger}
}
\references{
Literature or other references for background information 
}
\keyword{ package }