% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/za-classes.R
\docType{class}
\name{ligerATACDataset-class}
\alias{ligerATACDataset-class}
\alias{ligerATACDataset}
\title{Subclass of ligerDataset for ATAC modality}
\description{
Inherits from \linkS4class{ligerDataset} class. Contained slots
can be referred with the link.
}
\section{Slots}{

\describe{
\item{\code{rawPeak}}{sparse matrix}

\item{\code{normPeak}}{sparse matrix}
}}

