% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ligerDataset-class.R
\name{rawData}
\alias{rawData}
\alias{rawData<-}
\alias{normData}
\alias{normData<-}
\alias{scaleData}
\alias{scaleData<-}
\alias{scaleUnsharedData}
\alias{scaleUnsharedData<-}
\alias{getH5File}
\alias{h5fileInfo}
\alias{h5fileInfo<-}
\alias{featureMeta}
\alias{featureMeta<-}
\alias{getMatrix}
\alias{rawData,ligerDataset-method}
\alias{rawData<-,ligerDataset,ANY,matrixLike_OR_NULL-method}
\alias{rawData<-,ligerDataset,ANY,H5D-method}
\alias{normData,ligerDataset-method}
\alias{normData<-,ligerDataset,ANY,matrixLike_OR_NULL-method}
\alias{normData<-,ligerDataset,ANY,H5D-method}
\alias{scaleData,ligerDataset,ANY-method}
\alias{scaleData,liger,character-method}
\alias{scaleData,liger,numeric-method}
\alias{scaleData<-,ligerDataset,ANY,matrixLike_OR_NULL-method}
\alias{scaleData<-,ligerDataset,ANY,H5D-method}
\alias{scaleUnsharedData,ligerDataset,ANY-method}
\alias{scaleUnsharedData,liger,character-method}
\alias{scaleUnsharedData,liger,numeric-method}
\alias{scaleUnsharedData<-,ligerDataset,ANY,matrixLike_OR_NULL-method}
\alias{scaleUnsharedData<-,ligerDataset,ANY,H5D-method}
\alias{getH5File,ligerDataset,missing-method}
\alias{getH5File,liger,character-method}
\alias{h5fileInfo,ligerDataset-method}
\alias{h5fileInfo<-,ligerDataset-method}
\alias{featureMeta,ligerDataset-method}
\alias{featureMeta<-,ligerDataset-method}
\alias{getMatrix,ligerDataset,ANY,missing,missing-method}
\alias{getMatrix,liger,ANY,ANY,ANY-method}
\title{Access data in a liger or ligerDataset object}
\usage{
rawData(x)

rawData(x, check = TRUE) <- value

normData(x)

normData(x, check = TRUE) <- value

scaleData(x, name = NULL)

scaleData(x, check = TRUE) <- value

scaleUnsharedData(x, name = NULL)

scaleUnsharedData(x, check = TRUE) <- value

getH5File(x, dataset = NULL)

h5fileInfo(x, info = NULL)

h5fileInfo(x, info = NULL, check = TRUE) <- value

featureMeta(x, check = NULL)

featureMeta(x, check = TRUE) <- value

getMatrix(x, slot = "rawData", dataset = NULL, returnList = FALSE)

\S4method{rawData}{ligerDataset}(x)

\S4method{rawData}{ligerDataset,ANY,matrixLike_OR_NULL}(x, check = TRUE) <- value

\S4method{rawData}{ligerDataset,ANY,H5D}(x, check = TRUE) <- value

\S4method{normData}{ligerDataset}(x)

\S4method{normData}{ligerDataset,ANY,matrixLike_OR_NULL}(x, check = TRUE) <- value

\S4method{normData}{ligerDataset,ANY,H5D}(x, check = TRUE) <- value

\S4method{scaleData}{ligerDataset,ANY}(x, name = NULL)

\S4method{scaleData}{liger,character}(x, name = NULL)

\S4method{scaleData}{liger,numeric}(x, name = NULL)

\S4method{scaleData}{ligerDataset,ANY,matrixLike_OR_NULL}(x, check = TRUE) <- value

\S4method{scaleData}{ligerDataset,ANY,H5D}(x, check = TRUE) <- value

\S4method{scaleUnsharedData}{ligerDataset,ANY}(x, name = NULL)

\S4method{scaleUnsharedData}{liger,character}(x, name = NULL)

\S4method{scaleUnsharedData}{liger,numeric}(x, name = NULL)

\S4method{scaleUnsharedData}{ligerDataset,ANY,matrixLike_OR_NULL}(x, check = TRUE) <- value

\S4method{scaleUnsharedData}{ligerDataset,ANY,H5D}(x, check = TRUE) <- value

\S4method{getH5File}{ligerDataset,missing}(x, dataset = NULL)

\S4method{getH5File}{liger,character}(x, dataset = NULL)

\S4method{h5fileInfo}{ligerDataset}(x, info = NULL)

\S4method{h5fileInfo}{ligerDataset}(x, info = NULL, check = TRUE) <- value

\S4method{featureMeta}{ligerDataset}(x, check = NULL)

\S4method{featureMeta}{ligerDataset}(x, check = TRUE) <- value

\S4method{getMatrix}{ligerDataset,ANY,missing,missing}(
  x,
  slot = c("rawData", "normData", "scaleData", "scaleUnsharedData", "H", "V", "U"),
  dataset = NULL
)

\S4method{getMatrix}{liger,ANY,ANY,ANY}(
  x,
  slot = c("rawData", "normData", "scaleData", "scaleUnsharedData", "H", "V", "U", "W",
    "H.norm"),
  dataset = NULL,
  returnList = FALSE
)
}
\arguments{
\item{x}{\linkS4class{ligerDataset} object or \linkS4class{liger} object}

\item{check}{Whether to perform object validity check on setting new value.}

\item{value}{matrix like or \code{NULL}}

\item{name, dataset}{Name or numeric index of a dataset when \code{x} is a
\linkS4class{liger} object, in order to only get the data from this specified
dataset. Default \code{NULL}.}

\item{info}{Name of the entry in \code{h5fileInfo} slot.}

\item{slot}{The slot name when using \code{getMatrix}.}

\item{returnList}{When using \code{getMatrix} and data from one dataset is
being returned, whether it should be contained in a list or return the data
directly.}
}
\description{
Access data in a liger or ligerDataset object
}
