% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ligerDataset-class.R
\name{ligerDataset}
\alias{ligerDataset}
\title{ligerDataset class}
\usage{
ligerDataset(
  raw.data = NULL,
  modal = c("default", "rna", "atac"),
  norm.data = NULL,
  scale.data = NULL,
  scale.unshared.data = NULL,
  var.unshared.features = NULL,
  W = NULL,
  V = NULL,
  A = NULL,
  B = NULL,
  U = NULL,
  h5file.info = NULL,
  ...
)

ligerDataset(
  raw.data = NULL,
  modal = c("default", "rna", "atac"),
  norm.data = NULL,
  scale.data = NULL,
  scale.unshared.data = NULL,
  var.unshared.features = NULL,
  W = NULL,
  V = NULL,
  A = NULL,
  B = NULL,
  U = NULL,
  h5file.info = NULL,
  ...
)
}
\arguments{
\item{raw.data}{matrix}

\item{norm.data}{matrix}

\item{W}{matrix}

\item{V}{matrix}

\item{A}{matrix}

\item{B}{matrix}

\item{U}{matrix}

\item{annotation}{data.frame}
}
\description{
Object for storing dastaset specific information. Will be embedded within a
higher level \linkS4class{liger} object
}
\section{Slots}{

\describe{
\item{\code{raw.data}}{Raw data.}

\item{\code{norm.data}}{Normalized data}

\item{\code{scale.data}}{Scaled data, usually with subset variable features}

\item{\code{scale.unshared.data}}{Scaled data of features not shared with other
datasets}

\item{\code{var.unshared.features}}{Variable features not shared with other datasets}

\item{\code{W}}{matrix}

\item{\code{V}}{matrix}

\item{\code{A}}{matrix}

\item{\code{B}}{matrix}

\item{\code{U}}{matrix}

\item{\code{agg.data}}{list}

\item{\code{h5file.info}}{list}
}}

