% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/za-classes.R
\docType{class}
\name{ligerRNADataset-class}
\alias{ligerRNADataset-class}
\alias{ligerRNADataset}
\title{Subclass of ligerDataset for RNA modality}
\description{
Inherits from \linkS4class{ligerDataset} class. Contained slots
can be referred with the link. This subclass does not have any different from
the default \code{ligerDataset} class except the class name.
}
