% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/za-classes.R
\docType{class}
\name{ligerSpatialDataset-class}
\alias{ligerSpatialDataset-class}
\alias{ligerSpatialDataset}
\title{Subclass of ligerDataset for Spatial modality}
\description{
Inherits from \linkS4class{ligerDataset} class. Contained slots
can be referred with the link.
}
\section{Slots}{

\describe{
\item{\code{coordinate}}{dense matrix}
}}

