% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ligerToH5AD.R
\name{ligerToH5AD}
\alias{ligerToH5AD}
\title{Write liger object to H5AD files}
\usage{
ligerToH5AD(
  object,
  filename,
  overwrite = FALSE,
  verbose = getOption("ligerVerbose", TRUE)
)
}
\arguments{
\item{object}{A \linkS4class{liger} object}

\item{filename}{A character string, the path to the H5AD file to be written}

\item{overwrite}{Logical, whether to overwrite the file if it exists.}

\item{verbose}{Logical. Whether to show information of the progress. Default
\code{getOption("ligerVerbose")} which is \code{TRUE} if users have not set.}
}
\value{
No return value, an H5AD file is written to disk
}
\description{
Write liger object to H5AD files
}
\examples{
ligerToH5AD(pbmc, filename = tempfile(fileext = ".h5ad"))
}
