% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analogizer.R
\name{make_riverplot}
\alias{make_riverplot}
\title{Makes a riverplot to show how separate cluster assignments from two datasets map onto a joint clustering.}
\usage{
make_riverplot(object, cluster1, cluster2, cluster_consensus = NULL,
  min.frac = 0.05, min.cells = 10, river.yscale = 1, river.lty = 0,
  river.node_margin = 0.1, label.cex = 1, label.col = "black",
  lab.srt = 0, node.order = "auto")
}
\arguments{
\item{object}{analogizer object. Should run quantile_align_SNF before calling.}

\item{cluster1}{Cluster assignments for dataset 1. Note that cluster names should be distinct across datasets.}

\item{cluster2}{Cluster assignments for dataset 2. Note that cluster names should be distinct across datasets.}

\item{cluster_consensus}{Optional external consensus clustering (to use instead of object clusters)}

\item{min.frac}{Minimum fraction of cluster for edge to be shown}

\item{min.cells}{Minumum number of cells for edge to be shown}

\item{river.yscale}{y-scale to pass to riverplot}

\item{river.lty}{lty to pass to riverplot}

\item{river.node_margin}{node_margin to pass to riverplot}

\item{label.cex}{Size of text labels}

\item{label.col}{Color of text labels}

\item{lab.srt}{Angle of text labels}

\item{node.order}{Order of clusters in each set (list with three vectors of ordinal numbers)}
}
\description{
Makes a riverplot to show how separate cluster assignments from two datasets map onto a joint clustering.
}
\examples{
\dontrun{
Y = matrix(c(1,2,3,4,5,6,7,8,9,10,11,12),nrow=4,byrow=T)
Z = matrix(c(1,2,3,4,5,6,7,6,5,4,3,2),nrow=4,byrow=T)
analogy = Analogizer(list(Y,Z))
analogy@var.genes = c(1,2,3,4)
analogy = scaleNotCenter(analogy)
}
}
