% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ATAC.R
\name{makeInteractTrack-deprecated}
\alias{makeInteractTrack-deprecated}
\title{[Deprecated] Export predicted gene-pair interaction}
\arguments{
\item{corr.mat}{A sparse matrix of correlation with peak names as rows and
gene names as columns.}

\item{path_to_coords}{Path to the gene coordinates file.}

\item{genes.list}{Character vector of gene names to be exported. Default
\code{NULL} uses all genes available in \code{corrMat}.}

\item{output_path}{Path of filename where the output file will be stored. If
a folder, a file named \code{"Interact_Track.bed"} will be created. Default
current working directory.}
}
\value{
No return value. A file located at \code{outputPath} will be created.
}
\description{
Export the predicted gene-pair interactions calculated by
upstream function \code{\link{linkGenesAndPeaks}} into an Interact Track file
which is compatible with \href{https://genome.ucsc.edu/cgi-bin/hgCustom}{UCSC
Genome Browser}.
}
\seealso{
\code{\link{rliger-deprecated}}, \code{\link{exportInteractTrack}}
}
