% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{Matrix.column_norm}
\alias{Matrix.column_norm}
\title{Perform fast and memory-efficient normalization operation on sparse matrix data.}
\usage{
Matrix.column_norm(A)
}
\arguments{
\item{A}{Sparse matrix DGE.}
}
\description{
Perform fast and memory-efficient normalization operation on sparse matrix data.
}
\examples{
\dontrun{
Y = matrix(c(1,2,3,4,5,6,7,8,9,10,11,12),nrow=4,byrow=T)
Z = matrix(c(1,2,3,4,5,6,7,6,5,4,3,2),nrow=4,byrow=T)
ligerex = createLiger(list(Y,Z))
ligerex@var.genes = c(1,2,3,4)
ligerex = scaleNotCenter(ligerex)
}
}
