% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liger.R
\name{mergeH5}
\alias{mergeH5}
\title{Merge hdf5 files}
\usage{
mergeH5(file.list, library.names, new.filename = "merged.h5")
}
\arguments{
\item{file.list}{List of path to hdf5 files.}

\item{library.names}{Vector of library names (corresponding to file.list)}

\item{new.filename}{String of new hdf5 file name after merging (default merged.h5).}
}
\value{
Directly generates newly merged hdf5 file.
}
\description{
This function merges hdf5 files generated from different libraries (cell ranger by default).
}
\examples{
\dontrun{
mergeH5(list("library1.h5","library2.h5"), c("lib1","lib2"), "merged.h5")
}
}
