% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liger-class.R
\name{names,liger-method}
\alias{names,liger-method}
\alias{names<-,liger,character-method}
\alias{length,liger-method}
\title{Access dataset names or length}
\usage{
\S4method{names}{liger}(x)

\S4method{names}{liger,character}(x) <- value

\S4method{length}{liger}(x)
}
\arguments{
\item{x}{\linkS4class{liger} object}

\item{value}{Character vector of new names for datasets.}
}
\value{
character vector of dataset names or number of datasets
}
\description{
Access dataset names or length
}
