% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{nnzeroGroups}
\alias{nnzeroGroups}
\alias{nnzeroGroups.dgCMatrix}
\alias{nnzeroGroups.matrix}
\title{nnzeroGroups}
\usage{
nnzeroGroups(X, y, MARGIN = 2)

\method{nnzeroGroups}{dgCMatrix}(X, y, MARGIN = 2)

\method{nnzeroGroups}{matrix}(X, y, MARGIN = 2)
}
\arguments{
\item{X}{matrix}

\item{y}{group labels}

\item{MARGIN}{whether observations are rows (=2) or columns (=1)}
}
\value{
Matrix of groups by features
}
\description{
Utility function to compute number of zeros-per-feature within group
}
