% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rliger.R
\name{nonneg}
\alias{nonneg}
\title{Perform thresholding on dense matrix}
\usage{
nonneg(x, eps = 1e-16)
}
\arguments{
\item{x}{Dense matrix.}

\item{eps}{Threshold. Should be a small positive value. (default 1e-16)}
}
\value{
Dense matrix with smallest values equal to eps.
}
\description{
Perform thresholding on the input dense matrix. Remove any values samller than eps by eps.
Helper function for online_iNMF
}
