% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ligerDataset-class.R
\name{norm.data,ligerDataset-method}
\alias{norm.data,ligerDataset-method}
\title{Access norm.data of ligerDataset object}
\usage{
\S4method{norm.data}{ligerDataset}(x)
}
\arguments{
\item{x}{ligerDataset object}
}
\description{
Access norm.data of ligerDataset object
}
