% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liger.R
\name{norm_scale_unshared}
\alias{norm_scale_unshared}
\title{This function selects the most variable genes from those not shared between datasets. 
These features are normalized and scaled.}
\usage{
norm_scale_unshared(object, num.feat = 2500)
}
\arguments{
\item{object}{\code{liger} object}

\item{num.feat}{Approximate number of most variable features to select. (default 2500)}
}
\value{
a normalized and scaled matrix of additional genes
}
\description{
This function selects the most variable genes from those not shared between datasets. 
These features are normalized and scaled.
}
