% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/online_iNMF.R
\name{online_iNMF}
\alias{online_iNMF}
\title{Online iNMF}
\usage{
online_iNMF(
  object,
  X_new = NULL,
  projection = FALSE,
  W.init = NULL,
  V.init = NULL,
  H.init = NULL,
  A.init = NULL,
  B.init = NULL,
  k = 20,
  lambda = 5,
  max.epochs = 5,
  miniBatch_max_iters = 1,
  miniBatch_size = 5000,
  seed = 123,
  verbose = TRUE
)
}
\description{
Online iNMF
}
