% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liger.R
\name{online_iNMF_h5}
\alias{online_iNMF_h5}
\title{Perform online iNMF on scaled datasets (hdf5 files)}
\usage{
online_iNMF_h5(object, X_new = NULL, projection = FALSE,
  W.init = NULL, V.init = NULL, H.init = NULL, A.init = NULL,
  B.init = NULL, k = 40, lambda = 5, max.epochs = 5,
  miniBatch_max_iters = 1, miniBatch_size = 1000, thresh = 1e-04,
  h5_chunk_size = 1000, seed = 123)
}
\arguments{
\item{object}{\code{liger} object. Should normalize, select genes, and scale before calling.}

\item{X_new}{List of new datasets.}

\item{projection}{Data integration by Metagene (W) projection. (default FALSE)}

\item{W.init}{Previous H matrices. (default NULL)}

\item{V.init}{Previous W matrix (default NULL)}

\item{H.init}{Previous V matrices (default NULL)}

\item{A.init}{Previous A matrix (default NULL)}

\item{B.init}{Previous B matrices (default NULL)}

\item{k}{Inner dimension of factorization (number of factors). Run suggestK to determine
appropriate value; a general rule of thumb is that a higher k will be needed for datasets with
more sub-structure. (default 20)}

\item{lambda}{Regularization parameter. Larger values penalize dataset-specific effects more
strongly (ie. alignment should increase as lambda increases). Run suggestLambda to determine
most appropriate value for balancing dataset alignment and agreement (default 5.0).}

\item{max.epochs}{Maximum number of epochs. (default 1)}

\item{miniBatch_max_iters}{Maximum number of block coordinate descent (HALS algorithm) iterations to perform (default 1).}

\item{miniBatch_size}{Total number of cells in each mini-batch (default 1000).}

\item{thresh}{Convergence threshold. Convergence occurs when |obj0-obj|/(mean(obj0,obj)) < thresh.
(default 1e-4)}

\item{h5_chunk_size}{Chunk size of input hdf5 files (default 1000).}

\item{seed}{Random seed to allow reproducible results (default 123).}
}
\value{
\code{liger} object with H, W, V, A and B slots set.
}
\description{
Perform online integrative non-negative matrix factorization to return factorized H, W, and V 
matrices. It optimizes the iNMF objective function using online learning (alternating non-negative
least squares for H matrix, Hierarchical alternating least squares for W and V matrices), where the 
number of factors is set by k.

For each dataset, this factorization produces an H matrix (cells by k), a V matrix (k by genes),
and a shared W matrix (k by genes). The H matrices represent the cell factor loadings.
W is held consistent among all datasets, as it represents the shared components of the metagenes
across datasets. The V matrices represent the dataset-specific components of the metagenes.
}
