% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analogizer.R
\name{optimize_als}
\alias{optimize_als}
\title{Optimize objective function using block coordinate descent (alternating nonnegative least squares)}
\usage{
optimize_als(E, k, lambda = 5, thresh = 1e-04, max_iters = 100,
  nrep = 20)
}
\arguments{
\item{E}{List of N data matrices. Samples (cells) are in rows, features (genes) in columns.}

\item{k}{Inner dimension of factorization (number of factors)}

\item{lambda}{Regularization parameter. Larger values penalize dataset-specific effects more strongly.}

\item{thresh}{Convergence threshold. Convergence occurs when |obj0-obj|/(mean(obj0,obj)) < thresh}

\item{max_iters}{Maximum number of block coordinate descent iterations to perform}

\item{nrep}{Number of restarts to perform (NMF objectives are non-convex, so taking the best objective from multiple successive initializations is recommended)}
}
\value{
list of length N+2 containing Hs, W, Vs, and run statistics
}
\description{
Optimize objective function using block coordinate descent (alternating nonnegative least squares)
}
\examples{
\dontrun{
Y = matrix(c(1,2,3,4,5,6,7,8,9,10,11,12),nrow=4,byrow=T)
Z = matrix(c(1,2,3,4,5,6,7,6,5,4,3,2),nrow=4,byrow=T)
optimize_als(list(Y,Z),k=2,nrep=1)
}
}
