% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liger.R
\name{optimize_UANLS}
\alias{optimize_UANLS}
\title{Perform iNMF on scaled datasets, and include unshared, scaled and normalized, features}
\usage{
optimize_UANLS(
  object,
  k = 30,
  lambda = 5,
  max.iters = 30,
  nrep = 1,
  thresh = 1e-10,
  rand.seed = 1,
  print.obj = FALSE
)
}
\arguments{
\item{object}{\code{liger} object. Should normalize, select genes, and scale before calling.}

\item{k}{Inner dimension of factorization (number of factors).}

\item{lambda}{The lambda penalty. Default 5}

\item{max.iters}{Maximum number of block coordinate descent iterations to perform (default 30).}

\item{nrep}{Number of restarts to perform (iNMF objective function is non-convex, so taking the
best objective from multiple successive initializations is recommended). For easier
reproducibility, this increments the random seed by 1 for each consecutive restart, so future
factorizations of the same dataset can be run with one rep if necessary. (default 1)}

\item{thresh}{Convergence threshold. Convergence occurs when |obj0-obj|/(mean(obj0,obj)) < thresh.
(default 1e-6)}

\item{rand.seed}{Random seed to allow reproducible results (default 1).}

\item{print.obj}{Print objective function values after convergence (default FALSE).}
}
\description{
Perform iNMF on scaled datasets, and include unshared, scaled and normalized, features
}
