% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integration.R
\name{optimizeALS-deprecated}
\alias{optimizeALS-deprecated}
\title{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} Perform iNMF on scaled datasets}
\arguments{
\item{object}{\code{liger} object. Should normalize, select genes, and scale
before calling.}

\item{k}{Inner dimension of factorization (number of factors). Run suggestK
to determine appropriate value; a general rule of thumb is that a higher k
will be needed for datasets with more sub-structure.}

\item{lambda}{Regularization parameter. Larger values penalize
dataset-specific effects more strongly (ie. alignment should increase as
lambda increases). Run suggestLambda to determine most appropriate value for
balancing dataset alignment and agreement (default 5.0).}

\item{thresh}{Convergence threshold. Convergence occurs when
|obj0-obj|/(mean(obj0,obj)) < thresh. (default 1e-6)}

\item{max.iters}{Maximum number of block coordinate descent iterations to
perform (default 30).}

\item{nrep}{Number of restarts to perform (iNMF objective function is
non-convex, so taking the best objective from multiple successive
initializations is recommended). For easier reproducibility, this increments
the random seed by 1 for each consecutive restart, so future factorizations
of the same dataset can be run with one rep if necessary. (default 1)}

\item{H.init}{Initial values to use for H matrices. (default NULL)}

\item{W.init}{Initial values to use for W matrix (default NULL)}

\item{V.init}{Initial values to use for V matrices (default NULL)}

\item{rand.seed}{Random seed to allow reproducible results (default 1).}

\item{print.obj}{Print objective function values after convergence (default
FALSE).}

\item{verbose}{Print progress bar/messages (TRUE by default)}

\item{...}{Arguments passed to other methods}
}
\value{
\code{liger} object with H, W, and V slots set.
}
\description{
\bold{Please turn to \code{\link{runINMF}} or \code{\link{runIntegration}}}.

Perform integrative non-negative matrix factorization to return factorized H,
W, and V matrices. It optimizes the iNMF objective function using block
coordinate descent (alternating non-negative least squares), where the number
of factors is set by k. TODO: include objective function equation here in
documentation (using deqn)

For each dataset, this factorization produces an H matrix (cells by k), a V
matrix (k by genes), and a shared W matrix (k by genes). The H matrices
represent the cell factor loadings. W is held consistent among all datasets,
as it represents the shared components of the metagenes across datasets. The
V matrices represent the dataset-specific components of the metagenes.
}
\seealso{
\code{\link{rliger-deprecated}}
}
