% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liger.R
\name{optimizeNewK}
\alias{optimizeNewK}
\title{Perform factorization for new value of k}
\usage{
optimizeNewK(object, k.new, lambda = NULL, thresh = 1e-04,
  max.iters = 100, rand.seed = 1)
}
\arguments{
\item{object}{\code{liger} object. Should call optimizeALS before calling.}

\item{k.new}{Inner dimension of factorization (number of factors)}

\item{lambda}{Regularization parameter. By default, this will use the lambda last used with
optimizeALS.}

\item{thresh}{Convergence threshold. Convergence occurs when |obj0-obj|/(mean(obj0,obj)) < thresh
(default 1e-4).}

\item{max.iters}{Maximum number of block coordinate descent iterations to perform (default 100).}

\item{rand.seed}{Random seed to set. Only relevant if k.new > k. (default 1)}
}
\value{
\code{liger} object with H, W, and V slots reset.
}
\description{
This uses an efficient strategy for updating that takes advantage of the information in the 
existing factorization. It is most recommended for values of k smaller than current value,
where it is more likely to speed up the factorization.
}
\examples{
\dontrun{
Y <- matrix(c(1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12), nrow = 4, byrow = T)
Z <- matrix(c(1, 2, 3, 4, 5, 6, 7, 6, 5, 4, 3, 2), nrow = 4, byrow = T)
ligerex <- newLiger(list(y_set = Y, z_set = Z))
ligerex <- normalize(ligerex)
ligerex <- selectGenes(ligerex)
ligerex <- scaleNotCenter(ligerex)
# get factorization using three restarts and 20 factors
ligerex <- optimizeALS(ligerex, k = 20, lambda = 5, nrep = 3)
# decide to run with k = 15 instead (keeping old lambda the same)
ligerex <- optimizeNewK(ligerex, k.new = 15)
}
}
