% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{org.Hs.GO2Symbol.list}
\alias{org.Hs.GO2Symbol.list}
\title{Human Gene Ontology to HUGO Symbol list}
\format{List with each entry as a Gene Ontology gene set}
\source{
\url{http://geneontology.org/docs/download-ontology/}
}
\usage{
org.Hs.GO2Symbol.list
}
\description{
Human Gene Ontology to HUGO Symbol list
}
\keyword{datasets}
