% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pbmc}
\alias{pbmc}
\title{liger object of PBMC subsample data with Control and Stimulated datasets}
\format{
\linkS4class{liger} object with two datasets named by "ctrl" and
"stim".
}
\source{
https://www.nature.com/articles/nbt.4042
}
\usage{
pbmc
}
\description{
liger object of PBMC subsample data with Control and Stimulated datasets
}
\references{
Hyun Min Kang and et. al., Nature Biotechnology, 2018
}
\keyword{datasets}
