% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pbmcPlot}
\alias{pbmcPlot}
\title{liger object of PBMC subsample data with plotting information available}
\format{
\linkS4class{liger} object with two datasets named by "ctrl" and
"stim".
}
\source{
https://www.nature.com/articles/nbt.4042
}
\usage{
pbmcPlot
}
\description{
This data was generated from data \code{"pbmc"} with default
parameter integration pipeline, with setting \code{k = 20, maxIter = 10}
for \code{\link{optimizeALS}}. UMAP was generated with \code{minDist = 0.5}.
Only the 2nd and 3rd factors, the normalized expression of the top 50
variable genes, and the clustering label were kept for visualization example.
}
\keyword{datasets}
