% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zb-generics.R, R/zc-liger-methods.R,
%   R/zc-ligerDataset_subclass-methods.R
\name{rawPeak}
\alias{rawPeak}
\alias{rawPeak<-}
\alias{normPeak}
\alias{normPeak<-}
\alias{rawPeak,liger,character-method}
\alias{rawPeak<-,liger,character-method}
\alias{normPeak,liger,character-method}
\alias{normPeak<-,liger,character-method}
\alias{rawPeak,ligerATACDataset,missing-method}
\alias{rawPeak<-,ligerATACDataset,missing-method}
\alias{normPeak,ligerATACDataset,missing-method}
\alias{normPeak<-,ligerATACDataset,missing-method}
\title{Access ligerATACDataset peak data}
\usage{
rawPeak(x, dataset)

rawPeak(x, dataset, check = TRUE) <- value

normPeak(x, dataset)

normPeak(x, dataset, check = TRUE) <- value

\S4method{rawPeak}{liger,character}(x, dataset)

\S4method{rawPeak}{liger,character}(x, dataset, check = TRUE) <- value

\S4method{normPeak}{liger,character}(x, dataset)

\S4method{normPeak}{liger,character}(x, dataset, check = TRUE) <- value

\S4method{rawPeak}{ligerATACDataset,missing}(x, dataset = NULL)

\S4method{rawPeak}{ligerATACDataset,missing}(x, dataset = NULL, check = TRUE) <- value

\S4method{normPeak}{ligerATACDataset,missing}(x, dataset = NULL)

\S4method{normPeak}{ligerATACDataset,missing}(x, dataset = NULL, check = TRUE) <- value
}
\arguments{
\item{x}{\linkS4class{ligerATACDataset} object or a \linkS4class{liger}
object.}

\item{dataset}{Name or numeric index of an ATAC dataset.}

\item{check}{Logical, whether to perform object validity check on setting new
value.}

\item{value}{\code{\link[Matrix]{dgCMatrix-class}} matrix.}
}
\value{
The retrieved peak count matrix or the updated \code{x} object.
}
\description{
Similar as how default \linkS4class{ligerDataset} data is
accessed.
}
