% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analogizer.R
\name{plot_gene_violin}
\alias{plot_gene_violin}
\title{Plot violin plots for expression of specified gene for each dataset}
\usage{
plot_gene_violin(object, gene, methylation_indices = NULL, return.plots = F)
}
\arguments{
\item{object}{analogizer object.}

\item{gene}{Gene for which to plot relative expression.}

\item{methylation_indices}{Indices of datasets in object with methylation data.}

\item{return.plots}{Return ggplot objects instead of printing directly}
}
\description{
Plot violin plots for expression of specified gene for each dataset
}
\examples{
\dontrun{
Y = matrix(c(1,2,3,4,5,6,7,8,9,10,11,12),nrow=4,byrow=T)
Z = matrix(c(1,2,3,4,5,6,7,6,5,4,3,2),nrow=4,byrow=T)
analogy = Analogizer(list(Y,Z))
analogy@var.genes = c(1,2,3,4)
analogy = scaleNotCenter(analogy)
}
}
