% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analogizer.R
\name{plot_genes}
\alias{plot_genes}
\title{Plot expression of multiple genes, each on a separate page.}
\usage{
plot_genes(object, genes)
}
\arguments{
\item{object}{analogizer object. Should call run_tSNE before calling.}

\item{genes}{vector of gene names.}
}
\description{
Plot expression of multiple genes, each on a separate page.
}
\examples{
\dontrun{
Y = matrix(c(1,2,3,4,5,6,7,8,9,10,11,12),nrow=4,byrow=T)
Z = matrix(c(1,2,3,4,5,6,7,6,5,4,3,2),nrow=4,byrow=T)
analogy = Analogizer(list(Y,Z))
analogy@var.genes = c(1,2,3,4)
analogy = scaleNotCenter(analogy)
}
}
