% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analogizer.R
\name{plot_word_clouds}
\alias{plot_word_clouds}
\title{Word cloud plots coloring t-SNE points by their loading on specifed factors as well as the
most highly loading shared and dataset-specific genes}
\usage{
plot_word_clouds(object, num_genes = 30, min_size = 1, max_size = 4,
  dataset1 = NULL, dataset2 = NULL)
}
\arguments{
\item{object}{analogizer object. Should call run_tSNE before calling.}

\item{num_genes}{Number of genes to show in word clouds}

\item{min_size}{Size of smallest gene symbol in word cloud}

\item{max_size}{Size of largest gene symbol in word cloud}

\item{dataset1}{Name of first dataset}

\item{dataset2}{Name of second dataset}
}
\description{
Word cloud plots coloring t-SNE points by their loading on specifed factors as well as the
most highly loading shared and dataset-specific genes
}
\examples{
\dontrun{
Y = matrix(c(1,2,3,4,5,6,7,8,9,10,11,12),nrow=4,byrow=T)
Z = matrix(c(1,2,3,4,5,6,7,6,5,4,3,2),nrow=4,byrow=T)
analogy = Analogizer(Y,Z)
analogy@var.genes = c(1,2,3,4)
analogy = scaleNotCenter(analogy)
analogy = optimize_als(analogy,k=2,nrep=1)
}
}
