% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplotting.R
\name{plotCellScatter}
\alias{plotCellScatter}
\title{Generate scatter plot(s) using liger object}
\usage{
plotCellScatter(
  object,
  x,
  y,
  colorBy = NULL,
  slot = c("cellMeta", "rawData", "normData", "scaleData", "H.norm", "H"),
  colorByFunc = NULL,
  splitBy = NULL,
  shapeBy = NULL,
  ...
)
}
\arguments{
\item{object}{A \linkS4class{liger} object}

\item{x, y}{Available variable name in \code{cellMeta} slot to look for
the dot coordinates. See details.}

\item{colorBy}{Available variable name in specified \code{slot} to look for
color annotation information. See details. Default \code{NULL} generates
all-black dots.}

\item{slot}{Choose the slot to find the \code{colorBy} variable. See details.
Default \code{"cellMeta"}.}

\item{colorByFunc}{Default \code{NULL}. A function object that expects a
vector/factor/data.frame retrieved by \code{colorBy} as the only input, and
returns an object of the same size, so that the all color "aes" are replaced
by this output. Useful when, for example, users need to scale the gene
expression shown on plot.}

\item{splitBy}{Character vector of categorical variable names in
\code{cellMeta} slot. Split all cells by groupings on this/these variable(s)
to produce a scatter plot containing only the cells in each group. Default
\code{NULL}.}

\item{shapeBy}{Available variable name in \code{cellMeta} slot to look for
categorical annotation to be reflected by dot shapes. Default \code{NULL}.}

\item{...}{More plot setting arguments. See \code{\link{.ggScatter}} and
\code{\link{.ggplotLigerTheme}}.}
}
\value{
A ggplot object when a single plot is intended. A list of ggplot
objects, when multiple \code{colorBy} variables and/or \code{splitBy} are
set. When \code{plotly = TRUE}, all ggplot objects become plotly (htmlwidget)
objects.
}
\description{
This function allows for using available cell metadata to build
the x-/y-axis. Available per-cell data can be used to form the color/shape
annotation, including cell metadata, raw or processed gene expression, and
unnormalized or aligned factor loading. Multiple coloring variable is allowed
from the same specification of \code{slot}, and this returns a list of plots
with different coloring values. Users can further split the plot(s) by
grouping on cells (e.g. datasets).
}
\details{
Available option for \code{slot} include: \code{"cellMeta"},
\code{"rawData"}, \code{"normData"}, \code{"scaleData"}, \code{"H.norm"}
and \code{"H"}. When \code{"rawData"}, \code{"normData"} or
\code{"scaleData"}, \code{colorBy} has to be a character vector of feature
names. When \code{"H.norm"} or \code{"H"}, \code{colorBy} can be any valid
index to select one factor of interests. Note that character index follows
\code{"Factor_[k]"} format, with replacing \code{[k]} with an integer.

When \code{"cellMeta"}, \code{colorBy} has to be an available column name in
the table. Note that, for \code{colorBy} as well as \code{x}, \code{y},
\code{shapeBy} and \code{splitBy}, since a matrix object is feasible in
\code{cellMeta} table, using a column (e.g. named as \code{"column1"} in a
certain matrix (e.g. named as \code{"matrixVar"}) should follow the syntax of
\code{"matrixVar.column1"}. When the matrix does not have a "colname"
attribute, the subscription goes with \code{"matrixVar.V1"},
\code{"matrixVar.V2"} and etc. Use \code{"UMAP.1"}, \code{"UMAP.2"},
\code{"TSNE.1"} or \code{"TSNE.2"} for the 2D embeddings generated with
rliger package. These are based on the nature of \code{as.data.frame} method
on a \code{\link[S4Vectors]{DataFrame}} object.
}
\examples{
data("pbmcPlot", package = "rliger")
plotCellScatter(pbmcPlot, x = "UMAP.1", y = "UMAP.2",
                colorBy = "dataset", slot = "cellMeta",
                labelText = FALSE)
plotCellScatter(pbmcPlot, x = "UMAP.1", y = "UMAP.2",
                colorBy = "S100A8", slot = "normData",
                dotOrder = "ascending", dotSize = 2)
plotCellScatter(pbmcPlot, x = "UMAP.1", y = "UMAP.2",
                colorBy = 2, slot = "H.norm",
                dotOrder = "ascending", dotSize = 2,
                colorPalette = "viridis")
}
