% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization2.R
\name{plotClusterDimRed}
\alias{plotClusterDimRed}
\alias{plotDatasetDimRed}
\alias{plotByDatasetAndCluster}
\alias{plotGeneDimRed}
\alias{plotFactorDimRed}
\title{Generate Dimensionality Reduction Plot with Coloring}
\usage{
plotClusterDimRed(
  object,
  useCluster = "louvain_cluster",
  useDimRed = "UMAP",
  ...
)

plotDatasetDimRed(object, useDimRed = "UMAP", ...)

plotByDatasetAndCluster(
  object,
  useDimRed = "UMAP",
  useCluster = "louvain_cluster",
  combinePlots = TRUE,
  ...
)

plotGeneDimRed(
  object,
  features,
  useDimRed = "UMAP",
  log = TRUE,
  scaleFactor = 10000,
  zeroAsNA = TRUE,
  colorPalette = "C",
  ...
)

plotFactorDimRed(
  object,
  factors,
  useDimRed = "UMAP",
  trimHigh = 0.03,
  zeroAsNA = TRUE,
  colorPalette = "D",
  ...
)
}
\arguments{
\item{object}{A \linkS4class{liger} object.}

\item{useCluster}{Name of variable in \code{cellMeta} slot. Default
\code{"louvain_cluster"}.}

\item{useDimRed}{Name of the variable storing dimensionality reduction result
in the \code{cellMeta} slot. Default \code{"UMAP"}.}

\item{...}{Additional graphic setting arguments passed to
\code{\link{plotCellScatter}}.}

\item{combinePlots}{Logical, whether to utilize
\code{\link[cowplot]{plot_grid}} to combine multiple plots into one. Default
\code{TRUE} returns combined ggplot. \code{FALSE} returns a list of ggplot.}

\item{features, factors}{Name of genes or index of factors that need to be
visualized.}

\item{log}{Logical. Whether to log transform the normalized expression of
genes. Default \code{TRUE}.}

\item{scaleFactor}{Number to be multiplied with the normalized expression of
genes before log transformation. Default \code{1e4}. \code{NULL} for not
scaling.}

\item{zeroAsNA}{Logical, whether to swap all zero values to \code{NA} so
\code{naColor} will be used to represent non-expressing features. Default
\code{TRUE}.}

\item{colorPalette}{Name of viridis palette. See
\code{\link[viridisLite]{viridis}} for options. Default \code{"C"} ("plasma")
for gene expression and \code{"D"} ("viridis") for factor loading.}

\item{trimHigh}{Number for highest cut-off to limit the outliers. Factor
loading above this value will all be trimmed to this value. Default
\code{0.03}.}
}
\value{
ggplot object when only one feature (e.g. cluster variable, gene,
factor) is set. List object when multiple of those are specified.
}
\description{
some text
}
\seealso{
Please refer to \code{\link{plotCellScatter}},
\code{\link{.ggScatter}}, \code{\link{.ggplotLigerTheme}} for additional
graphic setting
}
