% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liger.R
\name{plotClusterProportions}
\alias{plotClusterProportions}
\title{Plot cluster proportions by dataset}
\usage{
plotClusterProportions(object, return.plot = F)
}
\arguments{
\item{object}{\code{liger} object. Should call quantileAlignSNF before calling.}

\item{return.plot}{Return ggplot object (default FALSE)}
}
\description{
Generates plot of clusters sized by the proportion of total cells
}
\examples{
\dontrun{
# liger object, factorization complete 
ligerex
ligerex <- quantileAlignSNF(ligerex)
# plot cluster proportions
plotClusterProportions(ligerex)
}
}
