% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplotting.R, R/visualization.R
\name{plotDimRed}
\alias{plotDimRed}
\alias{plotClusterDimRed}
\alias{plotDatasetDimRed}
\alias{plotByDatasetAndCluster}
\alias{plotGeneDimRed}
\alias{plotPeakDimRed}
\alias{plotFactorDimRed}
\title{Generate scatter plot(s) using liger object}
\usage{
plotDimRed(
  object,
  colorBy = NULL,
  useDimRed = NULL,
  slot = c("cellMeta", "rawData", "normData", "scaleData", "H.norm", "H", "normPeak",
    "rawPeak"),
  colorByFunc = NULL,
  cellIdx = NULL,
  splitBy = NULL,
  shapeBy = NULL,
  titles = NULL,
  axisArrows = TRUE,
  ...
)

plotClusterDimRed(object, useCluster = NULL, useDimRed = NULL, ...)

plotDatasetDimRed(object, useDimRed = NULL, ...)

plotByDatasetAndCluster(
  object,
  useDimRed = NULL,
  useCluster = NULL,
  combinePlot = TRUE,
  ...
)

plotGeneDimRed(
  object,
  features,
  useDimRed = NULL,
  log = TRUE,
  scaleFactor = 10000,
  zeroAsNA = TRUE,
  colorPalette = "C",
  ...
)

plotPeakDimRed(
  object,
  features,
  useDimRed = NULL,
  log = TRUE,
  scaleFactor = 10000,
  zeroAsNA = TRUE,
  colorPalette = "C",
  ...
)

plotFactorDimRed(
  object,
  factors,
  useDimRed = NULL,
  trimHigh = 0.03,
  zeroAsNA = TRUE,
  colorPalette = "D",
  ...
)
}
\arguments{
\item{object}{A \linkS4class{liger} object.}

\item{colorBy}{Available variable name in specified \code{slot} to look for
color annotation information. See details. Default \code{NULL} generates
all-black dots.}

\item{useDimRed}{Name of the variable storing dimensionality reduction result
in the \code{cellMeta(object)}. Default \code{NULL} use default dimRed.}

\item{slot}{Choose the slot to find the \code{colorBy} variable. See details.
Default \code{"cellMeta"}.}

\item{colorByFunc}{Default \code{NULL}. A function object that expects a
vector/factor/data.frame retrieved by \code{colorBy} as the only input, and
returns an object of the same size, so that the all color "aes" are replaced
by this output. Useful when, for example, users need to scale the gene
expression shown on plot.}

\item{cellIdx}{Character, logical or numeric index that can subscribe cells.
Missing or \code{NULL} for all cells.}

\item{splitBy}{Character vector of categorical variable names in
\code{cellMeta} slot. Split all cells by groupings on this/these variable(s)
to produce a scatter plot containing only the cells in each group. Default
\code{NULL}.}

\item{shapeBy}{Available variable name in \code{cellMeta} slot to look for
categorical annotation to be reflected by dot shapes. Default \code{NULL}.}

\item{titles}{Title text. A character scalar or a character vector with as
many elements as multiple plots are supposed to be generated. Default
\code{NULL}.}

\item{axisArrows}{Whether to draw short fixed-length arrows for axis. Default
\code{FALSE}.}

\item{...}{
  Arguments passed on to \code{\link[=.ggScatter]{.ggScatter}}, \code{\link[=.ggplotLigerTheme]{.ggplotLigerTheme}}
  \describe{
    \item{\code{dotOrder}}{Controls the order that each dot is added to the plot. Choose
from \code{"shuffle"}, \code{"ascending"}, or \code{"descending"}. Default
\code{"shuffle"}, useful when coloring by categories that overlaps (e.g.
"dataset"), \code{"ascending"} can be useful when coloring by a continuous
variable (e.g. gene expression) where high values needs more
highlight. \code{NULL} use default order.}
    \item{\code{dotSize,dotAlpha}}{Numeric, controls the size or transparency of all
dots. Default \code{getOption("ligerDotSize")} (1) and \code{0.9}.}
    \item{\code{trimHigh,trimLow}}{Numeric, limit the largest or smallest value of
continuous \code{colorBy} variable. Default \code{NULL}.}
    \item{\code{raster}}{Logical, whether to rasterize the plot. Default \code{NULL}
automatically rasterize the plot when number of total dots to be plotted
exceeds 100,000.}
    \item{\code{labelBy}}{A variable name available in \code{plotDF}. If the variable is
categorical (a factor), the label position will be the median coordinates of
all dots within the same group. Unique labeling in character vector for each
dot is also acceptable. Default \code{colorBy}.}
    \item{\code{labelText}}{Logical, whether to show text label at the median position
of each categorical group specified by \code{colorBy}. Default \code{TRUE}.
Does not work when continuous coloring is specified.}
    \item{\code{labelTextSize}}{Numeric, controls the size of label size when
\code{labelText = TRUE}. Default \code{4}.}
    \item{\code{seed}}{Random seed for reproducibility. Default \code{1}.}
    \item{\code{title,subtitle,xlab,ylab}}{Main title, subtitle or X/Y axis title text.
By default, no main title or subtitle will be set, and X/Y axis title will be
the names of variables used for plotting. Use \code{NULL} to hide elements.
\code{TRUE} for \code{xlab} or \code{ylab} shows default values.}
    \item{\code{legendColorTitle}}{Legend title text for color aesthetics, often used
for categorical or continuous coloring of dots. Default \code{NULL} shows the
original variable name.}
    \item{\code{legendShapeTitle}}{Legend title text for shape aesthetics, often used
for shaping dots by categorical variable. Default \code{NULL} shows the
original variable name.}
    \item{\code{showLegend}}{Whether to show the legend. Default \code{TRUE}.}
    \item{\code{legendPosition}}{Text indicating where to place the legend. Choose from
\code{"top"}, \code{"bottom"}, \code{"left"} or \code{"right"}. Default
\code{"right"}.}
    \item{\code{baseSize}}{One-parameter control of all text sizes. Individual text
element sizes can be controlled by other size arguments. "Title" sizes are
2 points larger than "text" sizes when being controlled by this.}
    \item{\code{titleSize,xTitleSize,yTitleSize,legendTitleSize}}{Size of main title,
axis titles and legend title. Default \code{NULL} controls by
\code{baseSize + 2}.}
    \item{\code{subtitleSize,xTextSize,yTextSize,legendTextSize}}{Size of subtitle text,
axis texts and legend text. Default \code{NULL} controls by \code{baseSize}.}
    \item{\code{legendDotSize}}{Allow dots in legend region to be large enough to see
the colors/shapes clearly. Default \code{4}.}
    \item{\code{showAxis}}{Whether to show axis lines or arrows. Default \code{TRUE}.}
    \item{\code{panelBorder}}{Whether to show rectangle border of the panel instead of
using ggplot classic bottom and left axis lines. Default \code{FALSE}.}
    \item{\code{colorLabels}}{Character vector for modifying category names in a
color legend. Passed to \code{ggplot2::scale_color_manual(labels)}. Default
\code{NULL} uses original levels of the factor.}
    \item{\code{colorValues}}{Character vector of colors for modifying category colors
in a color legend. Passed to \code{ggplot2::scale_color_manual(values)}.
Default \code{NULL} uses internal selected palette when <= 26 categories are
presented, otherwise ggplot hues.}
    \item{\code{legendNRow,legendNCol}}{Integer, when too many categories in one
variable, arranges number of rows or columns. Default \code{NULL},
automatically split to \code{ceiling(levels(variable)/15)} columns.}
    \item{\code{colorDirection}}{Choose \code{1} or \code{-1}. Applied when
\code{colorPalette} is from Viridis options. Default \code{-1} use darker
color for higher value, while \code{1} reverses this direction.}
    \item{\code{colorLow,colorMid,colorHigh,colorMidPoint}}{All four of these must be
specified to customize palette with}
    \item{\code{naColor}}{The color code for \code{NA} values. Default \code{"#DEDEDE"}.
\code{\link[ggplot2]{scale_colour_gradient2}}. Default \code{NULL}.}
    \item{\code{plotly}}{Whether to use plotly to enable web based interactive browsing
for the plot. Requires installation of package "plotly". Default
\code{FALSE}.}
  }}

\item{useCluster}{Name of variable in \code{cellMeta(object)}. Default
\code{NULL} uses default cluster.}

\item{combinePlot}{Logical, whether to utilize
\code{\link[cowplot]{plot_grid}} to combine multiple plots into one. Default
\code{TRUE} returns combined ggplot. \code{FALSE} returns a list of ggplot.}

\item{features, factors}{Name of genes or index of factors that need to be
visualized.}

\item{log}{Logical. Whether to log transform the normalized expression of
genes. Default \code{TRUE}.}

\item{scaleFactor}{Number to be multiplied with the normalized expression of
genes before log transformation. Default \code{1e4}. \code{NULL} for not
scaling.}

\item{zeroAsNA}{Logical, whether to swap all zero values to \code{NA} so
\code{naColor} will be used to represent non-expressing features. Default
\code{TRUE}.}

\item{colorPalette}{Name of viridis palette. See
\code{\link[viridisLite]{viridis}} for options. Default \code{"C"} ("plasma")
for gene expression and \code{"D"} ("viridis") for factor loading.}

\item{trimHigh}{Number for highest cut-off to limit the outliers. Factor
loading above this value will all be trimmed to this value. Default
\code{0.03}.}
}
\value{
A ggplot object when a single plot is intended. A list of ggplot
objects, when multiple \code{colorBy} variables and/or \code{splitBy} are
set. When \code{plotly = TRUE}, all ggplot objects become plotly (htmlwidget)
objects.

ggplot object when only one feature (e.g. cluster variable, gene,
factor) is set. List object when multiple of those are specified.
}
\description{
This function allows for using available cell metadata to build
the x-/y-axis. Available per-cell data can be used to form the color/shape
annotation, including cell metadata, raw or processed gene expression, and
unnormalized or aligned factor loading. Multiple coloring variable is allowed
from the same specification of \code{slot}, and this returns a list of plots
with different coloring values. Users can further split the plot(s) by
grouping on cells (e.g. datasets).
}
\details{
Available option for \code{slot} include: \code{"cellMeta"},
\code{"rawData"}, \code{"normData"}, \code{"scaleData"}, \code{"H.norm"}
and \code{"H"}. When \code{"rawData"}, \code{"normData"} or
\code{"scaleData"}, \code{colorBy} has to be a character vector of feature
names. When \code{"H.norm"} or \code{"H"}, \code{colorBy} can be any valid
index to select one factor of interests. Note that character index follows
\code{"Factor_[k]"} format, with replacing \code{[k]} with an integer.

When \code{"cellMeta"}, \code{colorBy} has to be an available column name in
the table. Note that, for \code{colorBy} as well as \code{x}, \code{y},
\code{shapeBy} and \code{splitBy}, since a matrix object is feasible in
\code{cellMeta} table, using a column (e.g. named as \code{"column1"} in a
certain matrix (e.g. named as \code{"matrixVar"}) should follow the syntax of
\code{"matrixVar.column1"}. When the matrix does not have a "colname"
attribute, the subscription goes with \code{"matrixVar.V1"},
\code{"matrixVar.V2"} and etc. Use \code{"UMAP.1"}, \code{"UMAP.2"},
\code{"TSNE.1"} or \code{"TSNE.2"} for the 2D embeddings generated with
rliger package. These are based on the nature of \code{as.data.frame} method
on a \code{\link[S4Vectors]{DataFrame}} object.
}
\examples{
plotDimRed(pbmcPlot, colorBy = "dataset", slot = "cellMeta",
           labelText = FALSE)
plotDimRed(pbmcPlot, colorBy = "S100A8", slot = "normData",
           dotOrder = "ascending", dotSize = 2)
plotDimRed(pbmcPlot, colorBy = 2, slot = "H.norm",
           dotOrder = "ascending", dotSize = 2, colorPalette = "viridis")
plotClusterDimRed(pbmcPlot)
plotDatasetDimRed(pbmcPlot)
plotByDatasetAndCluster(pbmcPlot)
plotGeneDimRed(pbmcPlot, varFeatures(pbmcPlot)[1])
plotFactorDimRed(pbmcPlot, 2)
}
