% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analogizer.R
\name{plotGene}
\alias{plotGene}
\title{Plot t-SNE coordinates by expression of specified gene}
\usage{
plotGene(object, gene, methylation.indices = NULL, pt.size = 0.1,
  min.clip = 0, max.clip = 1, points.only = F, low.col = "yellow",
  high.col = "red", return.plots = F)
}
\arguments{
\item{object}{Analogizer object. Should call runTSNE before calling.}

\item{gene}{Gene for which to plot relative expression.}

\item{methylation.indices}{Indices of datasets in object with methylation data (this data is not
magnified and put on log scale).}

\item{pt.size}{Point size for plots (default 0.1)}

\item{min.clip}{Quantile probability for lower bound of methylation data (everything lower set 
to this quantile value) (default 0)}

\item{max.clip}{Quantile probability for upper bound of methylation data (everything greater set 
to this quantile value) (default 1)}

\item{points.only}{Remove axes when plotting t-sne coordinates (default FALSE).}

\item{low.col}{Color to plot for lowest gene expression (default "yellow").}

\item{high.col}{Color to plot for highest gene expression (default "red").}

\item{return.plots}{Return ggplot objects instead of printing directly (default FALSE).}
}
\description{
Generates plot of t-SNE coordinates colored by expression of specified gene, for each dataset.
Color scale can be modified.
}
\examples{
\dontrun{
# analogizer object, factorization complete 
analogy
analogy <- runTSNE(analogy)
# plot expression for CD4 and return plots
gene_plots <- plotGene(analogy, "CD4", return.plots = T)
}
}
