% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGeneLoadings.R
\name{plotGeneLoadings}
\alias{plotGeneLoadings}
\alias{plotGeneLoadingRank}
\title{Visualize factor expression and gene loading}
\usage{
plotGeneLoadings(
  object,
  markerTable,
  useFactor,
  useDimRed = "UMAP",
  topNPval = 12,
  topNLoad = 30,
  ...
)

plotGeneLoadingRank(
  object,
  markerTable,
  useFactor,
  topNPval = 12,
  topNLoad = 30,
  ...
)
}
\description{
Visualize factor expression and gene loading
}
