% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analogizer.R
\name{plotGenes}
\alias{plotGenes}
\title{Plot expression of multiple genes}
\usage{
plotGenes(object, genes)
}
\arguments{
\item{object}{Analogizer object. Should call runTSNE before calling.}

\item{genes}{Vector of gene names.}
}
\description{
Uses plotGene to plot each gene (and dataset) on a separate page. It is recommended to call this
function into a PDF due to the large number of plots produced.
}
\examples{
\dontrun{
# analogizer object, factorization complete 
analogy
analogy <- runTSNE(analogy)
# plot expression for CD4 and FCGR3A
pdf("gene_plots.pdf")
plotGenes(analogy, c("CD4", "FCGR3A"))
dev.off()
}
}
