% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization2.R
\name{plotGeneViolin}
\alias{plotGeneViolin}
\title{Visualize gene expression with violin plot}
\usage{
plotGeneViolin(object, gene, byDataset = TRUE, groupBy = NULL, ...)
}
\arguments{
\item{object}{A \linkS4class{liger} object.}

\item{gene}{Character gene names.}

\item{byDataset}{Logical, whether the violin plot should be splitted by
dataset. Default \code{TRUE}.}

\item{groupBy}{Names of available categorical variable in \code{cellMeta}
slot. Use \code{FALSE} for no grouping. Default \code{NULL} looks clustering
result but will not group if no clustering found.}

\item{...}{Additional arguments passed to \code{\link{plotCellViolin}}.}
}
\value{
ggplot if using a single gene and not splitting by dataset.
Otherwise, list of ggplot.
}
\description{
Visualize gene expression with violin plot
}
