% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plotSankey}
\alias{plotSankey}
\alias{makeRiverplot}
\title{Make Riverplot/Sankey plot for showing label mapping}
\usage{
plotSankey(
  object,
  cluster1,
  cluster2,
  clusterConsensus = NULL,
  minFrac = 0.01,
  minCell = 10,
  titles = NULL,
  labelSize = getOption("ligerBaseSize"),
  titleSize = getOption("ligerBaseSize") + 2,
  colorValues = scPalette
)

makeRiverplot(object, cluster1, cluster2)
}
\arguments{
\item{object}{A \linkS4class{liger} object with all three clustering
variables available.}

\item{cluster1, cluster2}{Name of the variables in \code{cellMeta(object)} for
the cluster assignments of dataset 1 and 2, respectively.}

\item{clusterConsensus}{Name of the joint cluster variable to use. Default
uses the default clustering of the object. Can select a variable name in
\code{cellMeta(object)}.}

\item{minFrac}{Numeric. Minimum fraction of cluster for a node to be shown.
Default \code{0.05}.}

\item{minCell}{Numeric. Minimum number of cells for a node to be shown.
Default \code{10}.}

\item{titles}{Character vector of three. Customizes the column text shown.
Default uses the variable names \code{cluster1}, \code{clusterConsensus} and
\code{cluster2}.}

\item{labelSize}{Numeric. Size of the text label of the cluster node in
pixel. Default uses user global option \code{getOption("ligerBaseSize")}
which by default is \code{10}.}

\item{titleSize}{Numeric. Size of the text label of the column title in
pixel. Default uses user global option \code{getOption("ligerBaseSize") + 2}
which by default is \code{12}.}

\item{colorValues}{Character vector of color codes to set color for each
level in the consensus clustering. Default \code{scPalette}.}
}
\value{
List object of class "sankeyNetwork" and "htmlwidget". Printing
the returned object shows interactive view of the network when running on
device with display capability.
}
\description{
Creates a riverplot/Sankey plot to show how independent cluster assignments
from two datasets map onto a joint clustering. Prior knowledge of cell
annotation for the given datasets is required to make sense from the
visualization. Dataset original annotation can be added with the syntax shown
in example code in this manual. The joint clustering could be generated with
\code{\link{runCluster}} or set by any other metadata annotation.
}
\note{
This function works as a replacement of the function \code{makeRiverplot}
in rliger <1.99. We decide to make a new function because the dependency
adopted by the older version is archived on CRAN and will be no longer
available.
}
\examples{
# Make fake dataset specific labels from joint clustering result
cellMeta(pbmcPlot, "ctrl_cluster", "ctrl") <-
    cellMeta(pbmcPlot, "leiden_cluster", "ctrl")
cellMeta(pbmcPlot, "stim_cluster", "stim") <-
    cellMeta(pbmcPlot, "leiden_cluster", "stim")
plotSankey(pbmcPlot, "ctrl_cluster", "stim_cluster",
           titles = c("control", "LIGER", "stim"))
}
