% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suggestParam.R
\name{plotSuggestK}
\alias{plotSuggestK}
\title{Visualize suggestK result statistics}
\usage{
plotSuggestK(
  result,
  xlab = "k",
  ylab = "Median KL divergence (across all cells)",
  legendColorTitle = "",
  ...
)
}
\arguments{
\item{result}{Returned data.frame from \code{\link{suggestK}}.}

\item{xlab, ylab}{The title on x-/y-axis of the plot. See Usage for default.}

\item{legendColorTitle}{The title of legend. Default \code{""}.}

\item{...}{Additional plot theme setting passed to
\code{\link{.ggplotLigerTheme}}.}
}
\value{
ggplot
}
\description{
Visualize suggestK result statistics
}
