% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suggestParam.R
\name{plotSuggestLambda}
\alias{plotSuggestLambda}
\title{Visualize suggestLambda result statistics}
\usage{
plotSuggestLambda(
  result,
  xlab = "Lambda",
  ylab = "Alignment",
  legendColorTitle = "",
  ...
)
}
\arguments{
\item{result}{Returned data.frame from \code{\link{suggestLambda}}.}

\item{xlab, ylab}{The title on x-/y-axis of the plot. See Usage for default.}

\item{legendColorTitle}{The title of legend. Default \code{""}.}

\item{...}{Additional plot theme setting passed to
\code{\link{.ggplotLigerTheme}}.}
}
\value{
ggplot
}
\description{
Visualize suggestLambda result statistics
}
