% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectGenes.R
\name{plotVarFeatures}
\alias{plotVarFeatures}
\title{Plot the variance vs mean of feature expression}
\usage{
plotVarFeatures(
  object,
  combinePlot = TRUE,
  dotSize = 0.8,
  legendDotSize = 4,
  titleSize = 14,
  subTitleSize = 12,
  textSize = 10
)
}
\arguments{
\item{object}{\linkS4class{liger} object. \code{\link{selectGenes}} need to
run in advance.}

\item{combinePlot}{Logical. If \code{TRUE}, sub-figures for all datasets will
be combined into one plot. if \code{FALSE}, a list of plots will be returned.
Default \code{TRUE}.}

\item{dotSize}{Controls the size of dots in the main plot. Default
\code{0.8}.}

\item{legendDotSize}{Controls the size of dots in the legend. Default
\code{4}.}

\item{titleSize}{Controls the size of the main title of each sub-plot.
Default \code{14}.}

\item{subTitleSize}{Controls the size of subtitle, axis title and legend
title. Default \code{12}.}

\item{textSize}{Controls the size of axis text and legend text. Default
\code{10}.}
}
\value{
\code{ggplot} object when \code{combinePlot = TRUE}, a list of
\code{ggplot} objects when \code{combinePlot = FALSE}
}
\description{
For each dataset where the feature variablitity is calculated,
a plot of log10 feature expression variance and log10 mean will be produced.
Features that are considered as variable would be highlighted in red.
}
