% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liger.R
\name{plotWilcoxon}
\alias{plotWilcoxon}
\title{Generate heatmap for visulization of Wilconxon rank sum test results. Display top expressed genes for each cluster.}
\usage{
plotWilcoxon(object, num_top = 10)
}
\arguments{
\item{object}{\code{liger} object. After \code{calculateWilcoxon} is completed.}

\item{num_top}{number of top expressed genes to be displayed (default=10)}
}
\value{
Updated \code{liger} object.
}
\description{
Generate heatmap for visulization of Wilconxon rank sum test results. Display top expressed genes for each cluster.
}
\examples{
\dontrun{
# liger object, Wilcoxon rank sum test complete 
ligerex
# fill wilcoxon slot with dataframe of information from presto's Wilcoxon rank sum test
p <- plotWilcoxon(ligerex)
}

}
