% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorGSEA.R
\name{print.factorGSEA}
\alias{print.factorGSEA}
\title{Show information about factorGSEA object}
\usage{
\method{print}{factorGSEA}(x, ...)
}
\arguments{
\item{x}{A \code{factorGSEA} object.}

\item{...}{S3 method convention, not used for now.}
}
\description{
Show information about factorGSEA object
}
