% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantileNorm.R
\name{quantile_norm,liger-method}
\alias{quantile_norm,liger-method}
\alias{quantile_norm,list-method}
\title{Quantile Normalize}
\usage{
\S4method{quantile_norm}{liger}(
  object,
  quantiles = 50,
  ref_dataset = NULL,
  min_cells = 20,
  knn_k = 20,
  dims.use = NULL,
  do.center = FALSE,
  max_sample = 1000,
  eps = 0.9,
  refine.knn = TRUE,
  rand.seed = 1
)

\S4method{quantile_norm}{list}(
  object,
  quantiles = 50,
  ref_dataset = NULL,
  min_cells = 20,
  knn_k = 20,
  dims.use = NULL,
  do.center = FALSE,
  max_sample = 1000,
  eps = 0.9,
  refine.knn = TRUE,
  rand.seed = 1
)
}
\description{
Quantile Normalize
}
