% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantileNorm.R
\name{quantileNorm}
\alias{quantileNorm}
\alias{quantileNorm,liger-method}
\alias{quantileNorm,list-method}
\title{Quantile Align (Normalize) Factor Loadings}
\usage{
quantileNorm(
  object,
  quantiles = 50,
  reference = NULL,
  minCells = 20,
  nNeighbors = 20,
  useDims = NULL,
  center = FALSE,
  maxSample = 1000,
  eps = 0.9,
  refineKNN = TRUE,
  clusterName = "H.norm_cluster",
  seed = 1,
  verbose = getOption("ligerVerbose")
)

\S4method{quantileNorm}{liger}(
  object,
  quantiles = 50,
  reference = NULL,
  minCells = 20,
  nNeighbors = 20,
  useDims = NULL,
  center = FALSE,
  maxSample = 1000,
  eps = 0.9,
  refineKNN = TRUE,
  clusterName = "H.norm_cluster",
  seed = 1,
  verbose = getOption("ligerVerbose")
)

\S4method{quantileNorm}{list}(
  object,
  quantiles = 50,
  reference = NULL,
  minCells = 20,
  nNeighbors = 20,
  useDims = NULL,
  center = FALSE,
  maxSample = 1000,
  eps = 0.9,
  refineKNN = TRUE,
  seed = 1,
  verbose = getOption("ligerVerbose")
)
}
\arguments{
\item{object}{A \linkS4class{liger} object with \code{\link{optimizeALS}} run
in advance. Or a list object of \eqn{H} matrices, all in \eqn{k * nCell}
dimensionality, the \code{H} entry of the output of \code{\link{optimizeALS}}
list method is ready for use (i.e. use \code{out$H}).}

\item{quantiles}{Number of quantiles to use for quantile normalization.
Default \code{50}.}

\item{reference}{Character, numeric or logical selection of one dataset, out
of all available datasets in \code{object}, to use as a "reference" for
normalization. Default \code{NULL} use the dataset with the largest number of
cells.}

\item{minCells}{Minimum number of cells to consider a cluster shared across
datasets. Default \code{20}.}

\item{nNeighbors}{Number of nearest neighbors for within-dataset knn graph.
Default \code{20}.}

\item{useDims}{Indices of factors to use for shared nearest factor
determination. Default \code{NULL} uses all factors.}

\item{center}{Whether to center the data when scaling factors (centering
could be useful for less sparse modalities like methylation data). Default
\code{FALSE}.}

\item{maxSample}{Maximum number of cells used for quantile normalization of
each cluster and factor. Default \code{1000}.}

\item{eps}{The error bound of the nearest neighbor search. Lower values give
more accurate nearest neighbor graphs but take much longer to compute.
Default \code{0.9}.}

\item{refineKNN}{whether to increase robustness of cluster assignments using
KNN graph. Default \code{TRUE}.}

\item{clusterName}{of the variable that will store the clustering result
in \code{cellMeta} slot of \code{object}. Default \code{"H.norm_cluster"}.}

\item{seed}{Random seed to allow reproducible results. Default \code{1}.}

\item{verbose}{Logical. Whether to show information of the progress. Default
\code{getOption("ligerVerbose")} which is \code{TRUE} if users have not set.}
}
\value{
For method on \linkS4class{liger} object, the object will be returned
with the \code{H.norm} slot and \code{clusterName} variable in
\code{cellMeta} slot updated. For method on list, a result list with entries
of \code{H.norm} and \code{clusters} will be returned.
}
\description{
This process builds a shared factor neighborhood graph to
jointly cluster cells, then quantile normalizes corresponding clusters.

The first step, building the shared factor neighborhood graph, is performed
in SNF(), and produces a graph representation where edge weights between
cells (across all datasets) correspond to their similarity in the shared
factor neighborhood space. An important parameter here is \code{knn_k}, the
number of neighbors used to build the shared factor space.

Next we perform quantile alignment for each dataset, factor, and cluster (by
stretching/compressing datasets' quantiles to better match those of the
reference dataset). These aligned factor loadings are combined into a single
matrix and returned as "H.norm".

This function is a replacement of \code{quantile_norm()}, which will be
deprecated. See \code{help("quantile_norm-deprecated")}.
}
\examples{
data("pbmcPlot", package = "rliger")
pbmc <- quantileNorm(pbmcPlot)
}
