% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ligerDataset-class.R
\name{raw.data,ligerDataset-method}
\alias{raw.data,ligerDataset-method}
\title{Access raw.data of ligerDataset object}
\usage{
\S4method{raw.data}{ligerDataset}(x)
}
\arguments{
\item{x}{ligerDataset object}
}
\description{
Access raw.data of ligerDataset object
}
