% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ligerDataset-class.R
\name{raw.data<-,ligerDataset-method}
\alias{raw.data<-,ligerDataset-method}
\title{Set raw.data to ligerDataset object}
\usage{
\S4method{raw.data}{ligerDataset}(x) <- value
}
\arguments{
\item{x}{ligerDataset object}

\item{value}{dgCMatrix}
}
\description{
Set raw.data to ligerDataset object
}
