% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tenX.R
\name{read10X}
\alias{read10X}
\title{Read 10X alignment data (including V3)}
\usage{
read10X(
  sampleDirs,
  sampleNames,
  maxNCells = NULL,
  useFiltered = FALSE,
  reference = NULL,
  modal = "rna",
  verbose = getOption("ligerVerbose"),
  sample.dirs = sampleDirs,
  sample.names = sampleNames,
  num.cells = maxNCells,
  use.filtered = useFiltered,
  data.type = modal,
  merge = FALSE,
  min.umis = 0
)
}
\arguments{
\item{sampleDirs}{Character vector of directory paths to either
matrix.mtx(.gz) file along with genes.tsv (features.tsv.gz) and
barcodes.tsv(.gz), or 10X sample output directory (which contains "outs/").}

\item{sampleNames}{Character vector of names to use for each sample. Must
match \code{sampleDirs}.}

\item{maxNCells}{Numeric, limit on number of cells returned for each sample
(only for Gene Expression data). Retains the cells with the highest numbers
of transcripts. Default \code{NULL}.}

\item{useFiltered}{Logical, whether to use 10X's filtered data (as opposed to
raw). Only relevant if \code{sampleDirs} are 10X sample output directories (
which contain "outs/"). Default \code{FALSE}.}

\item{reference}{For 10X V<3, which reference directory to use if
\code{sampleDirs} is 10X sample output directory. Only use if more than one
reference is used for alignment. Default \code{NULL}.}

\item{modal}{Choose from \code{"rna"} or \code{"atac"}. The protocol of the
input data. If not specified, input data will be considered scRNA-seq data.
Default \code{"rna"}.}

\item{verbose}{Logical. Whether to show information of the progress. Default
\code{getOption("ligerVerbose")} which is \code{TRUE} if users have not set.}

\item{sample.dirs, sample.names, num.cells, use.filtered, data.type}{\bold{Deprecated}. See Usage section for replacement.}

\item{merge}{\bold{Defuncted}. Whether to merge all matrices of the same data
type across samples or leave as list of matrices (default TRUE).}

\item{min.umis}{\bold{Defuncted}. Minimum UMI threshold for cells (default
0).}
}
\value{
List of samples. Each sample has a list of matrices across data
types.
}
\description{
This function generates a sparse matrix (genes x cells) from the data
generated by 10X's cellranger count pipeline. It can process V2 and V3 data
together, producing either a single merged matrix or list of matrices. Also
handles multiple data types produced by 10X V3 (Gene Expression, Antibody
Capture, CRISPR, CUSTOM).
}
\examples{
if (FALSE) {
    # 10X output directory V2 --
    #  contains outs/raw_gene_bc_matrices/<reference>/...
    sample.dir1 <- "home/welchjd/run/sample1" # folder "outs/" inside
    # 10X output directory V3 --
    #  for two data types, Gene Expression and CUSTOM
    sample.dir2 <- "home/welchjd/run/sample2" # folder "outs/" inside
    dataList <- read10X(sampleDirs = c(sample.dir1, sample.dir2),
                        sampleNames = c("sample1", "sample2"))
    lig <- createLiger(list(sample1 = dataList$sample1[["Gene Expression"]],
                            sample2 = dataList$sample2[["Gene Expression"]]))
}
}
