% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{readLiger}
\alias{readLiger}
\title{Read liger object from RDS file}
\usage{
readLiger(
  filename,
  dimredName = "tsne_coords",
  clusterName = "clusters",
  update = TRUE
)
}
\arguments{
\item{filename}{Path to an RDS file of a \code{liger} object of old versions.}

\item{dimredName}{The name of variable in \code{cellMeta} slot to store the
dimensionality reduction matrix, which originally located in
\code{tsne.coords} slot. Default \code{"tsne.coords"}.}

\item{clusterName}{The name of variable in \code{cellMeta} slot to store the
clustering assignment, which originally located in \code{clusters} slot.
Default \code{"clusters"}.}

\item{update}{Logical, whether to update an old (<=1.0.0) \code{liger} object
to the currect version of structure. Default \code{TRUE}.}
}
\value{
New version of \linkS4class{liger} object
}
\description{
Read liger object from RDS file
}
