% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liger.R
\name{readSubset}
\alias{readSubset}
\title{Sample data for plotting}
\usage{
readSubset(object, slot.use = "norm.data", balance = NULL,
  max.cells = 1000, chunk = 1000, datasets.use = NULL,
  genes.use = NULL, rand.seed = 1)
}
\arguments{
\item{object}{\code{liger} object. Should call normalize and selectGenes before calling.}

\item{slot.use}{Type of data for sampling (raw.data, norm.data(default), scale.data).}

\item{balance}{Type of sampling. NULL means that max_cells are sampled from among all cells;
balance="dataset" samples up to max_cells from each dataset;
balance="cluster" samples up to max_cells from each cluster.}

\item{max.cells}{Total number of cell to sample (default 5000).}

\item{chunk}{is the max number of cells at a time to read from disk (default 1000).}

\item{datasets.use}{uses only the specified datasets for sampling. Default is NULL (all datasets)}

\item{genes.use}{samples from only the specified genes. Default is NULL (all genes)}

\item{rand.seed}{(default 1).}

\item{rand.seed}{for reproducibility (default 1).
#Note:}
}
\value{
\code{liger} object with sample.data slot set.
}
\description{
This function samples raw/normalized/scaled data from on-disk HDF5 files for plotting. 
This function assumes that the cell barcodes are unique across all datasets.
}
\examples{
\dontrun{
Y <- matrix(c(1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12), nrow = 4, byrow = T)
Z <- matrix(c(1, 2, 3, 4, 5, 6, 7, 6, 5, 4, 3, 2), nrow = 4, byrow = T)
ligerex <- createLiger(list(y_set = Y, z_set = Z))
ligerex <- normalize(ligerex)
# select genes
ligerex <- selectGenes(ligerex)
ligerex <- scaleNotCenter(ligerex)
ligerex <- readSubset(ligerex, slot.use = "norm.data", max.cells = 5000)
}
}
