% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc.R
\name{removeMissing}
\alias{removeMissing}
\title{Remove missing cells or features from liger object}
\usage{
removeMissing(
  object,
  orient = c("both", "feature", "cell"),
  useDatasets = NULL,
  filenameSuffix = "removeMissing",
  verbose = getOption("ligerVerbose"),
  ...
)
}
\arguments{
\item{object}{\linkS4class{liger} object}

\item{orient}{Choose to remove non-expressing features (\code{"feature"}),
empty barcodes (\code{"cell"}), or both of them (\code{"both"}). Default
\code{"both"}.}

\item{useDatasets}{A character vector of the names, a numeric or logical
vector of the index of the datasets to be processed. Default
\code{NULL} removes empty entries from all datasets.}

\item{filenameSuffix}{When subsetting H5 based datasets to new H5 files, this
suffix will be added to all the filenames. Default \code{"removeMissing"}.}

\item{verbose}{Logical. Whether to show information of the progress. Default
\code{getOption("ligerVerbose")} which is \code{TRUE} if users have not set.}

\item{...}{Arguments passed to \code{\link{subsetLigerDataset}}}
}
\value{
Updated (subset) \code{object}.
}
\description{
Remove missing cells or features from liger object
}
\examples{
data("pbmc", package = "rliger")
# The example dataset does not contain non-expressing genes or empty barcodes
pbmc <- removeMissing(pbmc)
}
