% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analogizer.R
\name{removeMissingCells}
\alias{removeMissingCells}
\title{Remove cells with no gene data}
\usage{
removeMissingCells(object)
}
\arguments{
\item{object}{Analogizer object (scale.data must be set).}
}
\value{
Analogizer object with modified scale.data (dataset names preserved).
}
\description{
Removes cells from scale.data with no expression in any selected genes.
}
\examples{
\dontrun{
# analogizer object
analogy
analogy <- removeMissingCells(analogy)
}
}
