% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h5Utility.R
\name{restoreH5Liger}
\alias{restoreH5Liger}
\title{Restore links (to HDF5 files) for reloaded liger/ligerDataset object}
\usage{
restoreH5Liger(object, filePath = NULL)
}
\arguments{
\item{object}{liger or ligerDataset object.}

\item{file.path}{Paths to HDF5 files. A single character path for
ligerDataset input or a list of paths named by the datasets for liger object
input. Default \code{NULL} looks for the path(s) of the last valid loading.}
}
\value{
\code{object} with restored links.
}
\description{
When loading the saved liger object with HDF5 data in a new R
session, the links to HDF5 files would be corrupted. This functions enables
the restoration of those links so that new analyses can be carried out.
}
