% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h5Utility.R
\name{restoreOnlineLiger}
\alias{restoreOnlineLiger}
\title{Restore links (to HDF5 files) for reloaded online liger object}
\usage{
restoreOnlineLiger(object, filePath = NULL)
}
\arguments{
\item{object}{liger or ligerDataset object.}

\item{file.path}{List of paths to hdf5 files.}
}
\value{
\code{object} with restored links.
}
\description{
When loading the saved online liger object in a new R session,
the links to HDF5 files may be corrupted. This functions enables the
restoration of those links so that new analyses can be carried out.
}
