% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analogizer.R
\name{riverplot_clusters}
\alias{riverplot_clusters}
\title{Makes a riverplot to show how separate cluster assignments from two datasets map onto a joint clustering.}
\usage{
riverplot_clusters(object, cluster1, cluster2)
}
\arguments{
\item{object}{analogizer object. Should run quantile_align_SNF before calling.}

\item{cluster1}{Cluster assignments for dataset 1. Note that cluster names should be distinct across datasets.}

\item{cluster1}{Cluster assignments for dataset 2. Note that cluster names should be distinct across datasets.}
}
\description{
Makes a riverplot to show how separate cluster assignments from two datasets map onto a joint clustering.
}
\examples{
\dontrun{
Y = matrix(c(1,2,3,4,5,6,7,8,9,10,11,12),nrow=4,byrow=T)
Z = matrix(c(1,2,3,4,5,6,7,6,5,4,3,2),nrow=4,byrow=T)
analogy = Analogizer(list(Y,Z))
analogy@var.genes = c(1,2,3,4)
analogy = scaleNotCenter(analogy)
}
}
