% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ATAC.R, R/clustering.R, R/deprecated.R,
%   R/metrics.R, R/quantileNorm.R
\name{makeInteractTrack}
\alias{makeInteractTrack}
\alias{louvainCluster}
\alias{rliger-deprecated}
\alias{calcAgreement}
\alias{calcAlignmentPerCluster}
\alias{calcARI}
\alias{calcPurity}
\alias{quantile_norm}
\alias{quantile_norm,ANY-method}
\title{Deprecated functions in package \pkg{rliger}.}
\usage{
makeInteractTrack(
  corr.mat,
  path_to_coords,
  genes.list = NULL,
  output_path = getwd()
)

louvainCluster(
  object,
  resolution = 1,
  k = 20,
  prune = 1/15,
  eps = 0.1,
  nRandomStarts = 10,
  nIterations = 100,
  random.seed = 1,
  verbose = getOption("ligerVerbose"),
  dims.use = NULL
)

calcAgreement(
  object,
  dr.method = "NMF",
  ndims = 40,
  k = 15,
  use.aligned = TRUE,
  rand.seed = 42,
  by.dataset = FALSE
)

calcAlignmentPerCluster(object, rand.seed = 1, k = NULL, by.dataset = FALSE)

calcARI(object, clusters.compare, verbose = TRUE)

calcPurity(object, classes.compare, verbose = TRUE)

quantile_norm(
  object,
  quantiles = 50,
  ref_dataset = NULL,
  min_cells = 20,
  knn_k = 20,
  dims.use = NULL,
  do.center = FALSE,
  max_sample = 1000,
  eps = 0.9,
  refine.knn = TRUE,
  clusterName = "H.norm_cluster",
  rand.seed = 1,
  verbose = getOption("ligerVerbose")
)

\S4method{quantile_norm}{ANY}(
  object,
  quantiles = 50,
  ref_dataset = NULL,
  min_cells = 20,
  knn_k = 20,
  dims.use = NULL,
  do.center = FALSE,
  max_sample = 1000,
  eps = 0.9,
  refine.knn = TRUE,
  rand.seed = 1,
  verbose = getOption("ligerVerbose")
)
}
\description{
The functions listed below are deprecated and will be defunct in
  the near future. When possible, alternative functions with similar
  functionality or a replacement are also mentioned. Help pages for
  deprecated functions are available at \code{help("<function>-deprecated")}.
}
\section{\code{makeInteractTrack}}{

For \code{makeInteractTrack}, use \code{\link{exportInteractTrack}}.
}

\section{\code{louvainCluster}}{

For \code{louvainCluster}, use \code{\link{runLouvainCluster}} as the
replacement, while \code{\link{runLeidenCluster}} is more recommended.
}

\section{\code{calcAgreement}}{

\code{calcAgreement} is deprecated and is now defunct.
}

\section{\code{calcAlignmentPerCluster}}{

\code{calcAlignmentPerCluster} is deprecated and is now defunct.
}

\section{\code{calcARI}}{

\code{calcARI} is deprecated and is now defunct.
}

\section{\code{calcPurity}}{

\code{calcPurity} is deprecated and is now defunct.
}

\section{\code{quantile_norm}}{

For \code{quantile_norm}, use \code{\link{quantileNorm}}.


For \code{quantile_norm}, use \code{\link{quantileNorm}}.
}

\keyword{internal}
