% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rliger-package.R
\docType{package}
\name{rliger-package}
\alias{rliger}
\alias{rliger-package}
\title{rliger: Linked Inference of Genomic Experimental Relationships}
\description{
Uses an extension of nonnegative matrix factorization to identify shared and dataset-specific factors. See Welch J, Kozareva V, et al (2019) \doi{10.1016/j.cell.2019.05.006}, and Liu J, Gao C, Sodicoff J, et al (2020) \doi{10.1038/s41596-020-0391-8} for more details.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://welch-lab.github.io/liger/}
  \item Report bugs at \url{https://github.com/welch-lab/liger/issues}
}

}
\author{
\strong{Maintainer}: Yichen Wang \email{wayichen@umich.edu}

Authors:
\itemize{
  \item Joshua Welch \email{welchjd@umich.edu}
  \item Chao Gao \email{gchao@umich.edu}
  \item Jialin Liu \email{alanliu@umich.edu}
  \item Joshua Sodicoff \email{sodicoff@umich.edu} [contributor]
  \item Velina Kozareva [contributor]
  \item Evan Macosko [contributor]
}

Other contributors:
\itemize{
  \item Paul Hoffman [contributor]
  \item Ilya Korsunsky [contributor]
  \item Robert Lee [contributor]
  \item Andrew Robbins \email{robbiand@med.umich.edu} [contributor]
}

}
\keyword{internal}
