% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DoubletFinder.R
\name{runDoubletFinder}
\alias{runDoubletFinder}
\title{Doublet detection with DoubletFinder}
\usage{
runDoubletFinder(
  object,
  useDatasets = NULL,
  nPCs = 10,
  nNeighbors = 20,
  nExp = NULL,
  verbose = getOption("ligerVerbose"),
  ...
)
}
\arguments{
\item{object}{A \linkS4class{liger} object.}

\item{useDatasets}{A character vector of the names, a numeric or logical
vector of the index of the datasets to run
\code{\link[DoubletFinder]{doubletFinder_v3}} with. Default \code{NULL}
applies to all datasets.}

\item{nPCs}{Number of top PC to use. Default \code{10}.}

\item{nNeighbors}{Number of the PC neighborhood size used to compute pANN.
See "See Also". Scalar for all used datasets or vector for each. Default
\code{20}.}

\item{nExp}{The total number of doublet predictions produced. Scalar for all
used datasets or vector for each. Default \code{NULL} sets a 0.15 proportion.}

\item{verbose}{Logical. Whether to show information of the progress. Default
\code{getOption("ligerVerbose")} which is \code{TRUE} if users have not set.}

\item{...}{Additional arguments passed to
\code{\link[DoubletFinder]{doubletFinder_v3}}.}
}
\value{
Updated \code{object} with variables \code{DoubletFinder_pANN} and
\code{DoubletFinder_classification} updated in \code{cellMeta} slot
}
\description{
Detect doublet with DoubletFinder. Package "Seurat" and
"DoubletFinder" would be required to run this function. "DoubletFinder" can
only be installed via
\href{https://github.com/chris-mcginnis-ucsf/DoubletFinder}{GitHub}.

Seurat default PCA workflow is performed prior to calling DoubletFinder.
}
\examples{
pbmc <- runDoubletFinder(pbmc)
cellMeta(pbmc)
}
\seealso{
\code{\link[DoubletFinder]{doubletFinder_v3}}
}
