% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSEA.R
\name{runFactorGeneGO}
\alias{runFactorGeneGO}
\title{Run Gene Ontology enrichment analysis on metagenes}
\usage{
runFactorGeneGO(
  object,
  useW = TRUE,
  useDatasets = NULL,
  sumLoading = TRUE,
  ...
)
}
\arguments{
\item{object}{A \linkS4class{liger} object with valid factorization result.}

\item{useW}{Logical, whether to consider shared gene loading value (i.e.
\eqn{W} matrix) when forming the genesets. Default \code{TRUE}.}

\item{useDatasets}{A character vector of the names, a numeric or logical
vector of the index of the datasets where the gene loading need to be
considered. Default \code{NULL} considers all datasets.}

\item{sumLoading}{Logical, whether to sum up the gene loading from all
considered datasets for each factor, or to query a geneset per factor and
dataset. Default \code{TRUE}.}

\item{...}{Additional arguments passed to \code{\link[gprofiler2]{gost}}.}
}
\value{
A list object with the following entries
\item{result}{data.frame of main GO analysis result.}
\item{meta}{Meta information for the query.}
}
\description{
This function forms genesets basing on the non-zero gene loading in each
factor, and calls gene ontology (GO) analysis method provided by gprofiler2.
}
\examples{
go <- runFactorGeneGO(pbmcPlot)
head(go$result)
}
\references{
Kolberg, L. et al, 2020 and Raudvere, U. et al, 2019
}
