% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSEA.R
\name{runGSEA}
\alias{runGSEA}
\title{Analyze biological interpretations of metagene}
\usage{
runGSEA(
  object,
  genesets = NULL,
  useW = TRUE,
  useDatasets = NULL,
  customGenesets = NULL,
  gene_sets = genesets,
  mat_w = useW,
  mat_v = useDatasets,
  custom_gene_sets = customGenesets
)
}
\arguments{
\item{object}{\code{liger} object.}

\item{gene_sets}{A list of the Reactome gene sets names to be tested. If not specified,
this function will use all the gene sets from the Reactome by default}

\item{mat_w}{This indicates whether to use the shared factor loadings 'W' (default TRUE)}

\item{mat_v}{This indicates which V matrix to be added to the analysis. It can be a numeric number or a list
of the numerics.}

\item{custom_gene_sets}{A named list of character vectors of entrez gene ids. If not specified,
this function will use all the gene symbols from the input matrix by default}
}
\value{
A list of matrices with GSEA analysis for each factor
}
\description{
Identify the biological pathways (gene sets from Reactome) that each metagene (factor) might belongs to.
}
