% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liger.R
\name{runJaccard}
\alias{runJaccard}
\title{Make Jaccard matrix}
\usage{
runJaccard(object)

runJaccard(object)
}
\arguments{
\item{object}{\code{liger} object. Should run scaleNotCenter first}

\item{object}{\code{liger} object. Should run scaleNotCenter first}
}
\value{
Updated \code{liger} object.

Updated \code{liger} object.
}
\description{
Make Jaccard matrix

Make Jaccard matrix (will have use when imputation is implemented)
}
\examples{
\dontrun{
# liger object, after running scaleNotCenter 
ligerex
# add Jaccard matrix to a list for each dataset
ligerex <- runJaccard(ligerex)
}
\dontrun{
# liger object, after running scaleNotCenter 
ligerex
# add Jaccard matrix to a list for each dataset
ligerex <- runJaccard(ligerex)
}
}
