% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudoBulkDEG.R
\name{runPseudoBulkDEG}
\alias{runPseudoBulkDEG}
\title{Run DESeq2 differential expression test on pseudo-bulk expression}
\usage{
runPseudoBulkDEG(
  object,
  comparison,
  useCellMetaVar = NULL,
  replicateAnn = NULL,
  nRep = 5,
  seed = 1,
  verbose = getOption("ligerVerbose")
)
}
\arguments{
\item{object}{A \linkS4class{liger} object.}

\item{comparison}{A named list with two elements, where each is a vector of
valid cell specification (numeric or logical indexing should be based on the
whole object).}

\item{useCellMetaVar}{Name(s) of \code{cellMeta} variable that specify the
replicate annotation. Default \code{NULL}.}

\item{replicateAnn}{A data.frame or factor to pass external replicate
annotation. See Details for requirements. Default \code{NULL}.}

\item{nRep}{Number of pseudo replicates per comparison group, applied only
when running pseudo replicate mode (i.e. both \code{useCellMetaVar} and
\code{replicateAnn} are \code{NULL}). Default \code{5}.}

\item{seed}{Random seed to allow reproducible results. Default \code{1}.}

\item{verbose}{Logical. Whether to show information of the progress. Default
\code{getOption("ligerVerbose")} which is \code{TRUE} if users have not set.}
}
\value{
A data.frame as the output of DESeq2. Ordered by adjusted p-value
(\code{padj}).
}
\description{
This function aggregate raw counts data of specified comparison groups of
cells basing on either given biological replicate annotation or pseudo
replicates. According to user prior knowledge, existing replicate annotation
should be used when it is true that the cells involved in comparison come
from multiple biological replicates, while when all cells in both groups
come from the same replicate, pseudo replicates should be created.

Known replicate annotation can be specified by \code{useCellMetaVar} to use
metadata stored in the \code{cellMeta} slot, or specify external annotation
through \code{replicateAnn}. To use pseudo replicates, users should leave the
two mentioned argument as default \code{NULL}.
}
\details{
For using \code{replicateAnn}. When passing a data.frame, rownames have to be
available and it should cover all cells involved in both comparison group.
When using a named factor/vector, the names should as well cover all involved
cells. When using a un-named factor/vector, it is recommended to have its
length matched with the total number of cells of the whole object. When the
length of the un-named factor/vector equals to \code{length(comparison[[1]])
+ length(comparison[[2]])}, it is assumed that the order matches to
\code{c(comparison[[1]], comparison[[2]])}.
}
\examples{
rawData(datasets(pbmcPlot)[[1]]) <- rawData(dataset(pbmc, 1))
rawData(datasets(pbmcPlot)[[2]]) <- rawData(dataset(pbmc, 2))
res <- runPseudoBulkDEG(
    pbmcPlot,
    comparison = list(
        c1 = pbmcPlot$leiden_cluster == 1,
        c2 = pbmcPlot$leiden_cluster == 2
    )
)
print(head(res))
}
