% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liger.R
\name{runTSNE}
\alias{runTSNE}
\title{Perform t-SNE dimensionality reduction}
\usage{
runTSNE(object, use.raw = F, dims.use = 1:ncol(object@H.norm),
  use.pca = F, perplexity = 30, rand.seed = 42)
}
\arguments{
\item{object}{\code{liger} object. Should run quantileAlignSNF before calling with defaults.}

\item{use.raw}{Whether to use un-aligned cell factor loadings (H matrices) (default FALSE).}

\item{dims.use}{Factors to use for computing tSNE embedding (default 1:ncol(H.norm)).}

\item{perplexity}{Parameter to pass to Rtsne (expected number of neighbors) (default 30).}

\item{rand.seed}{Random seed for reproducibility (default 42).}
}
\value{
\code{liger} object with tsne.coords slot set.
}
\description{
Runs t-SNE on the normalized cell factors (or raw cell factors) to generate a 2D embedding for 
visualization. Has option to run on subset of factors. Note that running multiple times will
reset tsne.coords values.
}
\examples{
\dontrun{
 # liger object
ligerex
# generate H.norm by quantile aligning factor loadings
ligerex <- quantileAlignSNF(ligerex)
# get tsne.coords for normalized data
ligerex <- runTSNE(ligerex)
# get tsne.coords for raw factor loadings
ligerex <- runTSNE(ligerex, use.raw = T)
}
}
