% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runUINMF.R
\name{runUINMF}
\alias{runUINMF}
\alias{runUINMF.liger}
\alias{runUINMF.list}
\alias{runUINMF.Seurat}
\title{Perform iNMF on scaled datasets, and include unshared features}
\usage{
runUINMF(
  object,
  k = 20,
  lambda = 5,
  nIteration = 30,
  nRandomStarts = 1,
  seed = 1,
  verbose = getOption("ligerVerbose"),
  ...
)

\method{runUINMF}{liger}(
  object,
  k = 20,
  lambda = 5,
  nIteration = 30,
  nRandomStarts = 1,
  seed = 1,
  verbose = getOption("ligerVerbose"),
  ...
)

\method{runUINMF}{list}(
  object,
  unsharedList,
  k = 20,
  lambda = 5,
  nIteration = 30,
  nRandomStarts = 1,
  seed = 1,
  verbose = getOption("ligerVerbose"),
  ...
)

\method{runUINMF}{Seurat}(
  object,
  unsharedList,
  datasetVar,
  k = 20,
  lambda = 5,
  nIteration = 30,
  nRandomStarts = 1,
  seed = 1,
  verbose = getOption("ligerVerbose"),
  ...
)
}
\arguments{
\item{object}{\linkS4class{liger} object. Should run
\code{\link{selectGenes}} with \code{unshared = TRUE} and then run
\code{\link{scaleNotCenter}} in advance.}

\item{k}{Integer, inner dimension of factorization (number of factors).
Default \code{20}.}

\item{lambda}{Numeric, the regularization parameter. Default \code{5}.}

\item{nIteration}{Total number of iterations to perform. Default \code{30}.}

\item{nRandomStarts}{Number of restarts to perform (iNMF objective function
is non-convex, so taking the best objective from multiple successive
initializations is recommended). For easier reproducibility, this increments
the random seed by 1 for each consecutive restart, so future factorizations
of the same dataset can be run with one rep if necessary. Default \code{1}.}

\item{seed}{Random seed to allow reproducible results. Default \code{1}.}

\item{verbose}{Logical. Whether to show information of the progress. Default
\code{getOption("ligerVerbose")} which is \code{TRUE} if users have not set.}

\item{unsharedList}{List of matrices for unshared features}

\item{datasetVar}{Variable for dataset belonging.}
}
\description{
Perform iNMF on scaled datasets, and include unshared features
}
