% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embedding.R
\name{runUMAP}
\alias{runUMAP}
\title{Perform UMAP Dimensionality Reduction}
\usage{
runUMAP(
  object,
  use.raw = FALSE,
  dims.use = NULL,
  k = 2,
  distance = c("euclidean", "cosine", "manhattan", "hamming"),
  n_neighbors = 10,
  min_dist = 0.1,
  rand.seed = 42
)
}
\arguments{
\item{object}{\linkS4class{liger} object with factorization results.}

\item{use.raw}{Whether to use un-aligned cell factor loadings (\eqn{H}
matrices). Default \code{FALSE}.}

\item{dims.use}{Index of factors to use for computing UMAP embedding. Default
\code{NULL} uses all factors.}

\item{k}{Number of dimensions to reduce to. Default \code{2}.}

\item{distance}{Character. Metric used to measure distance in the input
space. Default \code{"euclidean"}, alternative options include:
\code{"cosine"}, \code{"manhattan"} and \code{"hamming"}.}

\item{n_neighbors}{Number of neighboring points used in local approximations
of manifold structure. Default \code{10}.}

\item{min_dist}{Numeric. Controls how tightly the embedding is allowed
compress points together. Default \code{0.1}.}

\item{rand.seed}{Random seed for reproducibility. Default \code{42}.}
}
\value{
The \code{object} with new variables, from \code{UMAP_1} to
\code{UMAP_[k]}, updated in the \code{cell.meta} slot.
}
\description{
Run UMAP on the quantile normalized cell factors (result from
\code{\link{quantile_norm}}), or unnormalized cell
factors (result from \code{\link{optimizeALS}} or \code{\link{online_iNMF}}))
to generate a 2D embedding for visualization (or general dimensionality
reduction). Has option to run on subset of factors. It is generally
recommended to use this method for dimensionality reduction with extremely
large datasets. The underlying UMAP calculation imports uwot
\code{\link[uwot]{umap}}.
}
\details{
For \code{n_neighbors}, larger values will result in more global
structure being preserved at the loss of detailed local structure. In general
this parameter should often be in the range 5 to 50, with a choice of 10 to
15 being a sensible default.

For \code{min_dist}, larger values ensure embedded points are more evenly
distributed, while smaller values allow the algorithm to optimise more
accurately with regard to local structure. Sensible values are in the range
0.001 to 0.5, with 0.1 being a reasonable default.
}
\seealso{
\code{\link{runTSNE}}
}
