% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ligerDataset-class.R
\name{scale.data<-,ligerDataset-method}
\alias{scale.data<-,ligerDataset-method}
\title{Set norm.data to ligerDataset object}
\usage{
\S4method{scale.data}{ligerDataset}(x, check = TRUE) <- value
}
\arguments{
\item{x}{ligerDataset object}

\item{value}{dgCMatrix}
}
\description{
Set norm.data to ligerDataset object
}
