% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaleNotCenter.R
\name{scaleNotCenter}
\alias{scaleNotCenter}
\title{Scale genes by root-mean-square across cells}
\usage{
scaleNotCenter(object, chunk = 1000, verbose = TRUE)
}
\arguments{
\item{object}{\linkS4class{liger} object, \code{link{selectGenes}} should
have been run in advance.}

\item{chunk}{Integer. Number of maximum number of cells in each chunk, when
scaling is applied to any HDF5 based dataset. Default \code{1000}.}

\item{verbose}{Logical. Whether to show information of the progress. Default
\code{TRUE}.}
}
\value{
Updated \code{object}, where the \code{scale.data} slot of each
\linkS4class{ligerDataset} object in the \code{datasets} slot is updated.
}
\description{
This function scales normalized gene expression data after
variable genes have been selected. Note that the data is not mean-centered
before scaling because expression values must remain positive (NMF only
accepts positive values).
}
