% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectGenes.R
\name{selectGenes.H5}
\alias{selectGenes.H5}
\title{Select variable genes from an H5 ligerDataset object}
\usage{
selectGenes.H5(
  object,
  var.thresh = 0.1,
  alpha.thresh = 0.99,
  capitalize = FALSE,
  cell.meta = NULL
)
}
\arguments{
\item{object}{ligerDataset Object}
}
\description{
Select variable genes from an H5 ligerDataset object
}
