% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analogizer.R
\name{selectGenes}
\alias{selectGenes}
\title{Select a subset of informative genes from each dataset and combine them.}
\usage{
selectGenes(object, alphathresh = 0.99, varthresh = 0.1, cex.use = 0.3,
  combine = "union", keep.unique = F, capitalize = F)
}
\arguments{
\item{object}{analogizer object. Should have already called normalize.}

\item{alphathresh}{alpha threshold}

\item{varthresh}{variance threshold}

\item{cex.use}{point size for plot}

\item{combine}{How to combine variable genes across experiments. Either "union" or "intersect"}

\item{keep.unique}{Keep genes that occur (i.e., there is a corresponding column in raw.data) only in one dataset?}

\item{capitalize}{Capitalize gene names to match homologous genes?}
}
\value{
analogizer object
}
\description{
Select a subset of informative genes from each dataset and combine them.
}
\examples{
\dontrun{
Y = matrix(c(1,2,3,4,5,6,7,8,9,10,11,12),nrow=4,byrow=T)
Z = matrix(c(1,2,3,4,5,6,7,6,5,4,3,2),nrow=4,byrow=T)
analogy = Analogizer(list(Y,Z))
analogy = normalize(analogy)
analogy = selectGenes(analogy)
}
}
