% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{selectGenesVST}
\alias{selectGenesVST}
\title{Select variable gene from one dataset with VST method}
\usage{
selectGenesVST(
  object,
  useDataset,
  n = 2000,
  loessSpan = 0.3,
  clipMax = "auto",
  verbose = getOption("ligerVerbose")
)
}
\arguments{
\item{object}{A \linkS4class{liger} object}

\item{useDataset}{The names, a numeric or logical index of the ONE dataset to
be considered for selection.}

\item{n}{Number of variable features needed. Default \code{2000}.}

\item{loessSpan}{Loess span parameter used when fitting the variance-mean
relationship. Default \code{0.3}.}

\item{clipMax}{After standardization values larger than \code{clipMax} will
be set to \code{clipMax}. Default \code{"auto"} sets this value to the square
root of the number of cells.}

\item{verbose}{Logical. Whether to show information of the progress. Default
\code{getOption("ligerVerbose")} which is \code{TRUE} if users have not set.}
}
\description{
A re-implimentation of Seurat FindVariableFeatures VST method. This allows
the selection of a fixed number of variable features from one dataset.
}
\examples{
pbmc <- selectGenesVST(pbmc, "ctrl", n = 50)
}
\references{
Seurat::FindVariableFeatures.default(selection.method = "vst")
}
