% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liger.R
\name{seuratToLiger}
\alias{seuratToLiger}
\title{Create liger object from one or more Seurat objects}
\usage{
seuratToLiger(objects, combined.seurat = F, names = "use-projects",
  meta.var = NULL, renormalize = T, use.seurat.genes = T,
  num.hvg.info = NULL, use.idents = T, use.tsne = T, cca.to.H = F)
}
\arguments{
\item{objects}{One or more Seurat objects. If passing multiple objects, should be in list.}

\item{combined.seurat}{Whether Seurat object (single) already contains multiple datasets (default
FALSE).}

\item{names}{Names to use for datasets in new liger object. If use-projects, takes project names 
from individual Seurat objects; if use-meta.var, takes value of object meta.data in meta.var
column for each dataset (becomes default if passing combined Seurat object). Otherwise, user can
pass in vector of names with same length as number of datasets (default use-projects).}

\item{meta.var}{Seurat meta.data column name to use in naming datasets. Required if 
combined.seurat is TRUE (default NULL).}

\item{renormalize}{Whether to automatically normalize raw.data once \code{liger} object is created 
(default TRUE).}

\item{use.seurat.genes}{Carry over variable genes from Seurat objects. If num.hvg.info is set, uses
that value to get top most highly variable genes from hvg.info slot in Seurat objects. Otherwise 
uses var.genes slot in Seurat objects. For multiple datasets, takes the union of the variable 
genes. (default TRUE)}

\item{num.hvg.info}{Number of highly variable genes to include from each object's hvg.info slot.
If set, recommended value is 2000 (default NULL).}

\item{use.idents}{Carry over cluster identities from Seurat objects. If multiple objects with 
overlapping cluster names, will preface cluster names by dataset names to distinguish. (default
TRUE).}

\item{use.tsne}{Carry over t-SNE coordinates from Seurat object (only meaningful for combined
analysis Seurat object). Useful for plotting directly afterwards. (default TRUE)}

\item{cca.to.H}{Carry over CCA (and aligned) loadings and insert them into H (and H.norm) slot in 
liger object (only meaningful for combined analysis Seurat object). Useful for plotting directly 
afterwards. (default FALSE)}
}
\value{
\code{liger} object.
}
\description{
This function creates a \code{liger} object from multiple (disjoint) Seurat objects or a single (combined-
analysis) Seurat object. It includes options for keeping the variable genes and cluster identities
from the original Seurat objects. It renormalizes the raw.data by default.
}
\examples{
\dontrun{
# Seurat objects for two pbmc datasets
tenx <- readRDS('tenx.RDS')
seqwell <- readRDS('seqwell.RDS')
# create liger object, using project names
ligerex <- seuratToLiger(list(tenx, seqwell))
# create liger object, passing in names explicitly, using hvg.info genes
ligerex2 <- seuratToLiger(list(tenx, seqwell), names = c('tenx', 'seqwell'), num.hvg.info = 2000)
# Seurat object for joint analysis
pbmc <- readRDS('pbmc.RDS')
# create liger object, using 'protocol' for dataset names
ligerex3 <- seuratToLiger(pbmc, meta.var = 'protocol', num.hvg.info = 2000)
}
}
