% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analogizer.R
\docType{methods}
\name{show}
\alias{show}
\alias{show,analogizer-method}
\title{show method for analogizer}
\usage{
\S4method{show}{analogizer}(object)
}
\arguments{
\item{object}{analogizer object}
}
\description{
show method for analogizer
}
