% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liger-class.R
\name{[,liger,ANY,ANY,ANY-method}
\alias{[,liger,ANY,ANY,ANY-method}
\title{Subset liger object by cell}
\usage{
\S4method{[}{liger,ANY,ANY,ANY}(x, i, j, ..., drop = TRUE)
}
\arguments{
\item{x}{liger object}

\item{i}{subscriber for cell, based on cell.meta(x)}
}
\description{
Subset liger object by cell
}
