% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liger-methods.R
\name{[[<-,liger,ANY,missing-method}
\alias{[[<-,liger,ANY,missing-method}
\alias{[[<-,liger-method}
\title{Set cell metadata variable}
\usage{
\S4method{[[}{liger,ANY,missing}(x, i, j, ...) <- value
}
\arguments{
\item{x}{A \linkS4class{liger} object}

\item{i}{Name or numeric index of cell meta to be replaced}

\item{j}{Ignored}

\item{...}{Other argument that might be applied to [[<-,S4Vectors::DataFrame}

\item{value}{Metadata value to be inserted}
}
\value{
Input liger object updated with replaced/new variable in
\code{cellMeta(x)}.
}
\description{
Set cell metadata variable
}
\examples{
cellMeta(pbmc)
# Add new variable
pbmc[["newVar"]] <- 1
cellMeta(pbmc)
# Change existing variable
pbmc[["newVar"]][1:3] <- 1:3
cellMeta(pbmc)
}
