% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analogizer.R
\name{subsetAnalogizer}
\alias{subsetAnalogizer}
\title{Construct an Analogizer object with a specified subset of cells or clusters.
Should only be called after tsne coordinates have been computed and alignment
performed.}
\usage{
subsetAnalogizer(object, clusters.use = NULL, cells.use = NULL)
}
\arguments{
\item{object}{analogizer object.}

\item{clusters.use}{Clusters to extract}

\item{cells.use}{Vector of cell names to keep from any dataset}
}
\value{
analogizer object
}
\description{
Construct an Analogizer object with a specified subset of cells or clusters.
Should only be called after tsne coordinates have been computed and alignment
performed.
}
\examples{
\dontrun{
Y = matrix(c(1,2,3,4,5,6,7,8,9,10,11,12),nrow=4,byrow=T)
Z = matrix(c(1,2,3,4,5,6,7,6,5,4,3,2),nrow=4,byrow=T)
analogy = Analogizer(list(Y,Z))
analogy =
}
}
