% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liger.R
\name{subsetWilcoxon}
\alias{subsetWilcoxon}
\title{Subset a dataframe produced by calculateWilcox by feature, dataset, or column}
\usage{
subsetWilcoxon(object, features = NULL, datasets = NULL,
  columns = NULL, return.subset = FALSE, save.subset = TRUE)
}
\arguments{
\item{object}{\code{liger} object. Should run calculateWilcox first}

\item{features}{A list of features to include in the subset (default NULL)}

\item{datasets}{A list of datasets to include in the subset  (defualt NULL).}

\item{columns}{A list of columns to include in the subset  (default NULL).}

\item{return.subset}{Return the subsetted dataframe instead of a liger object (default FALSE)}

\item{save.subset}{Replace the unsubset dataframe with the subset (default TRUE)}
}
\value{
Updated \code{liger} object.
}
\description{
Subset a dataframe produced by calculateWilcox by feature, dataset, or column
}
\examples{
\dontrun{
# liger object, after running CalculateWilcoxon 
ligerex
# create lists of features, datasets, and columns to subset by
genes = c("UBBP4","KIAA1715","TMEM50B")
datasets = c("tenx")
#
ligerex <- subsetWilcoxon(ligerex, features = genes, datasets = datasets)
}
}
