% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suggestK.R
\name{suggestK}
\alias{suggestK}
\title{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}} Perform iterative cross-validation to
suggest the best k for the given data}
\usage{
suggestK(
  object,
  kTest = c(5, 10, 15, 20, 25, 30, 35, 40, 45, 50),
  nFold = 5,
  lambda = 5,
  nIteration = 30,
  verbose = getOption("ligerVerbose", TRUE)
)
}
\description{
For a series of k values, we perform n-fold cross validation on the given
data. We use the train data to get factorized low-rank matrices (the \eqn{W}
and \eqn{V} matrices for gene loading in metagenes) and predict the \eqn{H}
matrices (for metagene loading in cells) for the test data. The objective
error (see \code{\link{runINMF}}) is calculated for the test data and shown
as the mean squared error (MSE) for each k. The k with the lowest MSE is
suggested as the best k for the given data.
}
