% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorGSEA.R
\name{summary.factorGSEA}
\alias{summary.factorGSEA}
\title{Show significant results from factorGSEA}
\usage{
\method{summary}{factorGSEA}(object, ...)
}
\arguments{
\item{object}{A \code{factorGSEA} object.}

\item{...}{S3 method convention, not used for now.}
}
\value{
A data frame of significant tests with gene set names, factor names
and other GSEA statistics.
}
\description{
Show significant results from factorGSEA
}
