% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liger-class.R
\name{varFeatures,liger-method}
\alias{varFeatures,liger-method}
\alias{varFeatures<-,liger,ANY,character-method}
\title{Access variable features of liger object}
\usage{
\S4method{varFeatures}{liger}(x)

\S4method{varFeatures}{liger,ANY,character}(x, check = TRUE) <- value
}
\arguments{
\item{x}{\linkS4class{liger} object}
}
\description{
Access variable features of liger object
}
