% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analogizer.R
\name{AnalogizerToSeurat}
\alias{AnalogizerToSeurat}
\title{Create a Seurat object containing the data from an Analogizer object.}
\usage{
AnalogizerToSeurat(object, need.sparse = F)
}
\arguments{
\item{object}{analogizer object.}
}
\description{
Create a Seurat object containing the data from an Analogizer object.
}
\examples{
\dontrun{
Y = matrix(c(1,2,3,4,5,6,7,8,9,10,11,12),nrow=4,byrow=T)
Z = matrix(c(1,2,3,4,5,6,7,6,5,4,3,2),nrow=4,byrow=T)
analogy = Analogizer(list(Y,Z))
analogy@var.genes = c(1,2,3,4)
analogy = scaleNotCenter(analogy)
}
}
