% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analogizer.R
\name{AnalogizerToSeurat}
\alias{AnalogizerToSeurat}
\title{Create a Seurat object containing the data from an Analogizer object.}
\usage{
AnalogizerToSeurat(object, need.sparse = T)
}
\arguments{
\item{object}{Analogizer object.}

\item{need.sparse}{Whether data needs to be converted to sparse format first; most relevant for 
older Analogizer objects (default TRUE).}
}
\description{
Create a Seurat object containing the data from an Analogizer object.
}
\examples{
\dontrun{
# Analogizer object
analogy
s.object <- AnalogizerToSeurat(analogy)
}
}
