% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classConversion.R
\name{as.liger}
\alias{as.liger}
\alias{as.liger.dgCMatrix}
\alias{as.liger.SingleCellExperiment}
\alias{as.liger.Seurat}
\alias{seuratToLiger}
\title{Converting other classes of data to a liger object}
\usage{
as.liger(object, ...)

\method{as.liger}{dgCMatrix}(object, datasetVar = NULL, modal = NULL, ...)

\method{as.liger}{SingleCellExperiment}(object, datasetVar = NULL, modal = NULL, ...)

\method{as.liger}{Seurat}(object, datasetVar = "orig.ident", modal = NULL, assay = NULL, ...)

seuratToLiger(
  object,
  datasetVar = "orig.ident",
  modal = NULL,
  assay = NULL,
  ...
)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments passed to \code{\link{createLiger}}}

\item{datasetVar}{Specify the dataset belonging by: 1. Select a variable from
existing metadata in the object (e.g. colData column); 2. Specify a
vector/factor that assign the dataset belonging. 3. Give a single character
string which means that all data is from one dataset (must not be a metadata
variable, otherwise it is understood as scenario 1.). Default \code{NULL}
gathers things into one dataset and names it "sample" for dgCMatrix, "SCE"
for SingleCellExperiment, or "Seurat" for Seurat.}

\item{modal}{Modality setting for each dataset. See
\code{\link{createLiger}}.}

\item{assay}{Name of assay to use. Default \code{NULL} uses current active
assay.}
}
\value{
a \linkS4class{liger} object.
}
\description{
Converting other classes of data to a liger object
}
\examples{
# dgCMatrix (common sparse matrix class), usually obtained from other
# container object, and contains multiple samples merged in one.
matList <- rawData(pbmc)
multiSampleMatrix <- mergeSparseAll(matList)
# The `datasetVar` argument expects the variable assigning the sample source
pbmc2 <- as.liger(multiSampleMatrix, datasetVar = pbmc$dataset)
pbmc2

sce <- SingleCellExperiment::SingleCellExperiment(
    assays = list(counts = multiSampleMatrix)
)
sce$sample <- pbmc$dataset
pbmc3 <- as.liger(sce, datasetVar = "sample")
pbmc3

seu <- SeuratObject::CreateSeuratObject(multiSampleMatrix)
# Seurat creates variable "orig.ident" by identifying the cell barcode
# prefixes, which is indeed what we need in this case. Users might need
# to be careful and have it confirmed first.
pbmc4 <- as.liger(seu, datasetVar = "orig.ident")
pbmc4
}
