% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classConversion.R
\name{as.ligerDataset}
\alias{as.ligerDataset}
\alias{as.ligerDataset.ligerDataset}
\alias{as.ligerDataset.matrixLike}
\alias{as.ligerDataset.Seurat}
\alias{as.ligerDataset.anndata._core.anndata.AnnData}
\title{Converting other classes of data to a as.ligerDataset object}
\usage{
as.ligerDataset(object, ...)

\method{as.ligerDataset}{ligerDataset}(object, modal = c("default", "rna", "atac"), ...)

\method{as.ligerDataset}{matrixLike}(object, modal = c("default", "rna", "atac"), ...)

\method{as.ligerDataset}{Seurat}(object, modal = c("default", "rna", "atac"), ...)

\method{as.ligerDataset}{anndata._core.anndata.AnnData}(object, modal = c("default", "rna", "atac"), ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments passed to \code{\link{createLigerDataset}}}

\item{modal}{Modality setting for each dataset.}
}
\value{
a \linkS4class{liger} object.
}
\description{
Converting other classes of data to a as.ligerDataset object
}
\examples{
data("pbmc", package = "rliger")
ctrl <- dataset(pbmc, "ctrl")
ctrl
as.ligerDataset(ctrl, modal = "atac")
rawCounts <- rawData(ctrl)
class(rawCounts)
as.ligerDataset(rawCounts)
}
