% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classConversion.R
\name{as.ligerDataset}
\alias{as.ligerDataset}
\alias{as.ligerDataset,SingleCellExperiment-method}
\alias{as.ligerDataset,ligerDataset-method}
\alias{as.ligerDataset,matrixLike-method}
\alias{as.ligerDataset,Seurat-method}
\alias{as.ligerDataset,anndata._core.anndata.AnnData-method}
\title{Convert an object of various classes to a ligerDataset object}
\usage{
as.ligerDataset(x, modal = c("default", "rna", "atac"))

\S4method{as.ligerDataset}{SingleCellExperiment}(x, modal = c("default", "rna", "atac"))

\S4method{as.ligerDataset}{ligerDataset}(x, modal = c("default", "rna", "atac"))

\S4method{as.ligerDataset}{matrixLike}(x, modal = c("default", "rna", "atac"))

\S4method{as.ligerDataset}{Seurat}(x, modal = c("default", "rna", "atac"))

\S4method{as.ligerDataset}{anndata._core.anndata.AnnData}(x, modal = c("default", "rna", "atac"))
}
\arguments{
\item{x}{An object to be converted}

\item{modal}{The modality of this dataset. Default \code{"default"} for RNA.
Can choose from \code{"rna"}, \code{"atac"}.}
}
\value{
A ligerDataset object by default, or a modality specific sub-class of
ligerDataset object according to \code{modal}.
}
\description{
S4 method to convert objects of various types to a
\linkS4class{ligerDataset} object or a modality specific sub-class of
\linkS4class{ligerDataset} object. Supported classes include a matrix like
object, a \code{SingleCellExperiment} object, a \code{Seurat} object, and
\code{AnnData} object. This method also supports modality setting to a
\linkS4class{ligerDataset} inherited object.
}
\author{
Yichen Wang
}
