% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analogizer.R
\name{calcPurity}
\alias{calcPurity}
\title{Calculate purity for Analogizer clustering and external clustering (base truth).
Should run quantile_align_SNF first.}
\usage{
calcPurity(object, classes.compare)
}
\arguments{
\item{object}{analogizer object.}

\item{classes.compare}{Clustering with which to compare (named vector)}
}
\value{
Purity value
}
\description{
Calculate purity for Analogizer clustering and external clustering (base truth).
Should run quantile_align_SNF first.
}
\examples{
\dontrun{
Y = matrix(c(1,2,3,4,5,6,7,8,9,10,11,12),nrow=4,byrow=T)
Z = matrix(c(1,2,3,4,5,6,7,6,5,4,3,2),nrow=4,byrow=T)
analogy = Analogizer(list(Y,Z))
analogy =
}
}
