% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liger-class.R
\name{cell.meta,liger,NULL-method}
\alias{cell.meta,liger,NULL-method}
\alias{cell.meta,liger,character-method}
\alias{cell.meta,liger,missing-method}
\alias{cell.meta<-,liger,missing-method}
\alias{cell.meta<-,liger,character-method}
\title{Access cell.meta of liger object}
\usage{
\S4method{cell.meta}{liger,`NULL`}(x, columns = NULL, cellIdx = NULL, as.data.frame = FALSE, ...)

\S4method{cell.meta}{liger,character}(x, columns = NULL, cellIdx = NULL, as.data.frame = FALSE, ...)

\S4method{cell.meta}{liger,missing}(x, columns = NULL, cellIdx = NULL, as.data.frame = FALSE, ...)

\S4method{cell.meta}{liger,missing}(x, columns = NULL, check = FALSE) <- value

\S4method{cell.meta}{liger,character}(x, columns = NULL, check = FALSE) <- value
}
\arguments{
\item{x}{\linkS4class{liger} object}
}
\description{
Access cell.meta of liger object

Access cell.meta of liger object

Access cell.meta of liger object
}
