% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classConversion.R
\name{convertOldLiger}
\alias{convertOldLiger}
\title{Convert old liger object to latest version}
\usage{
convertOldLiger(object)
}
\arguments{
\item{object}{\code{liger} object from rliger version <1.2.0}
}
\description{
Convert old liger object to latest version
}
