% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liger.R
\name{convertOldLiger}
\alias{convertOldLiger}
\title{Convert older liger object into most current version (based on class definition)}
\usage{
convertOldLiger(object, override.raw = F)
}
\arguments{
\item{object}{\code{liger} object.}

\item{override.raw}{Keep original raw.data without any modifications (removing missing cells
etc.) (defualt FALSE).}
}
\value{
Updated \code{liger} object.
}
\description{
Also works for Analogizer objects (but must have both liger and Analogizer loaded). Transfers
data in slots with same names from old class object to new, leaving slots defined only in new
class NULL.
}
\examples{
\dontrun{
# old Analogizer object
analogy
# convert to latest class definition
ligerex <- convertOldLiger(analogy)
}
}
