% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rliger.R
\name{convertOldLiger}
\alias{convertOldLiger}
\title{Convert older liger object into most current version (based on class definition)}
\usage{
convertOldLiger(object, override.raw = FALSE, verbose = TRUE)
}
\arguments{
\item{object}{\code{liger} object.}

\item{override.raw}{Keep original raw.data without any modifications (removing missing cells
etc.) (defualt FALSE).}

\item{verbose}{Print progress bar/messages (TRUE by default)}
}
\value{
Updated \code{liger} object.
}
\description{
Also works for Analogizer objects (but must have both liger and Analogizer loaded). Transfers
data in slots with same names from old class object to new, leaving slots defined only in new
class NULL.
}
\examples{
\dontrun{
# Not able to generate old object from current version, thus not run
ligerex <- convertOldLiger(analogy)
}
}
