% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liger-class.R
\name{createLiger}
\alias{createLiger}
\title{Create liger object}
\usage{
createLiger(
  raw.data,
  modal = NULL,
  cell.meta = NULL,
  take.gene.union = FALSE,
  remove.missing = TRUE,
  format.type = "10X",
  data.name = NULL,
  indices.name = NULL,
  indptr.name = NULL,
  genes.name = NULL,
  barcodes.name = NULL,
  verbose = TRUE
)
}
\arguments{
\item{raw.data}{Named list of datasets. Required. Elements allowed include a
matrix, a \linkS4class{Seurat} object, a \linkS4class{SingleCellExperiment}
object, an \code{AnnData} object, a \linkS4class{ligerDataset} object or a
filename to an HDF5 file. See detail for HDF5 reading.}

\item{modal}{Character vector for modality setting. See detail.}

\item{cell.meta}{data.frame of metadata at single-cell level. Default}

\item{remove.missing}{Whether to remove cells that do not have any counts and
features not expressed in any cells from each dataset. Default \code{TRUE}.}

\item{format.type}{Select preset of H5 file structure. Current available
options are \code{"10X"} and \code{"AnnData"}. Can be either a single
specification for all datasets or a character vector that match with each
dataset.}
}
\description{
This function allows creating \linkS4class{liger} object from
multiple datasets of various forms (See \code{raw.data}).
}
\seealso{
\code{\link{createLigerDataset}},
\code{\link{createLigerDataset.h5}}
}
