% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liger.R
\name{createLiger}
\alias{createLiger}
\title{Create a liger object.}
\usage{
createLiger(raw.data, sparse.dcg = T)
}
\arguments{
\item{raw.data}{List of expression matrices (gene by cell). Should be named by dataset.}

\item{sparse.dcg}{Whether to convert raw data into sparse matrices (default: T).}
}
\value{
\code{liger} object with raw.data slot set.
}
\description{
This function initializes a liger object with the raw data passed in. It requires a list of 
expression (or another single-cell modality) matrices (gene by cell) for at least two datasets. 
By default, it converts all passed data into sparse matrices (dgCMatrix) to reduce object size.
}
\examples{
\dontrun{
Y <- matrix(c(1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12), nrow = 4, byrow = T)
Z <- matrix(c(1, 2, 3, 4, 5, 6, 7, 6, 5, 4, 3, 2), nrow = 4, byrow = T)
ligerex <- createLiger(list(y_set = Y, z_set = Z))
}
}
