% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liger-class.R
\name{createLiger}
\alias{createLiger}
\title{Create liger object}
\usage{
createLiger(
  datasets = NULL,
  type = NULL,
  cell.meta = NULL,
  var.features = NULL,
  H = NULL,
  H.norm = NULL,
  uns = NULL
)
}
\arguments{
\item{datasets}{Named list of datasets. List elements can be one of in memory
expression matrix, file name (.h5), constructed ligerDataset object.}

\item{cell.meta}{data.frame}

\item{var.features}{vector of feature names}

\item{H}{matrix}

\item{H.norm}{matrix}
}
\description{
Create liger object
}
