% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liger-class.R
\name{createLiger}
\alias{createLiger}
\title{Create liger object}
\usage{
createLiger(
  raw.data = NULL,
  format.type = NULL,
  modal = NULL,
  cell.meta = NULL,
  take.gene.union = FALSE,
  remove.missing = FALSE,
  h5.format = "10X",
  data.name = NULL,
  indices.name = NULL,
  indptr.name = NULL,
  genes.name = NULL,
  barcodes.name = NULL,
  verbose = TRUE
)
}
\arguments{
\item{cell.meta}{data.frame}

\item{datasets}{Named list of datasets. List elements can be one of in memory
expression matrix, file name (.h5), constructed ligerDataset object.}

\item{var.features}{vector of feature names}

\item{H}{matrix}

\item{H.norm}{matrix}
}
\description{
Create liger object
}
