% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liger.R
\name{createLiger}
\alias{createLiger}
\title{Create a liger object.}
\usage{
createLiger(raw.data, make.sparse = T, take.gene.union = F,
  remove.missing = T)
}
\arguments{
\item{raw.data}{List of expression matrices (gene by cell). Should be named by dataset.}

\item{make.sparse}{Whether to convert raw data into sparse matrices (default TRUE).}

\item{take.gene.union}{Whether to fill out raw.data matrices with union of genes across all 
datasets (filling in 0 for missing data) (requires make.sparse=T) (default FALSE).}

\item{remove.missing}{Whether to remove cells not expressing any measured genes, and genes not
expressed in any cells (if take.gene.union = T, removes only genes not expressed in any 
dataset) (default TRUE).}
}
\value{
\code{liger} object with raw.data slot set.
}
\description{
This function initializes a liger object with the raw data passed in. It requires a list of 
expression (or another single-cell modality) matrices (gene by cell) for at least two datasets. 
By default, it converts all passed data into sparse matrices (dgCMatrix) to reduce object size.
}
\examples{
\dontrun{
Y <- matrix(c(1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12), nrow = 4, byrow = T)
Z <- matrix(c(1, 2, 3, 4, 5, 6, 7, 6, 5, 4, 3, 2), nrow = 4, byrow = T)
ligerex <- createLiger(list(y_set = Y, z_set = Z))
}
}
