% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ligerDataset-class.R
\name{createLigerDataset}
\alias{createLigerDataset}
\title{Create ligerDataset object}
\usage{
createLigerDataset(
  raw.data = NULL,
  modal = c("default", "rna", "atac"),
  norm.data = NULL,
  scale.data = NULL,
  scale.unshared.data = NULL,
  var.unshared.features = NULL,
  H = NULL,
  V = NULL,
  A = NULL,
  B = NULL,
  U = NULL,
  h5file.info = NULL,
  feature.meta = NULL,
  ...
)
}
\arguments{
\item{raw.data}{matrix}

\item{norm.data}{matrix}

\item{V}{matrix}

\item{A}{matrix}

\item{B}{matrix}

\item{U}{matrix}

\item{annotation}{data.frame}
}
\description{
Create ligerDataset object
}
