% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ligerDataset-class.R
\name{raw.data}
\alias{raw.data}
\alias{raw.data<-}
\alias{norm.data}
\alias{norm.data<-}
\alias{scale.data}
\alias{scale.data<-}
\alias{scale.unshared.data}
\alias{scale.unshared.data<-}
\alias{getH5File}
\alias{h5file.info}
\alias{h5file.info<-}
\alias{feature.meta}
\alias{feature.meta<-}
\alias{getMatrix}
\alias{raw.data,ligerDataset-method}
\alias{raw.data<-,ligerDataset,ANY,matrixLike_OR_NULL-method}
\alias{raw.data<-,ligerDataset,ANY,H5D-method}
\alias{norm.data,ligerDataset-method}
\alias{norm.data<-,ligerDataset,ANY,matrixLike_OR_NULL-method}
\alias{norm.data<-,ligerDataset,ANY,H5D-method}
\alias{scale.data,ligerDataset,ANY-method}
\alias{scale.data,liger,character-method}
\alias{scale.data,liger,numeric-method}
\alias{scale.data<-,ligerDataset,ANY,matrixLike_OR_NULL-method}
\alias{scale.data<-,ligerDataset,ANY,H5D-method}
\alias{scale.unshared.data,ligerDataset,ANY-method}
\alias{scale.unshared.data,liger,character-method}
\alias{scale.unshared.data,liger,numeric-method}
\alias{scale.unshared.data<-,ligerDataset,ANY,matrixLike_OR_NULL-method}
\alias{scale.unshared.data<-,ligerDataset,ANY,H5D-method}
\alias{getH5File,ligerDataset,missing-method}
\alias{getH5File,liger,character-method}
\alias{h5file.info,ligerDataset-method}
\alias{h5file.info<-,ligerDataset-method}
\alias{feature.meta,ligerDataset-method}
\alias{feature.meta<-,ligerDataset-method}
\alias{getMatrix,ligerDataset,ANY,missing,missing-method}
\alias{getMatrix,liger,ANY,ANY,ANY-method}
\title{Access data in a liger or ligerDataset object}
\usage{
raw.data(x)

raw.data(x, check = TRUE) <- value

norm.data(x)

norm.data(x, check = TRUE) <- value

scale.data(x, name = NULL)

scale.data(x, check = TRUE) <- value

scale.unshared.data(x, name = NULL)

scale.unshared.data(x, check = TRUE) <- value

getH5File(x, dataset = NULL)

h5file.info(x, info = NULL)

h5file.info(x, info = NULL, check = TRUE) <- value

feature.meta(x, check = NULL)

feature.meta(x, check = TRUE) <- value

getMatrix(x, slot = "raw.data", dataset = NULL, returnList = FALSE)

\S4method{raw.data}{ligerDataset}(x)

\S4method{raw.data}{ligerDataset,ANY,matrixLike_OR_NULL}(x, check = TRUE) <- value

\S4method{raw.data}{ligerDataset,ANY,H5D}(x, check = TRUE) <- value

\S4method{norm.data}{ligerDataset}(x)

\S4method{norm.data}{ligerDataset,ANY,matrixLike_OR_NULL}(x, check = TRUE) <- value

\S4method{norm.data}{ligerDataset,ANY,H5D}(x, check = TRUE) <- value

\S4method{scale.data}{ligerDataset,ANY}(x, name = NULL)

\S4method{scale.data}{liger,character}(x, name = NULL)

\S4method{scale.data}{liger,numeric}(x, name = NULL)

\S4method{scale.data}{ligerDataset,ANY,matrixLike_OR_NULL}(x, check = TRUE) <- value

\S4method{scale.data}{ligerDataset,ANY,H5D}(x, check = TRUE) <- value

\S4method{scale.unshared.data}{ligerDataset,ANY}(x, name = NULL)

\S4method{scale.unshared.data}{liger,character}(x, name = NULL)

\S4method{scale.unshared.data}{liger,numeric}(x, name = NULL)

\S4method{scale.unshared.data}{ligerDataset,ANY,matrixLike_OR_NULL}(x, check = TRUE) <- value

\S4method{scale.unshared.data}{ligerDataset,ANY,H5D}(x, check = TRUE) <- value

\S4method{getH5File}{ligerDataset,missing}(x, dataset = NULL)

\S4method{getH5File}{liger,character}(x, dataset = NULL)

\S4method{h5file.info}{ligerDataset}(x, info = NULL)

\S4method{h5file.info}{ligerDataset}(x, info = NULL, check = TRUE) <- value

\S4method{feature.meta}{ligerDataset}(x, check = NULL)

\S4method{feature.meta}{ligerDataset}(x, check = TRUE) <- value

\S4method{getMatrix}{ligerDataset,ANY,missing,missing}(
  x,
  slot = c("raw.data", "norm.data", "scale.data", "scale.unshared.data", "H", "V", "U"),
  dataset = NULL
)

\S4method{getMatrix}{liger,ANY,ANY,ANY}(
  x,
  slot = c("raw.data", "norm.data", "scale.data", "scale.unshared.data", "H", "V", "U",
    "W", "H.norm"),
  dataset = NULL,
  returnList = FALSE
)
}
\arguments{
\item{x}{\linkS4class{ligerDataset} object or \linkS4class{liger} object}

\item{check}{Whether to perform object validity check on setting new value.}

\item{value}{matrix like or \code{NULL}}

\item{name, dataset}{Name or numeric index of a dataset when \code{x} is a
\linkS4class{liger} object, in order to only get the data from this specified
dataset. Default \code{NULL}.}

\item{info}{Name of the entry in \code{h5file.info} slot.}

\item{slot}{The slot name when using \code{getMatrix}.}

\item{returnList}{When using \code{getMatrix} and data from one dataset is
being returned, whether it should be contained in a list or return the data
directly.}
}
\description{
Access data in a liger or ligerDataset object
}
