% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplotting.R
\name{.ggplotLigerTheme}
\alias{.ggplotLigerTheme}
\title{Generic ggplot theme setting for rliger package}
\usage{
.ggplotLigerTheme(
  plot,
  title = NULL,
  subtitle = NULL,
  xlab = TRUE,
  ylab = TRUE,
  legendColorTitle = NULL,
  legendShapeTitle = NULL,
  legendSizeTitle = NULL,
  legend = TRUE,
  baseSize = 10,
  titleSize = NULL,
  subtitleSize = NULL,
  xTextSize = NULL,
  xTitleSize = NULL,
  yTextSize = NULL,
  yTitleSize = NULL,
  legendTextSize = NULL,
  legendTitleSize = NULL,
  legendDotSize = 4,
  legendNRow = NULL,
  legendNCol = NULL,
  colorPalette = "plasma",
  colorDirection = -1,
  colorLow = NULL,
  colorMid = NULL,
  colorHigh = NULL,
  colorMidPoint = NULL,
  plotly = FALSE
)
}
\arguments{
\item{plot}{ggplot object passed from wrapper plotting functions}

\item{title, subtitle, xlab, ylab}{Main title, subtitle or X/Y axis title text.
By default, no main title or subtitle will be set, and X/Y axis title will be
the names of variables used for plotting. Use \code{NULL} to hide elements.
\code{TRUE} for \code{xlab} or \code{ylab} shows default values.}

\item{legendColorTitle, legendShapeTitle}{Set alternative title text for
legend on color or shape. Default \code{NULL} shows the original variable
name.}

\item{legend}{Whether to show the legend. Default \code{TRUE}.}

\item{baseSize}{One-parameter control of all text sizes. Individual text
element sizes can be controlled by other size arguments. "Title" sizes are
2 points larger than "text" sizes when being controlled by this.}

\item{titleSize, xTitleSize, yTitleSize, legendTitleSize}{Size of main title,
axis titles and legend title. Default \code{NULL} controls by
\code{baseSize + 2}.}

\item{subtitleSize, xTextSize, yTextSize, legendTextSize}{Size of subtitle text,
axis texts and legend text. Default \code{NULL} controls by \code{baseSize}.}

\item{legendDotSize}{Allow dots in legend region to be large enough to see
the colors/shapes clearly. Default \code{4}.}

\item{legendNRow, legendNCol}{Integer, when too many categories in one
variable, arranges number of rows or columns. Default \code{NULL},
automatically split to \code{ceiling(levels(variable)/10)} columns.}

\item{colorPalette}{For continuous coloring, a palette name or an index to
select from available options from ggplot
\code{\link[ggplot2]{scale_brewer}}. Additionally, \code{"viridis"} TODO are
also supported. Default \code{"viridis"}.}

\item{colorLow, colorMid, colorHigh, colorMidPoint}{All four of these must be
specified to customize palette with
\code{\link[ggplot2]{scale_colour_gradient2}}. Default \code{NULL}.}

\item{plotly}{Whether to use plotly to enable web based interactive browsing
for the plot. Requires installation of package "plotly". Default
\code{FALSE}.}
}
\value{
Updated ggplot object by default. When \code{plotly = TRUE}, returns
plotly (htmlwidget) object.
}
\description{
Controls content and size of all peripheral texts.
}
