% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplotting.R
\name{element_line_fixlen}
\alias{element_line_fixlen}
\title{Line element with fixed length setting}
\usage{
element_line_fixlen(
  colour = NULL,
  length = NULL,
  linewidth = NULL,
  linetype = NULL,
  lineend = NULL,
  color = NULL,
  arrow = NULL,
  inherit.blank = FALSE
)
}
\arguments{
\item{colour, color}{Line/border colour. Color is an alias for colour.}

\item{length}{a \code{\link[grid:unit]{grid::unit()}} object to specify length of the line.}

\item{linewidth}{Line/border size in mm.}

\item{linetype}{Line type. An integer (0:8), a name (blank, solid,
dashed, dotted, dotdash, longdash, twodash), or a string with
an even number (up to eight) of hexadecimal digits which give the
lengths in consecutive positions in the string.}

\item{lineend}{Line end Line end style (round, butt, square)}

\item{arrow}{Arrow specification, as created by \code{\link[grid:arrow]{grid::arrow()}}}

\item{inherit.blank}{Should this element inherit the existence of an
\code{element_blank} among its parents? If \code{TRUE} the existence of
a blank element among its parents will cause this element to be blank as
well. If \code{FALSE} any blank parent element will be ignored when
calculating final element state.}
}
\value{
An S3 object of class \code{element_line_fixlen}, \code{element_line}, and \code{element}.
}
\description{
Used at the place in ggplot theme \verb{axis.line*}
}
\examples{
ggplot(mtcars, aes(mpg, wt)) +
    geom_point() +
    theme(
        axis.line = element_line_fixlen(length = unit(1, "cm")),
        axis.ticks = element_blank(),
        axis.text = element_blank(),
        axis.title = element_text(hjust = 0)
    )
}
