% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liger-class.R
\name{fortify.liger}
\alias{fortify.liger}
\title{Fortify method for classes from the liger package}
\usage{
\method{fortify}{liger}(x)
}
\arguments{
\item{x}{\code{liger} object to convert into a dataframe.}
}
\value{
A \code{data.frame} combining \code{cellMeta(x)} with \code{H.norm}
matrix.
}
\description{
Fortify method for classes from the liger package
}
