% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{getProportionMito}
\alias{getProportionMito}
\title{Calculate proportion mitochondrial contribution}
\usage{
getProportionMito(object, use.norm = FALSE, pattern = "^mt-")
}
\arguments{
\item{object}{\code{liger} object.}

\item{use.norm}{\bold{Deprecated} Whether to use cell normalized data in
calculating contribution. Default \code{FALSE}.}

\item{pattern}{Regex pattern for identifying mitochondrial genes. Default
\code{"^mt-"} for mouse.}
}
\value{
Named vector containing proportion of mitochondrial contribution for
each cell.
}
\description{
Calculates proportion of mitochondrial contribution based on raw or
normalized data.
}
\note{
\code{getProportionMito} will be deprecated because
\code{\link{runGeneralQC}} generally covers and expands its use case.
}
\examples{
# Example dataset does not contain MT genes, expected to see a message
pbmc$mito <- getProportionMito(pbmc)
}
