% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h5Utility.R
\name{H5Apply}
\alias{H5Apply}
\title{Apply function to chunks of H5 data}
\usage{
H5Apply(
  h5file,
  init,
  FUN,
  dataPath = "matrix/data",
  indicesPath = "matrix/indices",
  indptrPath = "matrix/indptr",
  cellIDPath = "matrix/barcodes",
  featureIDPath = "matrix/features/name",
  chunkSize = 1000,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{h5file}{H5File object}

\item{FUN}{A function where the first argument must be the chunk of data, the
return value must be a vector.}
}
\description{
h5 calculation wrapper, that run user specified calculation with on-disk
matrix in chunks
}
