% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ATAC.R
\name{imputeKNN}
\alias{imputeKNN}
\title{Impute the query cell expression matrix}
\usage{
imputeKNN(
  object,
  reference,
  queries = NULL,
  nNeighbors = 20,
  weight = TRUE,
  norm = TRUE,
  scale = FALSE,
  verbose = TRUE,
  ...,
  knn_k = nNeighbors
)
}
\arguments{
\item{object}{\linkS4class{liger} object with aligned factor loading computed
in advance.}

\item{reference}{Name of a dataset containing values to impute into query
dataset(s).}

\item{queries}{Names of datasets to be augmented by imputation. Should not
include \code{reference}. Default \code{NULL} uses all datasets except the
reference.}

\item{nNeighbors}{The maximum number of nearest neighbors to search. Default
\code{20}.}

\item{weight}{Logical. Whether to use KNN distances as weight matrix. Default
\code{FALSE}.}

\item{norm}{Logical. Whether to normalize the imputed data. Default
\code{TRUE}.}

\item{scale}{Logical. Whether to scale but not center the imputed data.
Default \code{TRUE}.}

\item{verbose}{Logical. Whether to show the progress and information. Default
\code{TRUE}.}

\item{...}{Optional arguments to be passed to \code{\link{normalize}} when
\code{norm = TRUE}.}

\item{knn_k}{\bold{Deprecated}. See Usage section for replacement.}
}
\value{
The input \code{object} where queried \linkS4class{ligerDataset}
objects in \code{datasets} slot are replaced. These datasets will all be
converted to \linkS4class{ligerATACDataset} class with an additional slot
\code{raw.peak} to store the imputed peak counts, and \code{norm.peak} for
normalized imputed peak counts if \code{norm = TRUE}.
}
\description{
Impute query features from a reference dataset using KNN.
}
