% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liger-class.R
\docType{class}
\name{liger-class}
\alias{liger-class}
\alias{liger}
\title{liger class}
\description{
Data container for LIGER analysis. The slot \code{datasets} is a
list where each element should be \linkS4class{ligerDataset} containing
dataset specific information, such as the expression matrices. The other
parts of liger object stores information that can be shared across the
analysis, such as the cell metadata and factorization result matrices.
}
\details{
Accessor methods are defined for getting and setting values from or
to the object, respectively.

A liger object with version number >= 1.2.0 is fortify-able which means users
can produce ggplots with the cell metadata simply by calling
\code{ggplot(liger_object)}.
}
\section{Slots}{

\describe{
\item{\code{datasets}}{list of \linkS4class{ligerDataset} objects.}

\item{\code{cell.meta}}{Cell metadata, \linkS4class{DFrame}}

\item{\code{var.features}}{vector of feature names}

\item{\code{H.norm}}{matrix}

\item{\code{H}}{matrix}

\item{\code{commands}}{Record of analysis}

\item{\code{uns}}{Unstructured meta-info of analyses}

\item{\code{version}}{Record of version of rliger package}
}}

