% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ligerDataset_subclasses.R
\docType{class}
\name{ligerATACDataset-class}
\alias{ligerATACDataset-class}
\alias{ligerATACDataset}
\title{Subclass of ligerDataset for ATAC modality}
\description{
Inherits from \linkS4class{ligerDataset} class. Contained slots
can be referred with the link.
}
\section{Slots}{

\describe{
\item{\code{raw.peak}}{sparse matrix}

\item{\code{norm.peak}}{sparse matrix}
}}

