% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ligerDataset-class.R
\docType{class}
\name{ligerDataset-class}
\alias{ligerDataset-class}
\alias{ligerDataset}
\title{ligerDataset class}
\description{
Object for storing dastaset specific information. Will be embedded within a
higher level \linkS4class{liger} object
}
\section{Slots}{

\describe{
\item{\code{rawData}}{Raw data.}

\item{\code{normData}}{Normalized data}

\item{\code{scaleData}}{Scaled data, usually with subset variable features}

\item{\code{scaleUnsharedData}}{Scaled data of features not shared with other
datasets}

\item{\code{varUnsharedFeatures}}{Variable features not shared with other datasets}

\item{\code{V}}{matrix}

\item{\code{A}}{matrix}

\item{\code{B}}{matrix}

\item{\code{H}}{matrix}

\item{\code{U}}{matrix}

\item{\code{h5fileInfo}}{list}

\item{\code{featureMeta}}{Feature metadata, DataFrame}

\item{\code{colnames}}{character}

\item{\code{rownames}}{character}
}}

