% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ligerDataset-class.R
\docType{class}
\name{ligerDataset-class}
\alias{ligerDataset-class}
\alias{ligerDataset}
\title{ligerDataset class}
\description{
Object for storing dastaset specific information. Will be embedded within a
higher level \linkS4class{liger} object
}
\section{Slots}{

\describe{
\item{\code{raw.data}}{Raw data.}

\item{\code{norm.data}}{Normalized data}

\item{\code{scale.data}}{Scaled data, usually with subset variable features}

\item{\code{scale.unshared.data}}{Scaled data of features not shared with other
datasets}

\item{\code{var.unshared.features}}{Variable features not shared with other datasets}

\item{\code{V}}{matrix}

\item{\code{A}}{matrix}

\item{\code{B}}{matrix}

\item{\code{U}}{matrix}

\item{\code{agg.data}}{list}

\item{\code{h5file.info}}{list}

\item{\code{feature.meta}}{Feature metadata, DataFrame}
}}

