% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{ligerMethDataset-class}
\alias{ligerMethDataset-class}
\alias{ligerMethDataset}
\title{Subclass of ligerDataset for Methylation modality}
\description{
Inherits from \linkS4class{ligerDataset} class. Contained slots
can be referred with the link. \code{\link{scaleNotCenter}} applied on
datasets of this class will automatically be taken by reversing the
normalized data instead of scaling the variable features.
}
