% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ligerToH5AD.R
\name{ligerToH5AD}
\alias{ligerToH5AD}
\title{Write liger object to H5AD files}
\usage{
ligerToH5AD(
  object,
  filename,
  useSlot = c("scaleData", "normData", "rawData"),
  saveRaw = useSlot != "rawData",
  overwrite = FALSE,
  verbose = getOption("ligerVerbose", TRUE)
)
}
\arguments{
\item{object}{A \linkS4class{liger} object}

\item{filename}{A character string, the path to the H5AD file to be written}

\item{useSlot}{Character scalar, which type of data is going to be stored
to \code{adata.X}. Default \code{"scaleData"}, choose from
\code{"scaleData"}, \code{"normData"}, or \code{"rawData"}.}

\item{saveRaw}{Logical, whether to save rawData to \code{adata.raw.X}.
Default \code{TRUE} when \code{useSlot} is not \code{"rawData"}, otherwise
\code{FALSE}.}

\item{overwrite}{Logical, whether to overwrite the file if it exists.}

\item{verbose}{Logical. Whether to show information of the progress. Default
\code{getOption("ligerVerbose")} which is \code{TRUE} if users have not set.}
}
\value{
No return value, an H5AD file is written to disk
}
\description{
Write liger object to H5AD files
}
\examples{
ligerToH5AD(pbmc, filename = tempfile(fileext = ".h5ad"))
}
