% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classConversion.R
\name{ligerToSeurat}
\alias{ligerToSeurat}
\title{#' @rdname as.ligerDataset
#' @export
#' @method as.ligerDataset anndata._core.anndata.AnnData
as.ligerDataset.anndata._core.anndata.AnnData <- function(
        object,
        modal = c("default", "rna", "atac"),
        ...
) {
    modal <- match.arg(modal)
    message("Python object AnnData input. ")
}
Convert between liger and Seurat object}
\usage{
ligerToSeurat(
  object,
  assay = NULL,
  nms = NULL,
  renormalize = NULL,
  use.liger.genes = NULL,
  by.dataset = FALSE
)
}
\arguments{
\item{object}{A \linkS4class{liger} object to be converted}

\item{assay}{Name of assay to store the data. Default \code{NULL} detects by
dataset modality. If the object contains various modality, default to
\code{"liger"}.}

\item{nms}{[Defunct] Will be ignored because new object structure does not
have related problem.}

\item{renormalize}{[Defunct] Will be ignored because since Seurat V5, layers
of data can exist at the same time and it is better to left it for users to
do it by theirselves.}

\item{use.liger.genes}{[Defunct] Will be ignored and will always set LIGER
variable features to the place.}

\item{by.dataset}{TODO. I don't know what this is.}
}
\description{
For converting a \linkS4class{liger} object to a Seurat object, the
\code{rawData}, \code{normData}, and \code{scaleData} from each dataset,
the \code{cellMeta}, \code{H.norm} and \code{varFeatures} slot will be
included. Compatible with V4 and V5. It is not recommended to use this
conversion if your \linkS4class{liger} object contains datasets from
various modalities.
}
