% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{normalize}
\alias{normalize}
\title{Normalize raw data in liger object}
\usage{
normalize(object, chunk = 1000, verbose = TRUE)
}
\description{
Perform normalization on each raw data matrix from each dataset
to account for total feature expression across a cell.
}
