% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{normalize}
\alias{normalize}
\alias{normalizePeak}
\title{Normalize rawData in liger object}
\usage{
normalize(
  object,
  log = FALSE,
  scaleFactor = NULL,
  useDatasets = NULL,
  chunk = 1000,
  verbose = getOption("ligerVerbose")
)

normalizePeak(
  object,
  log = FALSE,
  scaleFactor = NULL,
  useDatasets = NULL,
  verbose = getOption("ligerVerbose")
)
}
\arguments{
\item{object}{\linkS4class{liger} object}

\item{log}{Logical. Whether to do a \code{log(x + 1)} transform on the
normalized data. Default \code{TRUE}.}

\item{scaleFactor}{Numeric. Scale the normalized expression value by this
factor before transformation. \code{NULL} for not scaling. Default
\code{1e4}.}

\item{useDatasets}{A character vector of the names, a numeric or logical
vector of the index of the datasets to be normalized. Should specify ATACseq
datasets when using \code{normalizePeak}. Default \code{NULL} normalizes all
valid datasets.}

\item{chunk}{Integer. Number of maximum number of cells in each chunk, when
normalization is applied to any HDF5 based dataset. Default \code{1000}.}

\item{verbose}{Logical. Whether to show information of the progress. Default
\code{getOption("ligerVerbose")} which is \code{TRUE} if users have not set.}
}
\value{
Updated \code{object}. When using \code{normalize()}, the
\code{normData} slot of each \linkS4class{ligerDataset} object in the
\code{datasets} slot is updated with normalized values calculated from
\code{rawData} slot. While \code{normalizePeak()} normalizes rawPeak counts
in \code{rawPeak} slot of a \linkS4class{ligerATACDataset} object and update
result in its \code{normPeak} slot.
}
\description{
Perform normalization on each rawData matrix from each dataset
to account for total feature expression across a cell.
}
\examples{
pbmc <- normalize(pbmc)
}
