% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analogizer.R
\name{normalize}
\alias{normalize}
\title{Normalize raw datasets to column sums}
\usage{
normalize(object)
}
\arguments{
\item{object}{analogizer object.}
}
\value{
analogizer object
}
\description{
Normalize raw datasets to column sums
}
\examples{
\dontrun{
Y = matrix(c(1,2,3,4,5,6,7,8,9,10,11,12),nrow=4,byrow=T)
Z = matrix(c(1,2,3,4,5,6,7,6,5,4,3,2),nrow=4,byrow=T)
analogy = Analogizer(list(Y,Z))
analogy = normalize(analogy)
}
}
