% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rliger.R
\name{normalize}
\alias{normalize}
\title{Normalize raw datasets to column sums}
\usage{
normalize(object, chunk = 1000, format.type = "10X")
}
\arguments{
\item{object}{\code{liger} object.}

\item{chunk}{size of chunks in hdf5 file. (default 1000)}

\item{format.type}{string of HDF5 format (10X CellRanger by default).}
}
\value{
\code{liger} object with norm.data slot set.
}
\description{
This function normalizes data to account for total gene expression across a cell.
}
\examples{
\dontrun{
Y <- matrix(c(1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12), nrow = 4, byrow = T)
Z <- matrix(c(1, 2, 3, 4, 5, 6, 7, 6, 5, 4, 3, 2), nrow = 4, byrow = T)
ligerex <- createLiger(list(y_set = Y, z_set = Z))
ligerex <- normalize(ligerex)
}
}
