% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analogizer.R
\name{optimizeNewData}
\alias{optimizeNewData}
\title{Optimize objective function for new data. Uses an efficient strategy for updating that takes advantage of the information in the existing factorization.}
\usage{
optimizeNewData(object, k, lambda = 5, thresh = 1e-04, max_iters = 25)
}
\arguments{
\item{object}{analogizer object. Should normalize, select genes, and scale before calling.}

\item{lambda}{Regularization parameter. Larger values penalize dataset-specific effects more strongly.}

\item{thresh}{Convergence threshold. Convergence occurs when |obj0-obj|/(mean(obj0,obj)) < thresh}

\item{max_iters}{Maximum number of block coordinate descent iterations to perform}

\item{new.data}{list of raw data matrices (one or more). Each list entry should be named.}

\item{add.to.existing}{add the new data to existing datasets or treat as totally new datasets?}

\item{which.datasets}{list of datasets to append new.data to if add.to.existing is true. Otherwise, the most similar existing datasets for each entry in new.data.}
}
\value{
list of length N+2 containing Hs, W, Vs, and run statistics
}
\description{
Optimize objective function for new data. Uses an efficient strategy for updating that takes advantage of the information in the existing factorization.
}
\examples{
\dontrun{
Y = matrix(c(1,2,3,4,5,6,7,8,9,10,11,12),nrow=4,byrow=T)
Z = matrix(c(1,2,3,4,5,6,7,6,5,4,3,2),nrow=4,byrow=T)
analogy = Analogizer(Y,Z)
analogy@var.genes = c(1,2,3,4)
analogy = scaleNotCenter(analogy)
analogy = optimize_als(analogy,k=2,nrep=1)
}
}
