% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analogizer.R
\name{optimizeSubset}
\alias{optimizeSubset}
\title{Optimize objective function for a subset of the data. Uses an efficient strategy for updating that takes advantage of the information in the existing factorization.}
\usage{
optimizeSubset(object, cell.subset, lambda = 5, thresh = 1e-04,
  max_iters = 25)
}
\arguments{
\item{object}{analogizer object. Should normalize, select genes, and scale before calling.}

\item{cell.subset}{list of cell names to retain from each dataset.}

\item{lambda}{Regularization parameter. Larger values penalize dataset-specific effects more strongly.}

\item{thresh}{Convergence threshold. Convergence occurs when |obj0-obj|/(mean(obj0,obj)) < thresh}

\item{max_iters}{Maximum number of block coordinate descent iterations to perform}
}
\value{
list of length N+2 containing Hs, W, Vs, and run statistics
}
\description{
Optimize objective function for a subset of the data. Uses an efficient strategy for updating that takes advantage of the information in the existing factorization.
}
\examples{
\dontrun{
Y = matrix(c(1,2,3,4,5,6,7,8,9,10,11,12),nrow=4,byrow=T)
Z = matrix(c(1,2,3,4,5,6,7,6,5,4,3,2),nrow=4,byrow=T)
analogy = Analogizer(Y,Z)
analogy@var.genes = c(1,2,3,4)
analogy = scaleNotCenter(analogy)
analogy = optimize_als(analogy,k=2,nrep=1)
}
}
