% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pbmcPlot}
\alias{pbmcPlot}
\title{liger object of PBMC subsample data with plotting information available}
\format{
\linkS4class{liger} object with two datasets named by "ctrl" and
"stim".
}
\source{
https://www.nature.com/articles/nbt.4042
}
\usage{
pbmcPlot
}
\description{
This data was generated from data \code{"pbmc"} with default
parameter integration pipeline: normalize, selectGenes, scaleNotCenter,
runINMF, runCluster, runUMAP. To minimize the object size distributed with
the package, rawData and scaleData were removed. Genes are downsampled to
the top 50 variable genes, for smaller normData and \eqn{W} matrix.
}
\references{
Hyun Min Kang and et. al., Nature Biotechnology, 2018
}
\keyword{datasets}
