% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analogizer.R
\name{plotByDatasetAndCluster}
\alias{plotByDatasetAndCluster}
\title{Plot t-SNE coordinates of aligned datasets, colored by dataset and by cluster}
\usage{
plotByDatasetAndCluster(object)
}
\arguments{
\item{object}{analogizer object. Should run optimizeALS before calling.}

\item{quantiles}{Number of quantiles to use for quantile normalization}

\item{ref_dataset}{Name of dataset to use as a "reference" for normalization. By default,
the dataset with the largest number of cells is used.}
}
\value{
analogizer object
}
\description{
Plot t-SNE coordinates of aligned datasets, colored by dataset and by cluster
}
\examples{
\dontrun{
Y = matrix(c(1,2,3,4,5,6,7,8,9,10,11,12),nrow=4,byrow=T)
Z = matrix(c(1,2,3,4,5,6,7,6,5,4,3,2),nrow=4,byrow=T)
analogy = Analogizer(list(Y,Z))
analogy@var.genes = c(1,2,3,4)
analogy = scaleNotCenter(analogy)
}
}
