% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplotting.R
\name{plotCellViolin}
\alias{plotCellViolin}
\title{Generate violin/box plot(s) using liger object}
\usage{
plotCellViolin(
  object,
  y,
  groupBy = NULL,
  slot = c("cell.meta", "raw.data", "norm.data", "scale.data", "H.norm", "H"),
  yFunc = NULL,
  colorBy = NULL,
  splitBy = NULL,
  ...
)
}
\arguments{
\item{object}{\linkS4class{liger} object}

\item{y}{Available variable name in \code{slot} to look for the value to
visualize.}

\item{groupBy, colorBy}{Available variable name in \code{cell.meta} slot to
look for categorical grouping. See details. Default \code{NULL} produces no
grouping and all-black graphic elements.}

\item{slot}{Choose the slot to find the \code{y} variable. See Details.
Default \code{"cell.meta"}.}

\item{yFunc}{A function object that expects a vector/factor/data.frame
retrieved by \code{y} as the only input, and returns an object of the same
size, so that the y-axis is replaced by this output. Useful when, for
example, users need to scale the gene expression shown on plot.}

\item{splitBy}{Character vector of categorical variable names in
\code{cell.meta} slot. Split all cells by groupings on this/these variable(s)
to produce a violin plot containing only the cells in each group. Default
\code{NULL}.}

\item{...}{More plot setting arguments. See \code{\link{.ggCellViolin}} and
\code{\link{.ggplotLigerTheme}}.}
}
\value{
A ggplot object when a single plot is intended. A list of ggplot
objects, when multiple \code{y} variables and/or \code{splitBy} are set. When
\code{plotly = TRUE}, all ggplot objects become plotly (htmlwidget) objects.
}
\description{
This function allows for using available cell metadata, feature
expression or factor loading to generate violin plot, and grouping the data
with available categorical cell metadata. Available categorical cell metadata
can be used to form the color annotation. When it is different from the
grouping, it forms a nested grouping. Multiple y-axis variables are allowed
from the same specification of \code{slot}, and this returns a list of violin
plot for each. Users can further split the plot(s) by grouping on cells (e.g.
datasets).
}
\details{
Available option for \code{slot} include: \code{"cell.meta"},
\code{"raw.data"}, \code{"norm.data"}, \code{"scale.data"}, \code{"H.norm"}
and \code{"H"}. When \code{"raw.data"}, \code{"norm.data"} or
\code{"scale.data"}, \code{y} has to be a character vector of feature names.
When \code{"H.norm"} or \code{"H"}, \code{colorBy} can be any valid index to
select one factor of interests. Note that character index follows
\code{"factor_[k]"} format, with replacing \code{[k]} with an integer.

When \code{"cell.meta"}, \code{y} has to be an available column name in
the table. Note that, for \code{y} as well as \code{groupBy}, \code{colorBy}
and \code{splitBy} since a matrix object is feasible in \code{cell.meta}
table, using a column (e.g. named as \code{"column1"} in a certain matrix
(e.g. named as \code{"matrixVar"}) should follow the syntax of
\code{"matrixVar.column1"}. When the matrix does not have a "colname"
attribute, the subscription goes with \code{"matrixVar.V1"},
\code{"matrixVar.V2"} and etc. These are based on the nature of
\code{as.data.frame} method on a \code{\link[S4Vectors]{DataFrame}} object.
}
