% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liger.R
\name{plotGene}
\alias{plotGene}
\title{Plot t-SNE coordinates by expression of specified gene}
\usage{
plotGene(object, gene, use.raw = F, methylation.indices = NULL,
  pt.size = 0.1, min.clip = 0, max.clip = 1, points.only = F,
  option = "plasma", zero.color = "#F5F5F5", return.plots = F)
}
\arguments{
\item{object}{\code{liger} object. Should call runTSNE before calling.}

\item{gene}{Gene for which to plot relative expression.}

\item{use.raw}{Plot raw values instead of normalized and log-transformed data (default FALSE).}

\item{methylation.indices}{Indices of datasets in object with methylation data (this data is not
magnified and put on log scale).}

\item{pt.size}{Point size for plots (default 0.1)}

\item{min.clip}{Quantile probability for lower bound of methylation data (everything lower set 
to this quantile value) (default 0)}

\item{max.clip}{Quantile probability for upper bound of methylation data (everything greater set 
to this quantile value) (default 1)}

\item{points.only}{Remove axes when plotting t-sne coordinates (default FALSE).}

\item{option}{Colormap option to use for ggplot2's scale_color_viridis (default 'plasma').}

\item{zero.color}{Color to use for zero values (no expression) (default '#F5F5F5').}

\item{return.plots}{Return ggplot objects instead of printing directly (default FALSE).}
}
\description{
Generates plot of t-SNE coordinates colored by expression of specified gene, for each dataset.
Color scale can be modified.
}
