% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc.R
\name{plotQCMetric}
\alias{plotQCMetric}
\alias{plotTotalCountViolin}
\alias{plotGeneDetectedViolin}
\title{Generate violin plot of QC metrics of each dataset}
\usage{
plotQCMetric(
  object,
  metric,
  groupBy = "dataset",
  colorBy = NULL,
  dotPlot = FALSE,
  dotColor = "black",
  dotSize = 1,
  violinPlot = TRUE,
  violinAlpha = 0.8,
  boxPlot = FALSE,
  boxAlpha = 0,
  xlab = groupBy,
  ylab = metric,
  ...
)

plotTotalCountViolin(object, groupBy = "dataset", ...)

plotGeneDetectedViolin(object, groupBy = "dataset", ...)
}
\arguments{
\item{object}{\linkS4class{liger} object, or a data.frame like object of cell
metadata, where \code{"dataset"} variable must exist.}

\item{metric}{Metric to use. Should be found in \code{cell.meta(object)}.}

\item{groupBy}{Group data by this categorical variable.}

\item{colorBy}{Color the dot, violin and box plot with this variable.}

\item{dotPlot, violinPlot, boxPlot}{Whether to add corresponding plot(s).}

\item{dotColor}{Specify a color to set uniform color to all dots. Default
\code{"black"}. Set \code{NULL} then colored by \code{colorBy}.}

\item{dotSize}{Size of dots.}

\item{violinAlpha, boxAlpha}{The transparency of violins or boxes,
respectively. \code{1} is opaque and \code{0} is transparent.}

\item{...}{Theme setting parameters. Check \code{\link{.ggplotLigerTheme}}
for more options and details.}
}
\description{
By default, a violin plot grouped by dataset variable, overlaid by a box plot
would be generated. Options are available for adding jittered dot plot
behind.
}
