% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantileNorm.R
\name{quantile_norm,liger-method}
\alias{quantile_norm,liger-method}
\alias{quantile_norm,list-method}
\title{Quantile Align (Normalize) Factor Loadings}
\usage{
\S4method{quantile_norm}{liger}(
  object,
  quantiles = 50,
  ref_dataset = NULL,
  min_cells = 20,
  knn_k = 20,
  dims.use = NULL,
  do.center = FALSE,
  max_sample = 1000,
  eps = 0.9,
  refine.knn = TRUE,
  rand.seed = 1
)

\S4method{quantile_norm}{list}(
  object,
  quantiles = 50,
  ref_dataset = NULL,
  min_cells = 20,
  knn_k = 20,
  dims.use = NULL,
  do.center = FALSE,
  max_sample = 1000,
  eps = 0.9,
  refine.knn = TRUE,
  rand.seed = 1
)
}
\arguments{
\item{object}{A \linkS4class{liger} object with \code{\link{optimizeALS}} run
in advance. Or a list object of \eqn{H} matrices, all in \eqn{k * nCell}
dimensionality, the \code{H} entry of the output of \code{\link{optimizeALS}}
list method is ready for use (i.e. use \code{out$H}).}

\item{quantiles}{Number of quantiles to use for quantile normalization.
Default \code{50}.}

\item{ref_dataset}{Name or numeric index of the dataset to use as a
"reference" for normalization. Default \code{NULL} use the dataset with the
largest number of cells.}

\item{min_cells}{Minimum number of cells to consider a cluster shared across
datasets. Default \code{20}.}

\item{knn_k}{Number of nearest neighbors for within-dataset knn graph.
Default \code{20}.}

\item{dims.use}{Indices of factors to use for shared nearest factor
determination. Default \code{NULL} uses all factors.}

\item{do.center}{Whether to center the data when scaling factors (centering
could be useful for less sparse modalities like methylation data). Default
\code{FALSE}.}

\item{max_sample}{Maximum number of cells used for quantile normalization of
each cluster and factor. Default \code{1000}.}

\item{eps}{The error bound of the nearest neighbor search. Lower values give
more accurate nearest neighbor graphs but take much longer to compute.
Default \code{0.9}.}

\item{refine.knn}{whether to increase robustness of cluster assignments using
KNN graph. Default \code{TRUE}.}

\item{rand.seed}{Random seed to allow reproducible results. Default \code{1}.}
}
\value{
For method on \linkS4class{liger} object, the object will be returned
with the \code{H.norm} slot and \code{"H.norm_cluster"} variable in
\code{cell.meta} slot updated. For method on list, a result list with entries
of \code{H.norm} and \code{clusters} will be returned.
}
\description{
This process builds a shared factor neighborhood graph to
jointly cluster cells, then quantile normalizes corresponding clusters.

The first step, building the shared factor neighborhood graph, is performed
in SNF(), and produces a graph representation where edge weights between
cells (across all datasets) correspond to their similarity in the shared
factor neighborhood space. An important parameter here is \code{knn_k}, the
number of neighbors used to build the shared factor space.

Next we perform quantile alignment for each dataset, factor, and cluster (by
stretching/compressing datasets' quantiles to better match those of the
reference dataset). These aligned factor loadings are combined into a single
matrix and returned as "H.norm".
}
