% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analogizer.R
\name{quantile_norm}
\alias{quantile_norm}
\title{Quantile normalizes cell factors}
\usage{
quantile_norm(object, quantiles = 50, ref_dataset = NULL)
}
\arguments{
\item{object}{analogizer object. Should run optimizeALS before calling.}

\item{quantiles}{Number of quantiles to use for quantile normalization}

\item{ref_dataset}{Name of dataset to use as a "reference" for normalization. By default,
the dataset with the largest number of cells is used.}
}
\value{
analogizer object
}
\description{
Quantile normalizes cell factors
}
\examples{
\dontrun{
Y = matrix(c(1,2,3,4,5,6,7,8,9,10,11,12),nrow=4,byrow=T)
Z = matrix(c(1,2,3,4,5,6,7,6,5,4,3,2),nrow=4,byrow=T)
analogy = Analogizer(list(Y,Z))
analogy@var.genes = c(1,2,3,4)
analogy = scaleNotCenter(analogy)
}
}
