% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{rank_matrix}
\alias{rank_matrix}
\alias{rank_matrix.dgCMatrix}
\alias{rank_matrix.matrix}
\title{rank_matrix}
\usage{
rank_matrix(X)

\method{rank_matrix}{dgCMatrix}(X)

\method{rank_matrix}{matrix}(X)
}
\arguments{
\item{X}{feature by observation matrix.}
}
\value{
List with 2 items
}
\description{
Utility function to rank columns of matrix
}
