% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liger.R
\name{removeMissingCells}
\alias{removeMissingCells}
\title{Remove cells with no gene data}
\usage{
removeMissingCells(object, slot.use = "scale.data")
}
\arguments{
\item{object}{\code{liger} object (scale.data or norm.data must be set).}

\item{slot.use}{The data slot to filter (takes "raw.data" and "scale.data") (default "scale.data")}
}
\value{
\code{liger} object with modified scale.data (or norm.data) (dataset names preserved).
}
\description{
Removes cells from scale.data with no expression in any selected genes.
}
\examples{
\dontrun{
# liger object
ligerex
ligerex <- removeMissingCells(ligerex)
}
}
