% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc.R
\name{runGeneralQC}
\alias{runGeneralQC}
\title{General QC for liger object}
\usage{
runGeneralQC(object, features = NULL, pattern = NULL)
}
\arguments{
\item{object}{liger object with \code{raw.data} available in
each ligerDataset embedded}

\item{features}{Feature names matching the feature subsets that users want to
calculate the expression percentage with. A vector for a single subset, or a
named list for multiple subset. Default \code{NULL}.}

\item{pattern}{Regex patterns for matching the feature subsets that users
want to calculate the expression percentage with. A vector for a single
subset, or a named list for multiple subset. Default \code{NULL}.}
}
\value{
A liger object with \code{nUMI}, \code{nGene} updated
in \code{cell.meta(object)}, as well as expression percentage value for each
feature subset
}
\description{
Calculate number of UMIs, number of detected features and
percentage of feature subset (e.g. mito) expression per cell.
}
