% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSEA.R
\name{runGOEnrich}
\alias{runGOEnrich}
\title{Run Gene Ontology enrichment analysis on metagenes}
\usage{
runGOEnrich(
  result,
  group = NULL,
  useBg = TRUE,
  orderBy = "logFC",
  logFCThresh = 1,
  padjThresh = 0.01,
  ...
)
}
\arguments{
\item{result}{Data frame of unfiltered output from \code{\link{runWilcoxon}}.}

\item{group}{Selection of one group available from \code{result$group}.
Default \code{NULL} uses all groups involved in DE \code{result} table.}

\item{useBg}{Logical, whether to set all genes involved in DE analysis
(before threshold filtering) as a domain background of GO analysis. Default
\code{TRUE}.}

\item{orderBy}{Name of DE statistics metric to order the gene list for each
group. Choose from \code{"logFC"} (default), \code{"pval"} or \code{"padj"}.
Or set \code{NULL} to turn off ranked mode.}

\item{logFCThresh}{The log2FC threshold above which the genes will be used.
Default \code{1}.}

\item{padjThresh}{The adjusted p-value threshold less than which the genes
will be used. Default \code{0.01}.}

\item{...}{Additional arguments passed to \code{\link[gprofiler2]{gost}}.}
}
\value{
A list object with the following entries
\item{result}{data.frame of main GO analysis result.}
\item{meta}{Meta information for the query.}

See \code{\link[gprofiler2]{gost}}. for detailed explanation.
}
\description{
This function forms genesets basing on the differential expression result,
and calls gene ontology (GO) analysis method provided by gprofiler2.
}
\examples{
res <- runWilcoxon(pbmcPlot)
# Setting `significant = FALSE` because it's hard for a gene list obtained
# from small test dataset to represent real-life biology.
go <- runGOEnrich(res, group = 0, significant = FALSE)
}
\references{
Kolberg, L. et al, 2020 and Raudvere, U. et al, 2019
}
