% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embedding.R
\name{runTSNE}
\alias{runTSNE}
\title{Perform t-SNE dimensionality reduction}
\usage{
runTSNE(
  object,
  use.raw = FALSE,
  dims.use = NULL,
  use.pca = FALSE,
  perplexity = 30,
  theta = 0.5,
  method = c("Rtsne", "fftRtsne"),
  fitsne.path = NULL,
  rand.seed = 42
)
}
\arguments{
\item{object}{\linkS4class{liger} object with factorization results.}

\item{use.raw}{Whether to use un-aligned cell factor loadings (\eqn{H}
matrices). Default \code{FALSE}.}

\item{dims.use}{Index of factors to use for computing UMAP embedding. Default
\code{NULL} uses all factors.}

\item{use.pca}{Whether to perform initial PCA step for Rtsne. Default
\code{FALSE}.}

\item{perplexity}{Numeric parameter to pass to Rtsne (expected number of
neighbors). Default \code{30}.}

\item{theta}{Speed/accuracy trade-off (increase for less accuracy), set to
\code{0.0} for exact TSNE. Default \code{0.5}.}

\item{method}{Choose from \code{"Rtsne"} or \code{"fftRtsne"}. See
Description. Default \code{"Rtsne"}.}

\item{fitsne.path}{Path to the cloned FIt-SNE directory (i.e.
\code{'/path/to/dir/FIt-SNE'}). Required only when first time using
\code{runTSNE} with \code{method = "fftRtsne"}. Default \code{NULL}.}

\item{rand.seed}{Random seed for reproducibility. Default \code{42}.}
}
\value{
The \code{object} with new variables, from \code{TSNE_1} to
\code{TSNE_[k]}, updated in the \code{cell.meta} slot.
}
\description{
Runs t-SNE on the quantile normalized cell factors (result from
\code{\link{quantile_norm}}), or unnormalized cell factors (result from
\code{\link{optimizeALS}} or \code{\link{online_iNMF}})) to generate a 2D
embedding for visualization. By default \code{\link[Rtsne]{Rtsne}}
(Barnes-Hut implementation of t-SNE) method is invoked, while alternative
"fftRtsne" method (FFT-accelerated Interpolation-based t-SNE, using Kluger
Lab implementation) is also supported.

In order to run fftRtsne (recommended for large datasets), FIt-SNE must be
installed as instructed in detailed
\href{https://github.com/KlugerLab/FIt-SNE}{here}. Include the path to the
cloned FIt-SNE directory as the \code{fitsne.path} parameter, though this is
only necessary for the first call to run \code{runTSNE}. For more detailed
FIt-SNE installation instructions, see the liger repo
\href{https://github.com/welch-lab/liger#readme}{README}.
}
\seealso{
\code{\link{runUMAP}}
}
