% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analogizer.R
\name{runTSNE}
\alias{runTSNE}
\title{Perform t-SNE dimensionality reduction}
\usage{
runTSNE(object, use.raw = F, dims.use = 1:ncol(object@H.norm),
  perplexity = 30, rand.seed = 42)
}
\arguments{
\item{object}{Analogizer object. Should run quantileAlignSNF before calling with defaults.}

\item{use.raw}{Whether to use un-aligned cell factor loadings (H matrices) (default FALSE).}

\item{dims.use}{Factors to use for computing tSNE embedding (default 1:ncol(H.norm)).}

\item{perplexity}{Parameter to pass to Rtsne (expected number of neighbors) (default 30).}

\item{rand.seed}{Random seed for reproducibility (default 42).}
}
\value{
Analogizer object with tsne.coords slot set.
}
\description{
Runs t-SNE on the normalized cell factors (or raw cell factors) to generate a 2D embedding for 
visualization. Has option to run on subset of factors. Note that running multiple times will
reset tsne.coords values.
}
\examples{
\dontrun{
 # analogizer object
analogy
# generate H.norm by quantile aligning factor loadings
analogy <- quantileAlignSNF(analogy)
# get tsne.coords for normalized data
analogy <- runTSNE(analogy)
# get tsne.coords for raw factor loadings
analogy <- runTSNE(analogy, use.raw = T)
}
}
