% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liger.R
\name{runWilcoxon}
\alias{runWilcoxon}
\title{Perform Wilcoxon rank-sum test}
\usage{
runWilcoxon(object, data.use = NULL, compare.method, balance.by = NULL,
  max.sample = 1000, chunk = 1000)
}
\arguments{
\item{object}{\code{liger} object.}

\item{data.use}{This selects which dataset(s) to use. (default NULL, which uses all datasets)}

\item{compare.method}{Compare genes across \code{clusters} (using all datasets) or across \code{datasets} (within each cluster).}

\item{balance.by}{Sample from all datasets proportional to dataset size (default), balance the sample by \code{cluster} or 
balance the sample by \code{dataset}}

\item{max.sample}{Maximum number of cells to sample from either all datasets, each cluster, or each dataset
(depending on the value of \code{balance.by}). Default 1000.}

\item{chunk}{Number of cells at a time to read when constructing subset of data. Default 1000.}
}
\value{
A 10-columns data.frame with test results.
}
\description{
Perform Wilcoxon rank-sum tests on specified dataset using given method.
}
\examples{
\dontrun{
Y <- matrix(c(1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12), nrow = 4, byrow = T)
Z <- matrix(c(1, 2, 3, 4, 5, 6, 7, 6, 5, 4, 3, 2), nrow = 4, byrow = T)
X <- matrix(c(1, 2, 3, 4, 5, 6, 7, 2, 3, 4, 5, 6), nrow = 4, byrow = T)
ligerex <- createLiger(list(y_set = Y, z_set = Z, x_set = X))
ligerex <- normalize(ligerex)
# select genes
ligerex <- selectGenes(ligerex)
ligerex <- scaleNotCenter(ligerex)
ligerex <- optimizeALS(ligerex, k = 20)
ligerex <- quantileAlignSNF(ligerex)
wilcox.results <- runWilcoxon(ligerex, compare.method = "cluster")
wilcox.results <- runWilcoxon(ligerex, compare.method = "datastes", data.use = c(1, 2))
}

}
