% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rliger.R
\name{runWilcoxon}
\alias{runWilcoxon}
\title{Perform Wilcoxon rank-sum test}
\usage{
runWilcoxon(object, data.use = "all", compare.method)
}
\arguments{
\item{object}{\code{liger} object.}

\item{data.use}{This selects which dataset(s) to use. (default 'all')}

\item{compare.method}{This indicates the metric of the test. Either 'clusters' or 'datasets'.}
}
\value{
A 10-columns data.frame with test results.
}
\description{
Perform Wilcoxon rank-sum tests on specified dataset using given method.
}
\examples{
\dontrun{
Y <- matrix(c(1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12), nrow = 4, byrow = T)
Z <- matrix(c(1, 2, 3, 4, 5, 6, 7, 6, 5, 4, 3, 2), nrow = 4, byrow = T)
X <- matrix(c(1, 2, 3, 4, 5, 6, 7, 2, 3, 4, 5, 6), nrow = 4, byrow = T)
ligerex <- createLiger(list(y_set = Y, z_set = Z, x_set = X))
ligerex <- normalize(ligerex)
# select genes
ligerex <- selectGenes(ligerex)
ligerex <- scaleNotCenter(ligerex)
ligerex <- optimizeALS(ligerex, k = 20)
ligerex <- quantileAlignSNF(ligerex)
wilcox.results <- runWilcoxon(ligerex, compare.method = "cluster")
wilcox.results <- runWilcoxon(ligerex, compare.method = "datastes", data.use = c(1, 2))
# HDF5 input
ligerex <- online_iNMF(ligerex, k = 20)
ligerex <- quantile_norm(ligerex)
ligerex <- readSubset(ligerex, slot.use = "norm.data", max.cells = 1000)
de_genes <- runWilcoxon(ligerex, compare.method = "clusters")
}

}
