% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wilcoxon.R
\name{runWilcoxon}
\alias{runWilcoxon}
\title{Perform Wilcoxon rank-sum test}
\usage{
runWilcoxon(
  object,
  useDatasets = NULL,
  method = c("clusters", "datasets"),
  useCluster = "leiden_cluster",
  usePeak = FALSE,
  verbose = getOption("ligerVerbose"),
  data.use = useDatasets,
  compare.method = method
)
}
\arguments{
\item{object}{A \linkS4class{liger} object with cluster assignment available.}

\item{useDatasets}{A character vector of the names, a numeric or logical
vector of the index of the datasets to be normalized. Default
\code{NULL} uses all datasets.}

\item{method}{Choose from \code{"clusters"} or \code{"datasets"}. Default
\code{"clusters"} compares between clusters across all datasets, while
\code{"datasets"} compares between datasets within each cluster.}

\item{useCluster}{The name of the column in \code{cellMeta} slot storing the
cluster assignment variable. Default \code{"leiden_cluster"}}

\item{usePeak}{Logical, whether to test peak counts instead of gene
expression. Requires presence of ATAC modility datasets. Default
\code{FALSE}.}

\item{verbose}{Logical. Whether to show information of the progress. Default
\code{getOption("ligerVerbose")} which is \code{TRUE} if users have not set.}

\item{data.use, compare.method}{\bold{Deprecated}. See Usage section for
replacement.}
}
\value{
A 10-columns data.frame with test results.
}
\description{
Perform Wilcoxon rank-sum tests on specified dataset using
given method.
}
\examples{
library(dplyr)
data("pbmcPlot", package = "rliger")
result <- runWilcoxon(pbmcPlot)
result \%>\% group_by(group) \%>\% top_n(2, logFC)
}
