% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analogizer.R
\name{scaleNotCenter}
\alias{scaleNotCenter}
\title{Scale genes to unit variance (but no mean centering).}
\usage{
scaleNotCenter(object)
}
\arguments{
\item{object}{analogizer object. Should normalize and select genes before calling.}
}
\value{
analogizer object
}
\description{
Scale genes to unit variance (but no mean centering).
}
\examples{
\dontrun{
Y = matrix(c(1,2,3,4,5,6,7,8,9,10,11,12),nrow=4,byrow=T)
Z = matrix(c(1,2,3,4,5,6,7,6,5,4,3,2),nrow=4,byrow=T)
analogy = Analogizer(list(Y,Z))
analogy@var.genes = c(1,2,3,4)
analogy = scaleNotCenter(analogy)
}
}
