% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analogizer.R
\name{scaleNotCenter_sparse}
\alias{scaleNotCenter_sparse}
\title{Perform fast and memory-efficient data scaling operation on sparse matrix data.}
\usage{
scaleNotCenter_sparse(object, cells = NULL)
}
\arguments{
\item{object}{Sparse matrix DGE.}
}
\description{
Perform fast and memory-efficient data scaling operation on sparse matrix data.
}
\examples{
\dontrun{
Y = matrix(c(1,2,3,4,5,6,7,8,9,10,11,12),nrow=4,byrow=T)
Z = matrix(c(1,2,3,4,5,6,7,6,5,4,3,2),nrow=4,byrow=T)
analogy = Analogizer(list(Y,Z))
analogy@var.genes = c(1,2,3,4)
analogy = scaleNotCenter(analogy)
}
}
