% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectGenes.R
\name{selectGenes}
\alias{selectGenes}
\title{Select a subset of informative genes}
\usage{
selectGenes(
  object,
  var.thresh = 0.1,
  alpha.thresh = 0.99,
  num.genes = NULL,
  tol = 1e-04,
  capitalize = FALSE,
  combine = c("intersection", "union"),
  verbose = TRUE,
  chunkSize = 1000,
  ...
)
}
\arguments{
\item{object}{ligerObject}
}
\description{
This function identifies highly variable genes from each dataset
and combines these gene sets (either by union or intersection) for use in
downstream analysis. Assuming that gene expression approximately follows a
Poisson distribution, this function identifies genes with gene expression
variance above a given variance threshold (relative to mean gene expression).
}
