% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rliger.R
\docType{methods}
\name{show}
\alias{show}
\alias{show,liger-method}
\title{show method for liger}
\usage{
\S4method{show}{liger}(object)
}
\arguments{
\item{object}{liger object}
}
\description{
show method for liger
}
