% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liger-methods.R
\name{sub-liger}
\alias{sub-liger}
\alias{[.liger}
\title{Subset liger with brackets}
\usage{
\method{[}{liger}(x, i, j, ...)
}
\arguments{
\item{x}{A \linkS4class{liger} object}

\item{i}{Feature subscriptor, passed to \code{featureIdx} of
\code{\link{subsetLiger}}.}

\item{j}{Cell subscriptor, passed to \code{cellIdx} of
\code{\link{subsetLiger}}.}

\item{...}{Additional arguments passed to \code{\link{subsetLiger}}.}
}
\value{
Subset of \code{x} with specified features and cells.
}
\description{
Subset liger with brackets
}
\examples{
pbmcPlot[varFeatures(pbmcPlot)[1:10], 1:10]
}
\seealso{
\code{\link{subsetLiger}}
}
