% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ligerDataset-methods.R
\name{sub-ligerDataset}
\alias{sub-ligerDataset}
\alias{[.ligerDataset}
\title{Subset ligerDataset object}
\usage{
\method{[}{ligerDataset}(x, i, j, ...)
}
\arguments{
\item{x}{A \linkS4class{ligerDataset} object}

\item{i}{Numeric, logical index or character vector of feature names to
subscribe. Leave missing for all features.}

\item{j}{Numeric, logical index or character vector of cell IDs to subscribe.
Leave missing for all cells.}

\item{...}{Additional arguments passed to \code{\link{subsetLigerDataset}}.}
}
\value{
If \code{i} is given, the selected metadata will be returned; if it
is missing, the whole cell metadata table in
\code{S4Vectors::\link[S4Vectors]{DataFrame}} class will be returned.
}
\description{
Subset ligerDataset object
}
\examples{
ctrl <- dataset(pbmc, "ctrl")
ctrl[1:5, 1:5]
}
