% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liger-methods.R
\name{sub-sub-liger}
\alias{sub-sub-liger}
\alias{[[.liger}
\title{Get cell metadata variable}
\usage{
\method{[[}{liger}(x, i, ...)
}
\arguments{
\item{x}{A \linkS4class{liger} object}

\item{i}{Name or numeric index of cell meta data to fetch}

\item{...}{Anything that \code{S4Vectors::\link[S4Vectors]{DataFrame}}
method allows.}
}
\value{
If \code{i} is given, the selected metadata will be returned; if it
is missing, the whole cell metadata table in
\code{S4Vectors::\link[S4Vectors]{DataFrame}} class will be returned.
}
\description{
Get cell metadata variable
}
\examples{
# Retrieve whole cellMeta
pbmc[[]]
# Retrieve a variable
pbmc[["dataset"]]
}
