% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analogizer.R
\name{subsetAnalogizer}
\alias{subsetAnalogizer}
\title{Construct an Analogizer object with a specified subset}
\usage{
subsetAnalogizer(object, clusters.use = NULL, cells.use = NULL)
}
\arguments{
\item{object}{Analogizer object. Should run quantileAlignSNF and runTSNE before calling.}

\item{clusters.use}{Clusters to use for subset.}

\item{cells.use}{Vector of cell names to keep from any dataset.}
}
\value{
Analogizer object with subsetting applied to raw.data, norm.data, scale.data, H, W, V,
  H.norm, tsne.coords, and clusters.
}
\description{
The subset can be based on cell names or clusters. This function applies the subsetting to 
raw.data, norm.data, scale.data, H, W, V, H.norm, tsne.coords, and clusters. Note that it does
NOT reoptimize the factorization. See optimizeSubset for this functionality.
}
\examples{
\dontrun{
# analogizer object, with clusters 0:10
# factorization, alignment, and t-SNE calculation have been performed
analogy 
# subset by clusters
analogy_subset <- subsetAnalogizer(analogy, clusters.use = c(1, 4, 5))
}
}
