% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ligerDataset-class.R, R/subsetObject.R
\name{[,ligerDataset,index,missing,ANY-method}
\alias{[,ligerDataset,index,missing,ANY-method}
\alias{[,ligerDataset,missing,index,ANY-method}
\alias{[,ligerDataset,index,index,ANY-method}
\alias{subsetLigerDataset}
\alias{subsetH5LigerDataset}
\alias{subsetMemLigerDataset}
\title{Subset ligerDataset object}
\usage{
\S4method{[}{ligerDataset,index,missing,ANY}(x, i, j, ..., drop = TRUE)

\S4method{[}{ligerDataset,missing,index,ANY}(x, i, j, ..., drop = TRUE)

\S4method{[}{ligerDataset,index,index,ANY}(x, i, j, ..., drop = TRUE)

subsetLigerDataset(
  object,
  featureIdx = NULL,
  cellIdx = NULL,
  useSlot = NULL,
  newH5 = "auto",
  filename = NULL,
  filenameSuffix = NULL,
  chunkSize = 1000,
  verbose = TRUE,
  returnObject = TRUE,
  ...
)

subsetH5LigerDataset(
  object,
  featureIdx = NULL,
  cellIdx = NULL,
  useSlot = NULL,
  newH5 = "auto",
  filename = NULL,
  filenameSuffix = NULL,
  chunkSize = 1000,
  verbose = TRUE,
  returnObject = TRUE
)

subsetMemLigerDataset(
  object,
  featureIdx = NULL,
  cellIdx = NULL,
  useSlot = NULL,
  returnObject = TRUE
)
}
\arguments{
\item{i, featureIdx}{Character, logical or numeric index that can subscribe
features. Missing or \code{NULL} for all features.}

\item{j, cellIdx}{Character, logical or numeric index that can subscribe cells.
Missing or \code{NULL} for all cells.}

\item{...}{Arguments passed to \code{subsetH5LigerDataset}}

\item{object}{\linkS4class{ligerDataset} object. HDF5 based object if using
\code{subsetH5LigerDataset}, in-memory data for \code{subsetMemLigerDataset}.}

\item{useSlot}{The slot(s) to only consider. Choose one or more from
\code{"rawData"}, \code{"normData"} and \code{"scaleData"}. Default
\code{NULL} subsets the whole object including analysis result matrices.}

\item{newH5}{Whether to create a new H5 file on disk for the subset dataset
if \code{object} is HDF5 based. \code{TRUE} writes a new one, \code{FALSE}
returns in memory data. Default \code{"auto"} writes a new one when more than
8000 cells.}

\item{filename}{Filename of the new H5 file if being created. Default
\code{NULL} adds suffix \code{".subset_{yymmdd_HHMMSS}.h5"} to the original
name.}

\item{filenameSuffix}{Instead of specifying the exact filename, set a suffix
for the new files so the new filename looks like
\code{original.h5.[suffix].h5}. Default \code{NULL}.}

\item{chunkSize}{Integer. Number of maximum number of cells in each chunk,
Default \code{1000}.}

\item{verbose}{Logical. Whether to show the progress. Default \code{TRUE}.}
}
\value{
Subset \code{object}
}
\description{
This function subsets a \linkS4class{ligerDataset} object with
valid feature and cell indices. For HDF5 based object, options are available
for subsetting data into memory or a new on-disk H5 file. Feature and cell
subscription is always based on the size of rawData. Therefore, the feature
subsetting on scaled data, which usually contains already a subset of
features, will select the intersection between the wanted features and the
set available from scaled data.
}
