% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suggestParam.R
\name{suggestK}
\alias{suggestK}
\title{Visually suggest appropiate k value}
\usage{
suggestK(
  object,
  kTest = seq(5, 50, 5),
  lambda = 5,
  thresh = 1e-04,
  maxIter = 100,
  nCores = 1,
  seed = 1,
  genNew = FALSE,
  plotLog2 = TRUE,
  nrep = 1,
  returnRaw = FALSE,
  verbose = getOption("ligerVerbose"),
  k.test = kTest,
  max.iters = maxIter,
  num.cores = nCores,
  rand.seed = seed,
  gen.new = genNew,
  plot.log2 = plotLog2,
  return.raw = returnRaw,
  return.data = TRUE
)
}
\arguments{
\item{object}{A \linkS4class{liger} object with scaled data.}

\item{kTest}{Numeric vector of "k" values to test. Default \code{seq(5, 50,
5)}.}

\item{lambda}{Lambda to use for all factorization runs. Default \code{5}.}

\item{thresh}{Convergence threshold. Convergence occurs when
\eqn{|obj_0-obj|/(mean(obj_0,obj)) < thresh}. Default \code{1e-4}.}

\item{maxIter}{Maximum number of block coordinate descent iterations to
perform. Default \code{100}.}

\item{nCores}{Number of cores to use for optimizing factorization in
parallel. Default \code{1}.}

\item{seed}{Random seed to allow reproducible results. Default \code{1}.}

\item{genNew}{Logical. Whether to use \code{\link{optimizeNewK}} in
factorization. Results in slower factorization. Default \code{FALSE}.}

\item{plotLog2}{Logical. Whether to plot log2 curve for reference on K-L
plot. log2 is upper bound and can sometimes help in identifying "elbow" of
plot. Default \code{TRUE}.}

\item{nrep}{See \code{\link{optimizeALS}} for detail. Number restarts to
perform at each k value tested. Increase to produce smoother curve if results
unclear. Default \code{1}.}

\item{returnRaw}{Logical. Whether to return raw data or data.frame. Raw data
is a list of matrices of K-L divergences (length(k.test) by n_cells). Length
of list corresponds to \code{nrep}. (default FALSE)}

\item{verbose}{Logical. Whether to show information of the progress. Default
\code{getOption("ligerVerbose")} which is \code{TRUE} if users have not set.}

\item{k.test, max.iters, num.cores, rand.seed, gen.new, plot.log2, return.raw}{\bold{Deprecated}. See Usage section for replacement.}

\item{return.data}{\bold{Defuncted}. Will always return data now. Use
\code{\link{plotSuggestK}} to show the plot.}
}
\value{
Matrix of results
}
\description{
This can be used to select appropriate value of k for factorization of
particular dataset. Plots median (across cells in all datasets) K-L
divergence from uniform for cell factor loadings as a function of k. This
should increase as k increases but is expected to level off above
sufficiently high number of factors (k). This is because cells should have
factor loadings which are not uniformly distributed when an appropriate
number of factors is reached.
}
