% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suggestParam.R
\name{suggestLambda}
\alias{suggestLambda}
\title{Visually suggest appropriate lambda value}
\usage{
suggestLambda(
  object,
  k,
  lambdaTest = NULL,
  seed = 1,
  nCores = 1,
  thresh = 1e-04,
  maxIter = 100,
  nNeighbors = 20,
  reference = NULL,
  genNew = FALSE,
  nrep = 1,
  returnRaw = FALSE,
  verbose = getOption("ligerVerbose"),
  lambda.test = lambdaTest,
  rand.seed = seed,
  num.cores = nCores,
  max.iters = maxIter,
  knn_k = nNeighbors,
  ref_dataset = reference,
  gen.new = genNew,
  return.raw = returnRaw,
  return.data = TRUE,
  k2 = 500,
  resolution = 1
)
}
\arguments{
\item{object}{A \linkS4class{liger} object with scaled data.}

\item{k}{Number of factors to use in the test.}

\item{lambdaTest}{Numeric vector of "lambda" values to test. Default
\code{NULL} will test the following sequence: 0.25, 0.5, 0.75; 1, 2 to 10;
and 15, 20, ... to 60, totally 23 numbers.}

\item{seed}{Random seed to allow reproducible results. Default \code{1}.}

\item{nCores}{Number of cores to use for optimizing factorization in
parallel. Default \code{1}.}

\item{thresh}{Convergence threshold. Convergence occurs when
\eqn{|obj_0-obj|/(mean(obj_0,obj)) < thresh}. Default \code{1e-4}.}

\item{maxIter}{Maximum number of block coordinate descent iterations to
perform. Default \code{100}.}

\item{nNeighbors}{Number of nearest neighbors for within-dataset KNN in
\code{\link{quantileNorm}}. Default \code{20}.}

\item{reference}{Reference dataset for \code{\link{quantileNorm}}. See link
for detail. Default \code{NULL}.}

\item{genNew}{Logical. Whether to use \code{\link{optimizeNewLambda}} in
factorization. Recommended \code{TRUE} when looking at only a small range of
lambdas (i.e. 1:7). Default \code{FALSE}.}

\item{nrep}{See \code{\link{optimizeALS}} for detail. Number restarts to
perform at each k value tested. Increase to produce smoother curve if results
unclear. Default \code{1}.}

\item{returnRaw}{Logical. Whether to return raw data or data.frame. Raw data
is matrix of alignment values for each lambda value tested (each column
represents a different rep for nrep). Default \code{FALSE}.}

\item{verbose}{Logical. Whether to show information of the progress. Default
\code{getOption("ligerVerbose")} which is \code{TRUE} if users have not set.}

\item{lambda.test, rand.seed, num.cores, max.iters, knn_k, ref_dataset, gen.new, return.raw}{\bold{Deprecated}. See Usage section for replacement.}

\item{return.data}{\bold{Defuncted}. Will always return data now. Use
\code{\link{plotSuggestLambda}} to show the plot.}

\item{k2, resolution}{\bold{Defuncted}. Parameter passed to
\code{\link{quantileAlignSNF}} which is deprecated.}
}
\value{
Matrix of results
}
\description{
Can be used to select appropriate value of lambda for factorization of particular dataset. Plot
alignment and agreement for various test values of lambda. Most appropriate lambda
is likely around the "elbow" of the alignment plot (when alignment stops increasing). This will
likely also correspond to slower decrease in agreement. Depending on number of cores used,
this process can take 10-20 minutes.
}
