% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liger.R
\name{suggestLambda}
\alias{suggestLambda}
\title{Visually suggest appropriate lambda value}
\usage{
suggestLambda(object, k, lambda.test = NULL, rand.seed = 1, num.cores = 1,
  thresh = 1e-04, max.iters = 100, k2 = 500, ref_dataset = NULL,
  resolution = 1, agree.method = "NMF", gen.new = F, return.results = F)
}
\arguments{
\item{object}{\code{liger} object. Should normalize, select genes, and scale before calling.}

\item{k}{Number of factors to use in test factorizations. See optimizeALS documentation.}

\item{lambda.test}{Vector of lambda values to test. If not given, use default set spanning
0.25 to 60}

\item{rand.seed}{Random seed for reproducibility (default 1).}

\item{num.cores}{Number of cores to use for optimizing factorizations in parallel (default 1).}

\item{thresh}{Convergence threshold. Convergence occurs when |obj0-obj|/(mean(obj0,obj)) < thresh}

\item{max.iters}{Maximum number of block coordinate descent iterations to perform}

\item{k2}{Horizon parameter for quantileAlignSNF (default 500)}

\item{ref_dataset}{Reference dataset for quantileAlignSNF}

\item{resolution}{Resolution for quantileAlignSNF (default 1)}

\item{agree.method}{Reference dr.method for calculating agreement (default 'NMF')}

\item{gen.new}{Do not use optimizeNewLambda in factorizations. Recommended to set TRUE 
when looking at only a small range of lambdas (ie. 1:7) (default FALSE)}

\item{return.results}{Return matrix of alignment and agreement values (default FALSE)}
}
\value{
Matrix of results if desired (first column is test lambda values; second column
  is alignment; third and fourth columns are agreement before and after quantile alignment).
  Plots alignment vs. lambda and agreement vs. lambda to console.
}
\description{
Can be used to select appropriate value of lambda for factorization of particular dataset. Plot 
alignment and agreement for various test values of lambda. Most appropriate lambda
is likely around the "elbow" of the alignment plot (when alignment stops increasing). This will
likely also correspond to slower decrease in agreement. Depending on number of cores used,
this process can take 10-20 minutes.
}
\examples{
\dontrun{
Y <- matrix(c(1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12), nrow = 4, byrow = T)
Z <- matrix(c(1, 2, 3, 4, 5, 6, 7, 6, 5, 4, 3, 2), nrow = 4, byrow = T)
ligerex <- createLiger(list(y_set = Y, z_set = Z))
ligerex <- normalize(ligerex)
ligerex <- selectGenes(ligerex)
ligerex <- scaleNotCenter(ligerex)
# examine plot for most appropriate lambda, use multiple cores for faster results
suggestLambda(ligerex, k = 20, num.cores = 4)
}
}
